### 2024-12-14 afa2c69
[betaFam-minor-3]
1. separated `SR` and `SS` tests for `test_gamma_d.py` into `test_gamma_sr_d.py` and `test_gamma_ss_d.py`
2. minor changes in `test_exponential_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
