### 2024-12-15 245f5ca
[betaFam-minor-4]
1. added minor improvements to `beta_d.py` in both `cdf` and `pdf`
2. added `test_arcSine_d.py`
3. minor improvement in `test_beta_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
