### 2024-12-15 9bc95f0
[normalTesting-minor-1]
1. correct `foldedNormal_d.py` to comply with scipy
2. minor changes in `generators.py` and `halfNormal_d.py`
3. added `test_foldedNormal_d.py` although stats are not tested as they're not complying with scipy values

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
