+++ src/tests/test_distributions/test_foldedNormal_d.py
@@ -1,0 +1,62 @@
+"""Created on Dec 15 23:23:05 2024"""
+
+import numpy as np
+import pytest
+from scipy.stats import foldnorm
+
+from ...pymultifit import EPSILON
+from ...pymultifit.distributions import FoldedNormalDistribution
+from ...pymultifit.distributions.backend import errorHandling as erH
+
+
+class TestFoldedNormalDistribution:
+
+    @staticmethod
+    def test_initialization():
+        dist = FoldedNormalDistribution(amplitude=2.0, mean=1.0, sigma=0.5, normalize=False)
+        assert dist.amplitude == 2.0
+        assert dist.mean == 1.0
+        assert dist.sigma == 0.5
+        assert not dist.norm
+
+        dist_normalized = FoldedNormalDistribution(amplitude=2.0, normalize=True)
+        assert dist_normalized.amplitude == 1.0
+
+    @staticmethod
+    def test_constraints():
+        with pytest.raises(erH.NegativeAmplitudeError, match=f"Amplitude {erH.neg_message}"):
+            FoldedNormalDistribution(amplitude=-1.0, normalize=False)
+
+        # amplitude should be internally updated to 1.0 if `normalize` is called
+        distribution = FoldedNormalDistribution(amplitude=-1.0, normalize=True)
+        assert distribution.amplitude == 1.0
+
+        with pytest.raises(erH.NegativeStandardDeviationError, match=f"Standard deviation {erH.neg_message}"):
+            FoldedNormalDistribution(sigma=-3.0)
+
+    @staticmethod
+    def test_edge_case():
+        dist = FoldedNormalDistribution()
+        x = np.array([])
+        result = dist.pdf(x)
+        assert result.size == 0  # Should return an empty array
+
+    @staticmethod
+    def test_pdf_cdf():
+        def _cdf_pdf_custom(x_, dist_, what='cdf'):
+            return dist_.cdf(x_) if what == 'cdf' else dist_.pdf(x_)
+
+        def _cdf_pdf_scipy(x_, loc, scale, what='cdf'):
+            c = abs(loc) / scale
+            return foldnorm.cdf(x_, c=c, loc=0, scale=scale) if what == 'cdf' else foldnorm.pdf(x_, c=c, loc=0, scale=scale)
+
+        for i in ['pdf', 'cdf']:
+            for _ in range(50):  # Run 50 random tests
+                loc_ = np.random.uniform(low=-10, high=10)
+                scale_ = np.random.uniform(low=0.1, high=5)
+                x = np.random.uniform(low=EPSILON, high=loc_ + 10, size=50)
+                distribution = FoldedNormalDistribution(mean=loc_, sigma=scale_, normalize=True)
+                expected = _cdf_pdf_scipy(x_=x, loc=loc_, scale=scale_, what=i)
+                expected = np.nan_to_num(expected, True, 0)
+                np.testing.assert_allclose(actual=_cdf_pdf_custom(x_=x, dist_=distribution, what=i),
+                                           desired=expected, rtol=1e-5, atol=1e-8)
