### 2024-12-16 cca9874
[normalTesting-minor-2]
1. added `test_halfNormal_d.py`
2. fixed `halfNormal_d.py`
3. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
