+++ src/pymultifit/distributions/__init__.py
@@ -8,7 +8,7 @@ from .exponential_d import ExponentialDistribution
 from .foldedNormal_d import FoldedNormalDistribution
 from .gamma_d import GammaDistributionSR, GammaDistributionSS
 from .gaussian_d import GaussianDistribution
-from .halfNormal_d import NormalDistribution
+from .halfNormal_d import HalfNormalDistribution
 from .laplace_d import LaplaceDistribution
 from .logNorm_d import LogNormalDistribution
 from .norris_d import Norris2005Distribution, Norris2011Distribution
