+++ src/pymultifit/distributions/halfNormal_d.py
@@ -4,7 +4,7 @@ from .backend import errorHandling as erH
 from .foldedNormal_d import FoldedNormalDistribution
 
 
-class NormalDistribution(FoldedNormalDistribution):
+class HalfNormalDistribution(FoldedNormalDistribution):
     """A class for half normal distribution."""
 
     def __init__(self, amplitude: float = 1.0, scale: float = 1.0, normalize: bool = False):
@@ -12,4 +12,5 @@ class NormalDistribution(FoldedNormalDistribution):
             raise erH.NegativeAmplitudeError()
         elif scale <= 0:
             raise erH.NegativeScaleError()
+        self.scale = scale
         super().__init__(amplitude=amplitude, mean=0, sigma=scale, normalize=normalize)
