### 2024-12-17 b5d3d59
[main-minor-8]
1. test exclusion from packaging and version update in `version.py`
2. added `requirements[dev].txt` for development
