+++ MANIFEST.in
@@ -1,2 +1,3 @@
 include README.md
 include examples
+exclude src/tests
