+++ sphinx_docs/source/conf.py
@@ -1,0 +1,29 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# For the full list of built-in configuration values, see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+
+project = 'pymultifit'
+copyright = '2024, Syed Ali Mohsin Bukhari'
+author = 'Syed Ali Mohsin Bukhari'
+release = 'v0.2.2'
+
+# -- General configuration ---------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
+
+extensions = ['sphinx.ext.autodoc',  # For automatic documentation from docstrings
+              'sphinx.ext.napoleon',  # To support Google or NumPy-style docstrings
+              'sphinx.ext.viewcode',  # Adds source code links
+              ]
+
+templates_path = ['_templates']
+exclude_patterns = []
+
+# -- Options for HTML output -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
+
+html_theme = 'alabaster'
+html_static_path = ['_static']
