+++ sphinx_docs/source/pymultifit.rst
@@ -1,0 +1,31 @@
+pymultifit package
+==================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pymultifit.distributions
+   pymultifit.fitters
+   pymultifit.generators
+
+Submodules
+----------
+
+pymultifit.version module
+-------------------------
+
+.. automodule:: pymultifit.version
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pymultifit
+   :members:
+   :undoc-members:
+   :show-inheritance:
