+++ sphinx_docs/source/modules.rst
@@ -1,0 +1,7 @@
+pymultifit
+==========
+
+.. toctree::
+   :maxdepth: 4
+
+   pymultifit
