+++ sphinx_docs/source/index.rst
@@ -1,0 +1,18 @@
+.. pymultifit documentation master file, created by
+   sphinx-quickstart on Tue Dec 17 09:53:46 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+pymultifit documentation
+========================
+
+Add your content using ``reStructuredText`` syntax. See the
+`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
+documentation for details.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   pymultifit
