+++ sphinx_docs/source/pymultifit.fitters.rst
@@ -1,0 +1,117 @@
+pymultifit.fitters package
+==========================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pymultifit.fitters.backend
+
+Submodules
+----------
+
+pymultifit.fitters.chiSquare\_f module
+--------------------------------------
+
+.. automodule:: pymultifit.fitters.chiSquare_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.exponential\_f module
+----------------------------------------
+
+.. automodule:: pymultifit.fitters.exponential_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.foldedNormal\_f module
+-----------------------------------------
+
+.. automodule:: pymultifit.fitters.foldedNormal_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.gaussian\_f module
+-------------------------------------
+
+.. automodule:: pymultifit.fitters.gaussian_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.halfNormal\_f module
+---------------------------------------
+
+.. automodule:: pymultifit.fitters.halfNormal_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.laplace\_f module
+------------------------------------
+
+.. automodule:: pymultifit.fitters.laplace_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.logNorm\_f module
+------------------------------------
+
+.. automodule:: pymultifit.fitters.logNorm_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.mixed\_f module
+----------------------------------
+
+.. automodule:: pymultifit.fitters.mixed_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.norris\_f module
+-----------------------------------
+
+.. automodule:: pymultifit.fitters.norris_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.powerLaw\_f module
+-------------------------------------
+
+.. automodule:: pymultifit.fitters.powerLaw_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.skewNorm\_f module
+-------------------------------------
+
+.. automodule:: pymultifit.fitters.skewNorm_f
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pymultifit.fitters.utilities module
+-----------------------------------
+
+.. automodule:: pymultifit.fitters.utilities
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pymultifit.fitters
+   :members:
+   :undoc-members:
+   :show-inheritance:
