+++ sphinx_docs/source/pymultifit.fitters.backend.rst
@@ -1,0 +1,21 @@
+pymultifit.fitters.backend package
+==================================
+
+Submodules
+----------
+
+pymultifit.fitters.backend.baseFitter module
+--------------------------------------------
+
+.. automodule:: pymultifit.fitters.backend.baseFitter
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pymultifit.fitters.backend
+   :members:
+   :undoc-members:
+   :show-inheritance:
