+++ .readthedocs.yml
@@ -1,0 +1,20 @@
+# Read the Docs configuration file
+# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
+
+# Required
+version: 2
+
+# Set the OS, Python version, and other tools you might need
+build:
+  os: ubuntu-24.04
+  tools:
+    python: "3.9"
+
+# Build documentation in the "sphinx_docs" directory with Sphinx
+sphinx:
+   configuration: sphinx_docs/source/conf.py
+
+# Declare the Python requirements to build your documentation
+python:
+   install:
+     - requirements: ./requirements[dev].txt
