### 2024-12-17 b17bdb8
[docs-minor-1-2]
1. testing readthedoc triggering using a dummy `requirements.txt`
2. added `sphinx-autobuilder` for local building
