+++ .readthedocs.yml
@@ -6,15 +6,21 @@ version: 2
 
 # Set the OS, Python version, and other tools you might need
 build:
-  os: ubuntu-24.04
-  tools:
-    python: "3.9"
+    os: ubuntu-24.04
+    tools:
+        python: "3.9"
 
 # Build documentation in the "sphinx_docs" directory with Sphinx
 sphinx:
-   configuration: ./sphinx_docs/source/conf.py
+    builder: html
+    configuration: ./sphinx_docs/source/conf.py
+    fail_on_warning: true
+
+# recursive addition of submodules
+submodules:
+    recursive: true
 
 # Declare the Python requirements to build your documentation
 python:
-   install:
-     - requirements: ./requirements[dev].txt
+    install:
+        -   requirements: ./requirements[dev].txt
