### 2024-12-17 ac128e6
[docs-minor-2]
1. updated `conf.py`
2. updated `requirements[dev].txt`
