+++ sphinx_docs/source/conf.py
@@ -3,6 +3,10 @@
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
+import datetime
+import os
+import sys
+
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
@@ -14,15 +18,101 @@ release = 'v0.2.2'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ['sphinx.ext.autodoc',  # For automatic documentation from docstrings
-              'sphinx.ext.napoleon',  # To support Google or NumPy-style docstrings
-              'sphinx.ext.viewcode']
+## COPIED from PYLOPS
+sys.path.insert(0, os.path.abspath("../../src"))
+extensions = ["sphinx.ext.autodoc",
+              "sphinx.ext.autosummary",
+              "sphinx.ext.coverage",
+              "sphinx.ext.mathjax",
+              "sphinx.ext.doctest",
+              "sphinx.ext.extlinks",
+              "sphinx.ext.intersphinx",
+              "numpydoc",
+              "nbsphinx",
+              "sphinx.ext.viewcode",
+              "matplotlib.sphinxext.plot_directive"]
+
+intersphinx_mapping = {"python": ("https://docs.python.org/3/", None),
+                       "numpy": ("https://docs.scipy.org/doc/numpy/", None),
+                       "scipy": ("https://docs.scipy.org/doc/scipy/reference", None),
+                       "pandas": ("http://pandas.pydata.org/pandas-docs/stable/", None),
+                       "matplotlib": ("https://matplotlib.org/", None)}
+
+autosummary_generate = True
+autodoc_default_flags = ["members"]
+autodoc_typehints = "none"
+
+numpydoc_show_class_members = False
+numpydoc_show_inherited_class_members = False
+numpydoc_class_members_toctree = False
+#
+# sphinx_gallery_conf = {
+#     # path to your examples scripts
+#     "examples_dirs": ["../../examples", ],
+#     # path where to save gallery generated examples
+#     "gallery_dirs": ["gallery", "examples"],
+#     "filename_pattern": r"\.py",
+#     # Remove the "Download all examples" button from the top level gallery
+#     "download_all_examples": False,
+#     # Sort gallery example by file name instead of number of lines (default)
+#     "within_subsection_order": ExampleTitleSortKey,
+#     # Modules for which function level galleries are created.
+#     "doc_module": "pymultifit",
+#     # Insert links to documentation of objects in the examples
+#     "reference_url": {"pymultifit": None},
+# }
+
+# Always show the source code that generates a plot
+plot_include_source = True
+plot_formats = ["png"]
 
 templates_path = ['_templates']
-exclude_patterns = []
+exclude_patterns = ["_build", "**.ipynb_checkpoints", "**.ipynb", "**.md5"]
+source_suffix = {'.rst': 'restructuredtext'}
 
+# General information about the project
+year = datetime.date.today().year
+
+# These enable substitutions using |variable| in the rst files
+rst_epilog = """
+.. |year| replace:: {year}
+""".format(year=year)
+html_static_path = ["_static"]
+html_css_files = ['custom.css']
+html_last_updated_fmt = "%b %d, %Y"
+html_title = "pyMultiFit"
+html_short_title = "pyMultiFit"
+html_extra_path = []
+pygments_style = "default"
+add_function_parentheses = True
+html_show_sourcelink = True
+html_show_sphinx = True
[…]
