### 2024-12-17 854b34b
[docs-minor-2-1]
1. checking `branches-ignore` property for doc branch
