+++ .github/workflows/code_coverage.yml
@@ -1,7 +1,9 @@
 name: Run tests and upload coverage
 
 on:
-    push
+    push:
+        branches-ignore:
+            - doc
 
 jobs:
     test:
