+++ .github/workflows/publish_pymultifit.yml
@@ -9,34 +9,34 @@
 name: publish_pymultifit
 
 on:
-  release:
-    types: [ published ]
+    release:
+        types: [ published ]
 
 permissions:
-  contents: read
+    contents: read
 
 jobs:
-  deploy:
+    deploy:
 
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [ "3.9" ]
+        runs-on: ubuntu-latest
+        strategy:
+            matrix:
+                python-version: [ "3.9" ]
 
-    steps:
-      - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install build
-      - name: Build package
-        run: python -m build
-      - name: Publish package
-        uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
-        with:
-          user: __token__
-          password: ${{ secrets.GH_AUTH_TOKEN_ }}
+        steps:
+            -   uses: actions/checkout@v4
+            -   name: Set up Python
+                uses: actions/setup-python@v4
+                with:
+                    python-version: ${{ matrix.python-version }}
+            -   name: Install dependencies
+                run: |
+                    python -m pip install --upgrade pip
+                    pip install build
+            -   name: Build package
+                run: python -m build
+            -   name: Publish package
+                uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
+                with:
+                    user: __token__
+                    password: ${{ secrets.GH_AUTH_TOKEN_ }}
