+++ sphinx_docs/source/conf.py
@@ -33,10 +33,10 @@ extensions = ["sphinx.ext.autodoc",
               "matplotlib.sphinxext.plot_directive"]
 
 intersphinx_mapping = {"python": ("https://docs.python.org/3/", None),
-                       "numpy": ("https://docs.scipy.org/doc/numpy/", None),
-                       "scipy": ("https://docs.scipy.org/doc/scipy/reference", None),
+                       "numpy": ("https://docs.scipy.org/stable/", None),
+                       'scipy': ('https://docs.scipy.org/doc/scipy/', None),
                        "pandas": ("http://pandas.pydata.org/pandas-docs/stable/", None),
-                       "matplotlib": ("https://matplotlib.org/", None)}
+                       "matplotlib": ("https://matplotlib.org/stable/", None)}
 
 autosummary_generate = True
 autodoc_default_flags = ["members"]
