### 2024-12-17 647d1ee
[docs-major-1]
1. updated `pyproject.toml`, `requirements[dev].txt`, `requirements.txt`, `environment.yaml` and `setup.py`
2. updated `index.rst`, `pymultifit.rst`, `modules.rst` and added `installation.rst`
3. minor changes in `chiSquare_d.py` and `conf.py`
