+++ sphinx_docs/source/modules.rst
@@ -1,4 +1,4 @@
-pymultifit
+modules
 ==========
 
 .. toctree::
