+++ sphinx_docs/source/pymultifit.rst
@@ -1,4 +1,4 @@
-pymultifit package
+pyMultiFit
 ==================
 
 Subpackages
