### 2024-12-19 #221 @SimonMolinsky
Hi @syedalimohsinbukhari

Thanks for submitting your package! It's a great tool, but at this point, I wonder if it is within the scope of pyOpenSci. You have written that `pymultifit` overlaps with other packages in the ecosystem, but the differences are unclear. The package must fulfill one of those conditions:

```
- More open in licensing or development practices
- Broader in functionality (e.g., providing access to more data sets, providing a greater suite of functions), but not only by duplicating additional packages
- Better in usability and performance
- Actively maintained while alternatives are poorly or no longer actively maintained
```

You have stated that your package has better usability and performance; I've checked your package's documentation (https://pymultifit.readthedocs.io/index.html#), and you don't provide examples of how to use your package - thus, I wasn't able to compare functionalities. Could you prepare some comparisons of overlapping functionalities as the code examples? It could be a part of your `README` in the future. Then we can decide if `pymultifit` can be accepted to pyOpenSci.
