### 2024-12-19 #221 @syedalimohsinbukhari
Hi @SimonMolinsky

Thank you for reaching out. I apologize for my oversight of incomplete documentation before presumission. Currently the documentation of the package is in development in the `docs` branch, which will be up ASAP, including proper usage examples and comparisons as well.
