### 2024-12-19 d72c970
[doc-minor-1]
1. updated `conf.py` to work for home PC
2. updated `installation.rst`
3. added `usage.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
