+++ sphinx_docs/source/conf.py
@@ -33,10 +33,10 @@ extensions = ["sphinx.ext.autodoc",
               "sphinx.ext.viewcode",
               "matplotlib.sphinxext.plot_directive"]
 
-intersphinx_mapping = {"python": ("https://docs.python.org/3/", None),
-                       'scipy': ('https://docs.scipy.org/doc/scipy/', None),
-                       "pandas": ("http://pandas.pydata.org/pandas-docs/stable/", None),
-                       "matplotlib": ("https://matplotlib.org/stable/", None)}
+# intersphinx_mapping = {"python": ("https://docs.python.org/3/", None),
+#                        'scipy': ('https://docs.scipy.org/doc/scipy/', None),
+#                        "pandas": ("http://pandas.pydata.org/pandas-docs/stable/", None),
+#                        "matplotlib": ("https://matplotlib.org/stable/", None)}
 
 autosummary_generate = True
 autodoc_default_flags = ["members"]
