### 2024-12-19 a13bf54
[doc-minor-2]
1. updated `conf.py`
2. added `generating_data.ipynb` and `generating_data.nblink` for `tutorials.rst`
3. updated `installation.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
