+++ sphinx_docs/source/tutorials.rst
@@ -1,0 +1,16 @@
+Tutorials
+=========
+
+
+The ``pyMultiFit`` library is divided into three modules,
+
+1. **distributions**, housing statistical (and non-statistical) distributions, such as ``GaussianDistribution``, ``LaplaceDistribution``, and more.
+2. **generators**, can generate single or n-model data based on the selected distribution. And finally
+3. **fitters**, that can fit n-model data on the given data.asdf
+
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   generating_data.ipynb
