### 2024-12-20 3fc206d
[docs-minor-3]
1. updated `conf.py` for warning suppressions, have to check in with the gallery though
2. updated `environment.yaml`, `requirements[dev].txt` with more dependencies
3. finalized `generating_data.ipynb` as the first tutorial
4. minor necessary changes to `utilities.py`

pushing to see if RTD is fixed or not
