+++ src/pymultifit/distributions/utilities.py
@@ -65,38 +65,38 @@ def beta_(x: np.ndarray,
 
 
 def chi_squared_(x, amplitude: float = 1., degree_of_freedom: float = 1., normalize: bool = False):
-    """
-    Compute the Chi-Squared distribution.
-
-    The Chi-Squared distribution is a special case of the Gamma distribution with `alpha = degree_of_freedom / 2` and `beta = 1 / 2`.
-
-    Parameters
-    ----------
-    x : array-like
-        The input values at which to evaluate the Chi-Squared distribution.
-    amplitude : float, optional
-        The amplitude of the distribution. Defaults to 1.
-        Ignored if `normalize` is set to True.
-    degree_of_freedom : float, optional
-        The degrees of freedom of the Chi-Squared distribution. Defaults to 1.
-    normalize : bool, optional
-        If True, the distribution will be normalized so that the PDF is at most 1. Defaults to False.
-
-    Returns
-    -------
-    array-like
-        The computed Chi-Squared distribution values for the input `x`.
-
-    Notes
-    -----
-    The Chi-Squared distribution is related to the Gamma distribution by:
-    .. math::
-        f(x; k) = \\frac{x^{(k/2) - 1} e^{-x/2}}{2^{k/2} \\Gamma(k/2)}
-
-    where `k` is the degrees of freedom, and `Γ(k)` is the Gamma function.
-
-    If `normalize` is True, the distribution will be scaled such that the maximum value of the PDF is 1.
-    """
+    # """
+    # Compute the Chi-Squared distribution.
+    #
+    # The Chi-Squared distribution is a special case of the Gamma distribution with `alpha = degree_of_freedom / 2` and `beta = 1 / 2`.
+    #
+    # Parameters
+    # ----------
+    # x : array-like
+    #     The input values at which to evaluate the Chi-Squared distribution.
+    # amplitude : float, optional
+    #     The amplitude of the distribution. Defaults to 1.
+    #     Ignored if `normalize` is set to True.
+    # degree_of_freedom : float, optional
+    #     The degrees of freedom of the Chi-Squared distribution. Defaults to 1.
+    # normalize : bool, optional
+    #     If True, the distribution will be normalized so that the PDF is at most 1. Defaults to False.
+    #
+    # Returns
+    # -------
+    # array-like
+    #     The computed Chi-Squared distribution values for the input `x`.
+    #
+    # Notes
+    # -----
+    # The Chi-Squared distribution is related to the Gamma distribution by:
+    # .. math::
+    #     f(x; k) = \\frac{x^{(k/2) - 1} e^{-x/2}}{2^{k/2} \\Gamma(k/2)}
+    #
+    # where `k` is the degrees of freedom, and `Γ(k)` is the Gamma function.
+    #
+    # If `normalize` is True, the distribution will be scaled such that the maximum value of the PDF is 1.
+    # """
     return gamma_ss_(x, amplitude=amplitude, shape=degree_of_freedom / 2., scale=2., normalize=normalize)
 
 
@@ -135,39 +135,39 @@ def exponential_(x: np.ndarray,
 def folded_normal_(x: np.ndarray,
                    amplitude: float = 1., mu: float = 0.0, variance: float = 1.0,
                    normalize: bool = False) -> np.ndarray:
-    """
-    Compute the folded half-normal distribution.
-
-    The folded half-normal distribution is the sum of two Gaussian distributions mirrored around `mu`.
-    It is defined as the sum of a normal distribution centered at `mu` and another mirrored at `-mu`.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        Input array where the folded half-normal distribution is evaluated.
-    amplitude : float, optional
-        The amplitude of the distribution. Defaults to 1. Ignored if `normalize` is set to True.
-    mu : float, optional
[…]
