+++ sphinx_docs/source/modules.rst
@@ -1,7 +1,0 @@
-modules
-==========
-
-.. toctree::
-   :maxdepth: 4
-
-   pymultifit
