### 2024-12-20 9f4cb27
[doc-minor-4]
1. renamed tuts to tutorials
2. minor changes to `conf.py`, `requirements[dev].txt` and `generators.py`
3. added more examples to `generating_data.ipynb`
4. properly updated `index.rst`, `installation.rst` to include `tutorials.rst`
