+++ src/pymultifit/generators/generators.py
@@ -63,14 +63,14 @@ def generate_multi_half_normal_data(x: np.ndarray, params: listOfTuplesOrArray,
                                     noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
     y = np.zeros_like(x, dtype=float)
     for pars in params:
-        y += dist.NormalDistribution(*pars, normalize=normalize).pdf(x)
+        y += dist.HalfNormalDistribution(*pars, normalize=normalize).pdf(x)
     if noise_level > 0:
         y += noise_level * np.random.normal(size=x.size)
     return y
 
 
-def generate_multi_gaussian_data(x: np.ndarray, params: listOfTuplesOrArray,
-                                 noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
+def multi_gaussian(x: np.ndarray, params: listOfTuplesOrArray,
+                   noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
     """
     Generate multi-Gaussian data with optional noise.
 
