+++ sphinx_docs/source/tutorials.rst
@@ -8,9 +8,9 @@ The ``pyMultiFit`` library is divided into three modules,
 2. **generators**, can generate single or n-model data based on the selected distribution. And finally
 3. **fitters**, that can fit n-model data on the given data.asdf
 
-
 .. toctree::
    :maxdepth: 2
    :hidden:
 
-   generating_data.ipynb
+   generating_data
+   generating_n_model_data
