### 2024-12-20 4f70ad9
[doc-minor-4-1]
1. added a template for n-model data generation
