+++ src/pymultifit/generators/generators.py
@@ -1,6 +1,6 @@
 """Created on Jul 18 00:35:26 2024"""
 
-from typing import List, Tuple, Union
+from typing import Callable, List, Tuple, Union
 
 import numpy as np
 
@@ -10,8 +10,49 @@ listOfTuples = List[Tuple[float, ...]]
 listOfTuplesOrArray = Union[listOfTuples, np.ndarray]
 
 
-def generate_multi_exponential_data(x: np.ndarray, params: listOfTuplesOrArray,
-                                    noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
+def multi_chi_squared(x: np.ndarray, params: listOfTuplesOrArray,
+                      noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
+    return multi_distribution(x=x, distribution_func=dist.ChiSquareDistribution, params=params, noise_level=noise_level, normalize=normalize)
+
+
+def multi_distribution(x: np.ndarray, distribution_func: Callable, params: listOfTuplesOrArray,
+                       noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
+    """
+    Generate data based on a combination of distributions with optional noise.
+
+    Parameters
+    ----------
+    x : np.ndarray
+        X values.
+    distribution_func : callable
+        The distribution function to use for generating data.
+        This function must accept the parameters in `params` as arguments.
+    params : List[Tuple]
+        List of tuples containing the parameters for each distribution.
+        The parameters should match the input required by the distribution function.
+    noise_level : float, optional
+        Standard deviation of the noise to be added to the data, by default 0.0.
+    normalize: bool
+        If True, the function produces normalized data (Integration[PDF] < 1). Defaults to False.
+
+    Returns
+    -------
+    np.ndarray
+        Y values of the generated data with added noise.
+    """
+    y = np.zeros_like(x, dtype=float)
+
+    for param_set in params:
+        y += distribution_func(*param_set, normalize=normalize).pdf(x)
+
+    if noise_level > 0:
+        y += noise_level * np.random.normal(size=x.size)
+
+    return y
+
+
+def multi_exponential(x: np.ndarray, params: listOfTuplesOrArray,
+                      noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
     """
     Generate multi-Exponential data with optional noise.
 
@@ -31,42 +72,12 @@ def generate_multi_exponential_data(x: np.ndarray, params: listOfTuplesOrArray,
     np.ndarray
         Y values of the multi-Exponential data with added noise.
     """
-    y = np.zeros_like(x, dtype=float)
-    for pars in params:
-        y += dist.ExponentialDistribution(*pars, normalize=normalize).pdf(x)
-    if noise_level > 0:
-        y += noise_level * np.random.normal(size=x.size)
-    return y
+    return multi_distribution(x=x, distribution_func=dist.ExponentialDistribution, params=params, noise_level=noise_level, normalize=normalize)
 
 
-def generate_multi_chi_squared_data(x: np.ndarray, params: listOfTuplesOrArray,
-                                    noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
-    y = np.zeros_like(x, dtype=float)
-    for pars in params:
-        y += dist.ChiSquareDistribution(*pars, normalize=normalize).pdf(x)
-    if noise_level > 0:
-        y += noise_level * np.random.normal(size=x.size)
-    return y
-
-
-def generate_multi_folded_normal_data(x: np.ndarray, params: listOfTuplesOrArray,
-                                      noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
[…]
