+++ examples/laplace.py
@@ -4,14 +4,14 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from src.pymultifit.fitters import LaplaceFitter
-from src.pymultifit.generators import generate_multi_laplace_data
+from src.pymultifit.generators import multi_laplace
 
 params = [(10, -20, 2), (4, -5.5, 10), (5, -1, 0.5), (10, 3, 1), (4, 15, 3)]
 
 x = np.linspace(-35, 35, 1500)
 
 noise_level = 0.2
-y = generate_multi_laplace_data(x, params=params, noise_level=noise_level)
+y = multi_laplace(x, params=params, noise_level=noise_level)
 
 fitter = LaplaceFitter(n_fits=5, x_values=x, y_values=y)
 
