### 2024-12-20 014fc21
[doc-minor-6]
1. added `generating_mixture_model_data.ipynb` and `generating_mixture_model_data.nblink`
2. updated `tutorials.rst`
3. minor changes to `generating_n_model_data.ipynb`
