+++ sphinx_docs/source/tutorials.rst
@@ -14,3 +14,4 @@ The ``pyMultiFit`` library is divided into three modules,
 
    generating_data
    generating_n_model_data
+   generating_mixture_model_data
