### 2024-12-20 81e3c4e
[doc-minor-7]
1. added `custom-inherit` to `environment.yaml`, `requirements.txt`, `requirements[dev].txt`, `pyproject.toml` and `setup.py`
2. added docstrings to `generators.py`
3. minor changes to `mixed_f.py` and `generating_mixture_model_data.ipynb`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
