+++ src/pymultifit/generators/generators.py
@@ -3,37 +3,43 @@
 from typing import Callable, List, Tuple, Union
 
 import numpy as np
+from custom_inherit import doc_inherit
 
 from .. import distributions as dist, GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, POWERLAW, SKEW_NORMAL
 
 listOfTuples = List[Tuple[float, ...]]
 listOfTuplesOrArray = Union[listOfTuples, np.ndarray]
 
+doc_style = 'numpy_napoleon_with_merge'
+
 
 def multi_chi_squared(x: np.ndarray, params: listOfTuplesOrArray,
                       noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
-    return multi_distribution(x=x, distribution_func=dist.ChiSquareDistribution, params=params, noise_level=noise_level, normalize=normalize)
-
-
-def multi_distribution(x: np.ndarray, distribution_func: Callable, params: listOfTuplesOrArray,
-                       noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
-    """
-    Generate data based on a combination of distributions with optional noise.
+    """Generate multi-Chi-Squared data with optional noise.
 
     Parameters
     ----------
     x : np.ndarray
         X values.
-    distribution_func : callable
-        The distribution function to use for generating data.
-        This function must accept the parameters in `params` as arguments.
-    params : List[Tuple]
-        List of tuples containing the parameters for each distribution.
-        The parameters should match the input required by the distribution function.
+    params : listOfTuplesOrArray
+        List of tuples containing the parameters for each Chi-Squared distribution.
     noise_level : float, optional
         Standard deviation of the noise to be added to the data, by default 0.0.
-    normalize: bool
+    normalize : bool, optional
         If True, the function produces normalized data (Integration[PDF] < 1). Defaults to False.
+
+    Returns
+    -------
+    np.ndarray
+        Y values of the multi-Chi-Squared data with added noise.
+    """
+    return multi_distribution(x=x, distribution_func=dist.ChiSquareDistribution, params=params, noise_level=noise_level, normalize=normalize)
+
+
+@doc_inherit(parent=multi_chi_squared, style=doc_style)
+def multi_distribution(x: np.ndarray, distribution_func: Callable, params: listOfTuplesOrArray,
+                       noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
+    """Generate data based on a combination of distributions with optional noise.
 
     Returns
     -------
@@ -51,21 +57,10 @@ def multi_distribution(x: np.ndarray, distribution_func: Callable, params: listO
     return y
 
 
+@doc_inherit(parent=multi_chi_squared, style=doc_style)
 def multi_exponential(x: np.ndarray, params: listOfTuplesOrArray,
                       noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
-    """
-    Generate multi-Exponential data with optional noise.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        X values.
-    params : List[Tuple[float, ...]]
-        List of tuples containing the parameters for each Exponential (amplitude, scale).
-    noise_level : float, optional
-        Standard deviation of the Exponential noise to be added to the data, by default 0.0.
-    normalize: bool
-        If True, the function produces normalized data (Integration[PDF] < 1). Defaults to False.
+    """Generate multi-Exponential data with optional noise.
 
     Returns
     -------
@@ -75,26 +70,23 @@ def multi_exponential(x: np.ndarray, params: listOfTuplesOrArray,
[…]
