+++ setup.py
@@ -30,7 +30,7 @@ setup(name='pymultifit',
       long_description=readme,
       long_description_content_type="text/markdown",
       python_requires=">=3.9",
-      install_requires=["setuptools", "numpy<2.1.0", "matplotlib", "scipy", "mpyez"],
+      install_requires=["setuptools", "numpy<2.1.0", "matplotlib", "scipy", "mpyez", "custom-inherit"],
       classifiers=["License :: OSI Approved :: MIT License",
                    "Programming Language :: Python :: 3.9",
                    "Programming Language :: Python :: 3.10",
