+++ requirements.txt
@@ -2,3 +2,4 @@ numpy<2.1.0
 matplotlib
 scipy
 mpyez
+custom-inherit
