+++ environment.yaml
@@ -7,6 +7,7 @@ dependencies:
     - numpy<2.1.0
     - matplotlib
     - scipy
+    - custom-inherit
     - pytest
     - jupyter
     - notebook
