### 2024-12-20 4fc087b
[doc-major-2]
1. minor changes to `baseDistribution.py`, `conf.py`, and `custom.css`
2. updated `environment.yaml` and `requirements[dev].txt` with `sphinxcontrib-napoleon`
3. added `fitting_mixture_model.ipynb` and `fitting_mixture_model.nblink`
4. minor docstring changes to `gamma_d.py`
5. minor changes to `generating_mixture_model_data.ipynb`
6. added `index.rst` to API folder
7. updated `source/index.rst`
8. updated `tutorials.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
