+++ sphinx_docs/source/conf.py
@@ -29,10 +29,11 @@ extensions = ["sphinx.ext.autodoc",
               "sphinx.ext.intersphinx",
               # "sphinx_gallery.gen_gallery", # have to work on gallery later
               "sphinx_copybutton",
-              "numpydoc",
+              # "numpydoc",
               "nbsphinx",
               "nbsphinx_link",
               "sphinx.ext.viewcode",
+              'sphinx.ext.napoleon',
               "matplotlib.sphinxext.plot_directive"]
 
 nbsphinx_execute = 'auto'
@@ -47,11 +48,8 @@ suppress_warnings = ["config.cache"]
 autosummary_generate = True
 autodoc_default_flags = ["members"]
 autodoc_typehints = "none"
+add_module_names = False
 
-numpydoc_show_class_members = False
-numpydoc_show_inherited_class_members = False
-numpydoc_class_members_toctree = False
-#
 # sphinx_gallery_conf = {
 #     # path to your examples scripts
 #     "examples_dirs": ["../../examples", ],
