+++ sphinx_docs/source/index.rst
@@ -55,3 +55,4 @@ With **pyMultiFit**, you can focus more on analyzing your data and less on redef
    self
    installation
    tutorials
+   api/index
