### 2024-12-21 0517b92
[doc-minor-8]
1. renamed `generating_data.ipynb` to `001__generating_data.ipynb`
2. renamed `generating_n_model_data.ipynb` to `002__generating_n_model_data.ipynb`
3. updated `conf.py`
4. alphabetical sorting in  `errorHandling.py`
5. updated `index.rst`, `pymultifit.distributions.rst` and `tutorials.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
