+++ sphinx_docs/source/api/index.rst
@@ -11,10 +11,10 @@ The Application Programming Interface (API) of `pyMultiFit` provides tools for s
 
 --------------------------------------------
 
-Distributions
+distributions
 -------------
 
-The `Distributions` module provides tools for defining, generating, and using statistical distributions.
+The `distributions` module provides tools for defining, generating, and using statistical distributions.
 It includes a base distributions and several specific implementations.
 
 Base Class
@@ -63,3 +63,96 @@ Derived Distributions
      - Skew-Normal distribution.
    * - :class:`UniformDistribution <pymultifit.distributions.uniform_d.UniformDistribution>`
      - Uniform distribution.
+
+--------------------------------------------
+
+fitters
+-------
+
+The `fitters` module provides tools for fitting single, multiple and mixed models.
+It includes a base fitter method for user-defined fitting algorithms and several specific implementations.
+
+Base Class
+^^^^^^^^^^
+.. list-table::
+   :widths: 25 75
+   :header-rows: 1
+
+   * - Name
+     - Description
+   * - :class:`BaseFitter <pymultifit.fitters.backend.baseFitter.BaseFitter>`
+     - The base class for multi-fitting functionality.
+   * - :class:`MixedFitter <pymultifit.fitters.mixed_f.MixedFitter>`
+     - Mixed model fitting class.
+
+Derived Fitters
+^^^^^^^^^^^^^^^^
+
+.. list-table::
+   :widths: 25 75
+   :header-rows: 1
+
+   * - Name
+     - Description
+   * - :class:`ChiSquaredFitter <pymultifit.fitters.chiSquare_f.ChiSquareFitter>`
+     - ChiSquare fitter.
+   * - :class:`ExponentialFitter <pymultifit.fitters.exponential_f.ExponentialFitter>`
+     - Exponential fitter.
+   * - :class:`FoldedNormalFitter <pymultifit.fitters.foldedNormal_f.FoldedNormalFitter>`
+     - Folded Normal fitter.
+   * - :class:`GaussianFitter <pymultifit.fitters.gaussian_f.GaussianFitter>`
+     - Gaussian fitter.
+   * - :class:`HalfNormalFitter <pymultifit.fitters.halfNormal_f.HalfNormalFitter>`
+     - Half-Normal fitter.
+   * - :class:`LaplaceFitter <pymultifit.fitters.laplace_f.LaplaceFitter>`
+     - Laplace fitter.
+   * - :class:`LogNormalFitter <pymultifit.fitters.logNorm_f.LogNormalFitter>`
+     - Log-Normal fitter.
+   * - :class:`SkewNormalFitter <pymultifit.fitters.skewNorm_f.SkewNormalFitter>`
+     - Skew-Normal fitter.
+
+--------------------------------------------
+
+generators
+----------
+
+The `generators` module provides functions for generating single or multimodal data for specific distributions or mixed models.
+
+Base function
+^^^^^^^^^^^^^
+.. list-table::
+   :widths: 25 75
+   :header-rows: 1
+
+   * - Name
+     - Description
+   * - :class:`BaseFitter <pymultifit.fitters.backend.baseFitter.BaseFitter>`
+     - The base class for multi-fitting functionality.
+   * - :class:`MixedFitter <pymultifit.fitters.mixed_f.MixedFitter>`
+     - Mixed model fitting class.
+
+Derived Fitters
+^^^^^^^^^^^^^^^^
+
+.. list-table::
+   :widths: 25 75
+   :header-rows: 1
+
+   * - Name
+     - Description
[…]
