+++ sphinx_docs/source/tutorials.rst
@@ -1,16 +1,8 @@
 Tutorials
 =========
 
-
-The ``pyMultiFit`` library is divided into three modules,
-
-1. **distributions**, housing statistical (and non-statistical) distributions, such as ``GaussianDistribution``, ``LaplaceDistribution``, and more.
-2. **generators**, can generate single or n-model data based on the selected distribution. And finally
-3. **fitters**, that can fit n-model data on the given data.asdf
-
 .. toctree::
-   :maxdepth: 2
-   :hidden:
+   :maxdepth: 1
 
    generating_data
    generating_n_model_data
