### 2024-12-21 4449adc
[doc-minor-8-1]
1. updated `environment.yaml` and `requirements[dev].txt`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
