+++ environment.yaml
@@ -17,6 +17,7 @@ dependencies:
     - sphinx-rtd-theme
     - sphinx-copybutton
     - sphinxcontrib-napoleon
+    - numpydoc
     - nbsphinx
     - nbsphinx-link
     - sphinx-gallery
