### 2024-12-21 7d31093
[doc-minor-9]
1. separated `api_index.rst` to `_distributions.rst`, `_fitters.rst` and `_generators.rst`
2. minor changes to `generators.py`, and `index.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
