+++ sphinx_docs/source/_generators.rst
@@ -1,0 +1,42 @@
+``generators``
+==============
+
+Template function
+-----------------
+
+.. list-table::
+   :align: center
+   :header-rows: 1
+
+   * - Name
+     - Description
+   * - :meth:`multi_base <pymultifit.generators.generators.multi_base>`
+     - The template function for the generator functions.
+
+Derived generators
+------------------
+
+.. list-table::
+   :align: center
+   :header-rows: 1
+
+   * - Name
+     - Description
+   * - :meth:`multi_chi_squared <pymultifit.generators.generators.multi_chi_squared>`
+     - Chi-Squared distribution generator.
+   * - :meth:`multi_exponential <pymultifit.generators.generators.multi_exponential>`
+     - Exponential distribution generator.
+   * - :meth:`multi_folded_normal <pymultifit.generators.generators.multi_folded_normal>`
+     - Folded Normal distribution generator.
+   * - :meth:`multi_gaussian <pymultifit.generators.generators.multi_gaussian>`
+     - Gaussian distribution generator.
+   * - :meth:`multi_half_normal <pymultifit.generators.generators.multi_half_normal>`
+     - Half-Normal distribution generator.
+   * - :meth:`multi_laplace <pymultifit.generators.generators.multi_laplace>`
+     - Laplace distribution generator.
+   * - :meth:`multi_log_normal <pymultifit.generators.generators.multi_log_normal>`
+     - Log-Normal distribution generator.
+   * - :meth:`multi_skewed_normal <pymultifit.generators.generators.multi_skewed_normal>`
+     - Skewed Normal distribution generator.
+   * - :meth:`multiple_models <pymultifit.generators.generators.multiple_models>`
+     - Generator for data based on a combination of different models.
