+++ sphinx_docs/source/_fitters.rst
@@ -1,0 +1,42 @@
+``fitters``
+===========
+
+Base Class
+----------
+
+.. list-table::
+   :align: center
+   :header-rows: 1
+
+   * - Name
+     - Description
+   * - :class:`BaseFitter <pymultifit.fitters.backend.baseFitter.BaseFitter>`
+     - The base class for multi-fitting functionality.
+   * - :class:`MixedFitter <pymultifit.fitters.mixed_f.MixedFitter>`
+     - Mixed model fitting class.
+
+Derived Fitters
+---------------
+
+.. list-table::
+   :align: center
+   :header-rows: 1
+
+   * - Name
+     - Description
+   * - :class:`ChiSquaredFitter <pymultifit.fitters.chiSquare_f.ChiSquareFitter>`
+     - ChiSquare fitter.
+   * - :class:`ExponentialFitter <pymultifit.fitters.exponential_f.ExponentialFitter>`
+     - Exponential fitter.
+   * - :class:`FoldedNormalFitter <pymultifit.fitters.foldedNormal_f.FoldedNormalFitter>`
+     - Folded Normal fitter.
+   * - :class:`GaussianFitter <pymultifit.fitters.gaussian_f.GaussianFitter>`
+     - Gaussian fitter.
+   * - :class:`HalfNormalFitter <pymultifit.fitters.halfNormal_f.HalfNormalFitter>`
+     - Half-Normal fitter.
+   * - :class:`LaplaceFitter <pymultifit.fitters.laplace_f.LaplaceFitter>`
+     - Laplace fitter.
+   * - :class:`LogNormalFitter <pymultifit.fitters.logNorm_f.LogNormalFitter>`
+     - Log-Normal fitter.
+   * - :class:`SkewNormalFitter <pymultifit.fitters.skewNorm_f.SkewNormalFitter>`
+     - Skew-Normal fitter.
