+++ sphinx_docs/source/api_index.rst
@@ -1,0 +1,19 @@
+pyMultiFit API
+==============
+
+The Application Programming Interface (API) of `pyMultiFit` provides tools for statistical data generation and model fitting, organized into three primary layers:
+
+* **Distributions**: Statistical distributions for data modeling, generation, and fitting.
+* **Fitters**: Classes implementing fitting algorithms for statistical models.
+* **Generators**: Functions for generating synthetic datasets.
+
+.. toctree::
+   :hidden:
+
+   _distributions
+   _fitters
+   _generators
+   pymultifit
+   pymultifit.distributions
+   pymultifit.fitters
+   pymultifit.generators

[…]
