### 2024-12-22 69859de
[doc-major-3]
1. minor changes to `api_index.rst`, starting to understand api structure now
2. major overhaul to `arcSine_d.py`, might have to do this to all other distributions soon
3. added `arcSine_d.rst` for API
4. minor changes to `pymultifit.distributions.rst` and `pymultifit.rst`
5. added arcsine cdf and pdf to `utilities.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
