+++ sphinx_docs/source/distributions/arcSine_d.rst
@@ -1,0 +1,34 @@
+ArcSine Distribution
+====================
+
+The parent module for this distribution is: :mod:`pymultifit.distributions.arcSine_d`
+
+Class
+-----
+
+.. autoclass:: pymultifit.distributions.arcSine_d.ArcSineDistribution
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Backend function
+----------------
+
+.. autofunction:: pymultifit.distributions.utilities.beta_
+
+
+--------------------------------------------
+
+Full Import
+-----------
+
+.. code-block:: python
+
+   from pymultifit.distributions.arcSine_d import ArcSineDistribution
+
+Recommended Import
+------------------
+
+.. code-block:: python
+
+   from pymultifit.distributions import ArcSineDistribution
