+++ sphinx_docs/source/pymultifit.rst
@@ -5,7 +5,7 @@ Subpackages
 -----------
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 2
 
    pymultifit.distributions
    pymultifit.fitters
