### 2024-12-22 cf10ce2
[cdf_separation-major-1]
1. added a lot of rst files
2. name changes to a lot of distribution files
3. finalized `pydata-sphinx-theme` as final in `requirements[dev].txt` and `environment.yaml`
4. minor changes to `custom.css` and `conf.py`
5. added _cdf to `baseDistribution.py`
6. added `_distribution_api_runner.py` and `_distribution_api_template.py` for DISTRIBUTION template generation

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
