+++ src/pymultifit/distributions/logNormal_d.py
@@ -10,7 +10,7 @@ from .utilities import log_normal_
 
 
 class LogNormalDistribution(BaseDistribution):
-    """Class for Log-Normal distribution."""
+    """Class for LogNormal distribution."""
 
     def __init__(self, amplitude: float = 1., mean: float = 0., standard_deviation: float = 1., normalize: bool = False):
         if not normalize and amplitude <= 0:
