+++ src/pymultifit/distributions/beta_d.py
@@ -28,15 +28,6 @@ class BetaDistribution(BaseDistribution):
     def _pdf(self, x: np.ndarray) -> np.ndarray:
         return beta_(x, amplitude=self.amplitude, alpha=self.alpha, beta=self.beta, normalize=self.norm)
 
-    def pdf(self, x: np.ndarray) -> np.ndarray:
-        y = np.zeros_like(x, dtype=float)
-        mask_ = np.logical_and(x > 0, x < 1)
-        y[mask_] = self._pdf(x[mask_])
-
-        # hack to match beta distribution at x = 0 and x = 1
-        y[np.logical_or(x == 0, x == 1)] = np.inf
-        return y
-
     def cdf(self, x: np.ndarray) -> np.ndarray:
         y = np.zeros_like(x, dtype=float)
         mask_ = np.logical_and(x > 0, x < 1)
