### 2024-12-23 0798eae
[cdf_separation-minor-1]
1. updated `beta_d.py` and `beta_d.rst` with more functionalities
2. minor changes in some rst files
3. reverted `intersphinx_mapping` in `conf.py`
4. added more tests in `test_beta_d.py`
5. more functionality in `utilities.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
