+++ src/pymultifit/distributions/utilities.py
@@ -1,10 +1,14 @@
 """Created on Aug 03 17:13:21 2024"""
 
-__all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'beta_', 'chi_squared_', 'exponential_', 'folded_normal_', 'gamma_sr_', 'gamma_ss_', 'gaussian_',
-           'half_normal_', 'integral_check', 'laplace_', 'log_normal_', 'norris2005', 'norris2011', 'power_law_', 'uniform_']
+__all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'beta_pdf_', 'beta_cdf_', 'beta_logpdf_', 'chi_squared_', 'exponential_', 'folded_normal_', 'gamma_sr_',
+           'gamma_ss_', 'gaussian_', 'half_normal_', 'integral_check', 'laplace_', 'log_normal_', 'norris2005', 'norris2011', 'power_law_',
+           'uniform_']
 
 import numpy as np
-from scipy.special import gamma
+from custom_inherit import doc_inherit
+from scipy.special import betainc, gamma, gammaln
+
+doc_style = 'numpy_napoleon_with_merge'
 
 
 def integral_check(pdf_function, x_range: tuple) -> float:
@@ -31,12 +35,8 @@ def integral_check(pdf_function, x_range: tuple) -> float:
 def arc_sine_pdf_(x: np.array,
                   amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0,
                   normalize: bool = False) -> np.array:
-    """
+    r"""
     Compute the probability density function (PDF) of the Arcsine distribution.
-
-    The Arcsine distribution is defined over a specified interval `[loc, loc + scale]`
-    and can be optionally normalized or scaled with a custom amplitude. Values
-    outside the interval are assigned a PDF of 0.
 
     Parameters
     ----------
@@ -49,7 +49,7 @@ def arc_sine_pdf_(x: np.array,
     scale : float, optional
         The scale parameter, specifying the width of the distribution. Defaults to 1.0.
     normalize : bool, optional
-        If True, normalize the PDF so that its integral over `[loc, loc + scale]` is 1.0.
+        If True, the distribution will be normalized such that the total area under the PDF equals 1.
         Defaults to False.
 
     Returns
@@ -62,16 +62,9 @@ def arc_sine_pdf_(x: np.array,
     - The standard Arcsine distribution is defined on `[0, 1]` with a PDF of:
 
       .. math::
-         f(x) = \\frac{1}{\\pi \\sqrt{x(1-x)}}
+         f(x) = \frac{1}{\pi \sqrt{x(1-x)}}
 
-      This implementation generalizes it to an interval `[loc, loc + scale]` using
-      a location-scale transformation.
-
-    - Numerical stability is ensured by adding a small epsilon (`1e-10`) to the denominator
-      to avoid division by zero at the boundaries of the interval.
-
-    - The PDF is scaled by `1 / scale` to ensure proper normalization for the location-scale
-      transformation.
+      This implementation generalizes it to an interval `[loc, loc + scale]` using a location-scale transformation.
     """
     x = (x - loc) / scale
 
@@ -135,34 +128,57 @@ def arc_sine_cdf_(x: np.array, loc: float = 0.0, scale: float = 1.0) -> np.array
     return cdf
 
 
-def beta_(x: np.ndarray,
-          amplitude: float = 1., alpha: float = 1., beta: float = 1.,
-          normalize: bool = False) -> np.ndarray:
+def beta_pdf_(x: np.array,
+              amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0, loc: float = 0.0, scale: float = 1.0,
+              normalize: bool = False) -> np.array:
     """
-    Compute the beta probability density function (PDF).
[…]
