+++ src/tests/test_distributions/test_beta_d.py
@@ -5,6 +5,7 @@ import pytest
 from scipy.special import betaincinv
 from scipy.stats import beta
 
+from pymultifit import EPSILON
 from ...pymultifit.distributions import BetaDistribution
 from ...pymultifit.distributions.backend import errorHandling as erH
 
@@ -58,34 +59,72 @@ class TestBetaDistribution:
         assert d_stats["variance"] == (a * b) / ((a + b)**2 * (a + b + 1))
 
     @staticmethod
-    def test_stats_scipy():
-        for _ in range(50):  # Run 50 random tests
-            # Generate random alpha and beta values
-            a_ = np.random.uniform(0.1, 5.0)  # Alpha (shape1) > 0
-            b_ = np.random.uniform(0.1, 5.0)  # Beta (shape2) > 0
+    def test_pdf_cdf_logpdf_scipy():
+        x_ = np.linspace(start=-3, stop=3, num=50)
+        loc, scale = 0, 1
 
-            dist_ = BetaDistribution(amplitude=1.0, alpha=a_, beta=b_, normalize=True)
-            a, b = dist_.alpha, dist_.beta
+        def _multipy(dist, func_type):
+            return dist.pdf if func_type == 'pdf' else dist.cdf if func_type == 'cdf' else dist.logpdf
 
-            # Get custom stats
-            d_stats = dist_.stats()
+        def _scipy(a_, b_, loc_, scale_, func_type):
+            return [beta(a=a_, b=b_, loc=loc_, scale=scale_).pdf(x_) if func_type == 'pdf' else
+                    beta(a=a_, b=b_, loc=loc_, scale=scale_).cdf(x_) if func_type == 'cdf' else
+                    beta(a=a_, b=b_, loc=loc_, scale=scale_).logpdf(x_)][0]
 
-            # Get stats from scipy
-            scipy_mean, scipy_variance = beta.stats(a=a, b=b, moments="mv")
-            scipy_median = beta.median(a=a, b=b)
-            scipy_mode = None
-            if a > 1 and b > 1:
-                scipy_mode = (a - 1) / (a + b - 2)
+        for func_ in ['pdf', 'cdf', 'logpdf']:
+            for _ in range(50):  # Run 50 random tests
+                alpha_ = np.random.uniform(low=EPSILON, high=5.0)
 
-            # Perform assertions
-            assert np.isclose(d_stats["mean"], scipy_mean, rtol=1e-5,
-                              atol=1e-8), f"Mean mismatch: {d_stats['mean']} != {scipy_mean} for alpha={a}, beta={b}"
-            assert np.isclose(d_stats["median"], scipy_median, rtol=1e-5,
-                              atol=1e-8), f"Median mismatch: {d_stats['median']} != {scipy_median} for alpha={a}, beta={b}"
-            if a > 1 and b > 1:
-                assert np.isclose(d_stats["mode"], scipy_mode, rtol=1e-5,
-                                  atol=1e-8), f"Mode mismatch: {d_stats['mode']} != {scipy_mode} for alpha={a}, beta={b}"
-            elif np.logical_or(a <= 0, b <= 0):
-                assert d_stats.get("mode", []) == [], f"Unexpected mode for invalid parameters: {d_stats['mode']} for alpha={a}, beta={b}"
-            assert np.isclose(d_stats["variance"], scipy_variance, rtol=1e-5,
-                              atol=1e-8), f"Variance mismatch: {d_stats['variance']} != {scipy_variance} for alpha={a}, beta={b}"
+                for case in [1, 2, 3]:
+                    if case == 1:
+                        beta_ = alpha_
[…]
