+++ sphinx_docs/source/distributions/skewNormal_d.rst
@@ -1,5 +1,5 @@
 SkewNormal Distribution
-====================
+=======================
 
 The parent module for this distribution is: :mod:`pymultifit.distributions.skewNormal_d`.
 
