### 2024-12-23 7fdcb92
[cdf_separation-minor-2]
1. added toctree to `_distributions.rst`
2. removed toctree from `api_index.rst`
3. minor changes in `arcSine_d.py`, `baseDistribution.py`, and `conf.py`
4. added some more tests in `test_arcSine_d.py`
5. added log_pdf for arcsine in `utilities.py`
