+++ src/pymultifit/distributions/utilities.py
@@ -36,35 +36,41 @@ def arc_sine_pdf_(x: np.array,
                   amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0,
                   normalize: bool = False) -> np.array:
     r"""
-    Compute the probability density function (PDF) of the Arcsine distribution.
+    Compute the probability density function (PDF) of the ArcSine distribution.
 
     Parameters
     ----------
     x : np.array
         Input array of values where the PDF is evaluated.
     amplitude : float, optional
-        The amplitude of the PDF. Defaults to 1.0. Ignored if `normalize` is True.
+        The amplitude of the PDF. Defaults to 1.0.
+        Ignored if **normalize** is ``True``.
     loc : float, optional
-        The location parameter, specifying the lower bound of the distribution. Defaults to 0.0.
+        The location parameter, specifying the lower bound of the distribution.
+        Defaults to 0.0.
     scale : float, optional
-        The scale parameter, specifying the width of the distribution. Defaults to 1.0.
+        The scale parameter, specifying the width of the distribution.
+        Defaults to 1.0.
     normalize : bool, optional
-        If True, the distribution will be normalized such that the total area under the PDF equals 1.
-        Defaults to False.
+        If True, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
 
     Returns
     -------
     np.array
-        Array of the same shape as `x`, containing the evaluated PDF values.
+        Array of the same shape as :math:`x`, containing the evaluated PDF values.
 
     Notes
     -----
-    - The standard Arcsine distribution is defined on `[0, 1]` with a PDF of:
+    The ArcSine PDF is defined as:
 
-      .. math::
-         f(x) = \frac{1}{\pi \sqrt{x(1-x)}}
+    .. math:: f(y) = \frac{1}{\pi \sqrt{y(1-y)}}
 
-      This implementation generalizes it to an interval `[loc, loc + scale]` using a location-scale transformation.
+    where, :math:`y` is a transformed value of :math:`x`, defined as:
+
+    .. math:: y = \frac{x - \text{loc}}{\text{scale}}
+
+    The final PDF is expressed as :math:`f(y)/\text{scale}`
     """
     x = (x - loc) / scale
 
@@ -84,37 +90,34 @@ def arc_sine_pdf_(x: np.array,
     return pdf_ / scale
 
 
-def arc_sine_cdf_(x: np.array, loc: float = 0.0, scale: float = 1.0) -> np.array:
-    """
-    Compute the Cumulative Distribution Function (CDF) of the Arcsine distribution.
-
-    The standard Arcsine distribution is defined on the interval [0, 1] with a CDF of:
-
-    .. math::
-       F(x) = \\frac{2}{\\pi}\\arcsin\\left[\\sqrt{x}\\right]
-
-    This implementation generalizes the distribution to an interval [loc, loc + scale] using a
-    location-scale transformation.
+@doc_inherit(parent=arc_sine_pdf_, style=doc_style)
+def arc_sine_cdf_(x: np.array,
+                  amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0,
+                  normalize: bool = False) -> np.array:
+    r"""
+    Compute the cumulative density function (CDF) of the ArcSine distribution.
 
     Parameters
     ----------
-    x : np.array
-        Input array of values where the CDF is evaluated.
-    loc : float, optional
[…]
