+++ src/tests/test_distributions/test_arcSine_d.py
@@ -1,7 +1,7 @@
 """Created on Dec 15 19:24:18 2024"""
 
 import numpy as np
-from scipy.stats import beta
+from scipy.stats import arcsine, beta
 
 from ...pymultifit.distributions.arcSine_d import ArcSineDistribution
 
@@ -24,18 +24,22 @@ class TestArcSineDistribution:
 
     @staticmethod
     def test_pdf_cdf():
-        x1 = np.linspace(-0.5, 1.5, 1000)
+        x1 = np.linspace(-0.5, 1.5, 10)
         x2 = np.array([0, 1])
 
         for x in [x1, x2]:
             distribution = ArcSineDistribution(normalize=True)
             pdf_custom = distribution.pdf(x)
             cdf_custom = distribution.cdf(x)
+            logpdf_custom = distribution.logpdf(x)
 
-            pdf_scipy = beta.pdf(x, a=0.5, b=0.5)
-            cdf_scipy = beta.cdf(x, a=0.5, b=0.5)
+            pdf_scipy = arcsine.pdf(x)
+            cdf_scipy = arcsine.cdf(x)
+            logpdf_scipy = arcsine.logpdf(x)
 
             np.testing.assert_allclose(pdf_custom, pdf_scipy, rtol=1e-5, atol=1e-8,
                                        err_msg="PDF does not match SciPy Beta(0.5, 0.5).")
             np.testing.assert_allclose(cdf_custom, cdf_scipy, rtol=1e-5, atol=1e-8,
                                        err_msg="CDF does not match SciPy Beta(0.5, 0.5).")
+            np.testing.assert_allclose(logpdf_custom, logpdf_scipy, rtol=1e-5, atol=1e-8,
+                                       err_msg="logPDF does not match SciPy Beta(0.5, 0.5).")
