+++ sphinx_docs/source/api_index.rst
@@ -10,6 +10,7 @@ The Application Programming Interface (API) of `pyMultiFit` provides tools for s
 The documentation first goes through a birdseye view for these three modules, followed by detailed documentation for each class and function.
 
 .. toctree::
+   :maxdepth: 2
    :hidden:
 
    _distributions
@@ -19,15 +20,3 @@ The documentation first goes through a birdseye view for these three modules, fo
    pymultifit.distributions
    pymultifit.fitters
    pymultifit.generators
-   distributions/arcSine_d
-   distributions/beta_d
-   distributions/chiSquare_d
-   distributions/exponential_d
-   distributions/foldedNormal_d
-   distributions/gamma_d
-   distributions/gaussian_d
-   distributions/halfNormal_d
-   distributions/laplace_d
-   distributions/logNormal_d
-   distributions/skewNormal_d
-   distributions/uniform_d

[…]
