### 2024-12-23 ba56cdb
[cdf_separation-minor-2-1]
1. minor correction in `test_beta_d.py` for CodeCOV CI
