### 2024-12-23 1c19847
[cdf_separation-minor-2-2]
1. minor changes in ArcSine
