+++ src/pymultifit/distributions/arcSine_d.py
@@ -1,11 +1,13 @@
-"""Created on Aug 14 02:02:42 2024"""
+r"""
+:strong:`pymultifit.distributions.arcSine_d.`
+"""
 
 from typing import Any, Dict
 
 import numpy as np
 
 from .backend import BaseDistribution
-from .utilities import arc_sine_cdf_, arc_sine_log_pdf, arc_sine_pdf_
+from .utilities import arc_sine_cdf_, arc_sine_logpdf_, arc_sine_pdf_
 
 
 class ArcSineDistribution(BaseDistribution):
@@ -25,7 +27,7 @@ class ArcSineDistribution(BaseDistribution):
         return arc_sine_cdf_(x=x, loc=self.loc, scale=self.scale)
 
     def logpdf(self, x: np.array) -> np.array:
-        return arc_sine_log_pdf(x=x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
+        return arc_sine_logpdf_(x=x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
 
     def stats(self) -> Dict[str, Any]:
         mean_ = 0.5
