### 2024-12-23 70666d5
[cdf_separation-minor-3]
1. adopted inheritance for major distributions
2. minor changes here and there
