+++ src/pymultifit/distributions/chiSquare_d.py
@@ -3,10 +3,10 @@
 from typing import Dict
 
 from .backend import errorHandling as erH
-from .gamma_d import GammaDistributionSS
+from .gamma_d import GammaDistributionSR
 
 
-class ChiSquareDistribution(GammaDistributionSS):
+class ChiSquareDistribution(GammaDistributionSR):
     """Class for ChiSquare distribution."""
 
     def __init__(self, amplitude: float = 1., degree_of_freedom: int = 1, normalize: bool = False):
@@ -15,7 +15,7 @@ class ChiSquareDistribution(GammaDistributionSS):
         elif not isinstance(degree_of_freedom, int) or degree_of_freedom <= 0:
             raise erH.DegreeOfFreedomError()
         self.dof = degree_of_freedom
-        super().__init__(amplitude=amplitude, shape=degree_of_freedom / 2., scale=2., normalize=normalize)
+        super().__init__(amplitude=amplitude, shape=degree_of_freedom / 2., rate=0.5, normalize=normalize)
 
     def stats(self) -> Dict[str, float]:
         stat_ = super().stats()
