+++ src/pymultifit/distributions/gamma_d.py
@@ -3,10 +3,9 @@
 from typing import Dict
 
 import numpy as np
-from scipy.special import gammainc
 
 from .backend import BaseDistribution, errorHandling as erH
-from .utilities import gamma_sr_
+from .utilities import gamma_sr_cdf_, gamma_sr_pdf_
 
 
 class GammaDistributionSR(BaseDistribution):
@@ -28,13 +27,10 @@ class GammaDistributionSR(BaseDistribution):
     def _pdf(self, x: np.ndarray) -> np.ndarray:
         if np.any(x < 0):
             raise erH.XOutOfRange()
-        return gamma_sr_(x, amplitude=self.amplitude, shape=self.shape, rate=self.rate, normalize=self.norm)
+        return gamma_sr_pdf_(x, amplitude=self.amplitude, shape=self.shape, rate=self.rate, normalize=self.norm)
 
-    def pdf(self, x: np.ndarray) -> np.ndarray:
-        return np.nan_to_num(x=self._pdf(x), copy=False, nan=0)
-
-    def cdf(self, x: np.ndarray) -> np.ndarray:
-        return np.nan_to_num(x=gammainc(self.shape, self.rate * x), copy=False, nan=0)
+    def _cdf(self, x: np.ndarray) -> np.ndarray:
+        return gamma_sr_cdf_(x, amplitude=self.amplitude, shape=self.shape, rate=self.rate, normalize=self.norm)
 
     def stats(self) -> Dict[str, float]:
         a, b = self.shape, self.rate
@@ -50,6 +46,7 @@ class GammaDistributionSR(BaseDistribution):
 
 class GammaDistributionSS(GammaDistributionSR):
     """Class for Gamma distribution with shape and scale parameters."""
+
     def __init__(self, amplitude: float = 1., shape: float = 1., scale: float = 1., normalize: bool = False):
         self.scale = scale
         if not normalize and amplitude <= 0:
