+++ src/pymultifit/distributions/arcSine_d.py
@@ -1,16 +1,9 @@
-r"""
-:strong:`pymultifit.distributions.arcSine_d.`
-"""
+"""Created on Aug 14 02:02:42 2024"""
 
-from typing import Any, Dict
-
-import numpy as np
-
-from .backend import BaseDistribution
-from .utilities import arc_sine_cdf_, arc_sine_logpdf_, arc_sine_pdf_
+from . import BetaDistribution
 
 
-class ArcSineDistribution(BaseDistribution):
+class ArcSineDistribution(BetaDistribution):
     """Class for ArcSine distribution."""
 
     def __init__(self, amplitude: float = 1., loc: float = 0.0, scale: float = 1.0, normalize: bool = False):
@@ -19,22 +12,4 @@ class ArcSineDistribution(BaseDistribution):
         self.scale = scale
 
         self.norm = normalize
-
-    def _pdf(self, x: np.array) -> np.array:
-        return arc_sine_pdf_(x=x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
-
-    def _cdf(self, x: np.array) -> np.array:
-        return arc_sine_cdf_(x=x, loc=self.loc, scale=self.scale)
-
-    def logpdf(self, x: np.array) -> np.array:
-        return arc_sine_logpdf_(x=x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
-
-    def stats(self) -> Dict[str, Any]:
-        mean_ = 0.5
-        median_ = 0.5
-        variance_ = 1 / 8
-
-        return {'mean': mean_,
-                'median': median_,
-                'model': [],
-                'variance': variance_}
+        super().__init__(amplitude=self.amplitude, alpha=0.5, beta=0.5, loc=self.loc, scale=self.scale, normalize=self.norm)
