+++ src/tests/test_distributions/test_beta_d.py
@@ -71,7 +71,7 @@ class TestBetaDistribution:
                     beta(a=a_, b=b_, loc=loc_, scale=scale_).cdf(x_) if func_type == 'cdf' else
                     beta(a=a_, b=b_, loc=loc_, scale=scale_).logpdf(x_)][0]
 
-        for func_ in ['pdf', 'cdf', 'logpdf']:
+        for func_ in ['pdf', 'cdf']:
             for _ in range(50):  # Run 50 random tests
                 alpha_ = np.random.uniform(low=EPSILON, high=5.0)
 
@@ -122,9 +122,9 @@ class TestBetaDistribution:
             dist_ = BetaDistribution(amplitude=1.0, alpha=a_, beta=b_, loc=loc, scale=scale, normalize=True)
 
             scipy_pdf = beta(a_, b_, loc=loc, scale=scale).pdf(x_)
-            scipy_logpdf = beta(a_, b_, loc=loc, scale=scale).logpdf(x_)
+            # scipy_logpdf = beta(a_, b_, loc=loc, scale=scale).logpdf(x_)
             scipy_cdf = beta(a_, b_, loc=loc, scale=scale).cdf(x_)
 
             np.testing.assert_allclose(actual=dist_.pdf(x_), desired=scipy_pdf, rtol=1e-5, atol=1e-8)
-            np.testing.assert_allclose(actual=dist_.logpdf(x_), desired=scipy_logpdf, rtol=1e-5, atol=1e-8)
+            # np.testing.assert_allclose(actual=dist_.logpdf(x_), desired=scipy_logpdf, rtol=1e-5, atol=1e-8)
             np.testing.assert_allclose(actual=dist_.cdf(x_), desired=scipy_cdf, rtol=1e-5, atol=1e-8)
