### 2024-12-24 66f6064
[cdf_separation-minor-4]
1. minor changes
2. added `substitutions.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
