+++ rough_docstring.txt
@@ -1,22 +1,22 @@
 PFD generic docstring
 """""""""""""""""""""
 
-r"""
-Compute PDF for {DISTRIBUTION}.
+    r"""
+    Compute PDF for the {DISTRIBUTION} Distribution.
 
-Parameters
-----------
-x : np.array
-    Input array of values where the PDF is evaluated.
-amplitude : float, optional
-    The amplitude of the PDF. Defaults to 1.0.
-    Ignored if **normalize** is ``True``.
-normalize : bool, optional
-    If True, the distribution is normalized so that the total area under the PDF equals 1.
-    Defaults to ``False``.
+    Parameters
+    ----------
+    x : np.array
+        Input array of values where the PDF is evaluated.
+    amplitude : float, optional
+        The amplitude of the PDF. Defaults to 1.0.
+        Ignored if **normalize** is ``True``.
+    normalize : bool, optional
+        If True, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
 
-Returns
--------
-np.array
-    Array of the same shape as :math:`x`, containing the evaluated values.
-"""
+    Returns
+    -------
+    np.array
+        Array of the same shape as :math:`x`, containing the evaluated values.
+    """
