### 2024-12-24 274ca39
[cdf_separation-minor-5]
1. added loc to `gamma_d.py`
2. updated `rough_docstring.txt` for better copy/pasting
3. removed `substitutions.rst`
4. added loc testing in `test_gamma_sr_d.py`
5. updated `utilities.py`
