+++ src/pymultifit/distributions/utilities.py
@@ -59,7 +59,7 @@ def arc_sine_pdf_(x: np.array,
         The amplitude of the PDF. Defaults to 1.0.
         Ignored if **normalize** is ``True``.
     loc : float, optional
-        The location parameter, specifying the lower bound of the distribution.
+        The location parameter specifying the lower bound of the distribution.
         Defaults to 0.0.
     scale : float, optional
         The scale parameter, specifying the width of the distribution.
@@ -85,21 +85,7 @@ def arc_sine_pdf_(x: np.array,
 
     The final PDF is expressed as :math:`f(y)/\text{scale}`
     """
-    y = (x - loc) / scale
-
-    un_normalized = 1 / np.sqrt(y * (1 - y))
-
-    if normalize:
-        normalization_factor = np.pi
-        amplitude = 1
-    else:
-        normalization_factor = 1
-
-    pdf_ = amplitude * (un_normalized / normalization_factor)
-    mask = np.logical_and(y >= 0, y <= 1)
-    pdf_[~mask] = 0
-
-    return pdf_ / scale
+    return beta_pdf_(x=x, amplitude=amplitude, alpha=0.5, beta=0.5, loc=loc, scale=scale, normalize=normalize)
 
 
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
@@ -131,12 +117,7 @@ def arc_sine_cdf_(x: np.array,
 
     .. math:: y = \frac{x - \text{loc}}{\text{scale}}
     """
-    y = (x - loc) / scale
-    cdf = np.zeros_like(a=y, dtype=float)
-    mask = np.logical_and(y >= 0, y <= 1)
-    cdf[mask] = (2 / np.pi) * np.arcsin(np.sqrt(y[mask]))
-    cdf[y > 1] = 1.0
-    return cdf
+    return beta_cdf_(x=x, amplitude=amplitude, alpha=0.5, beta=0.5, loc=loc, scale=scale, normalize=normalize)
 
 
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
@@ -151,34 +132,14 @@ def arc_sine_logpdf_(x: np.array,
     np.array
         Array of the same shape as `x`, containing the evaluated log-PDF values.
     """
-    if x.size == 0:
-        return np.array([])
-
-    y = (x - loc) / scale
-    logpdf_ = np.full_like(a=y, fill_value=-np.inf, dtype=float)
-
-    mask_ = np.logical_and(y > 0, y < 1)
-
-    log_numerator = 0.5 * np.log(y) + 0.5 * np.log(1 - y)
-
-    if normalize:
-        log_normalization = np.log(np.pi)  # Log of normalization constant
-        amplitude = 1.0
-    else:
-        log_normalization = 0.0
-
-    logpdf_[mask_] = np.log(amplitude) - log_numerator[mask_] - log_normalization
-    logpdf_[y == 0] = np.inf
-    logpdf_[y == 1] = np.inf
-
-    return logpdf_ - np.log(scale)
+    return beta_logpdf_(x=x, amplitude=amplitude, alpha=0.5, beta=0.5, loc=loc, scale=scale, normalize=normalize)
 
 
 def beta_pdf_(x: np.array,
               amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0, loc: float = 0.0, scale: float = 1.0,
               normalize: bool = False) -> np.array:
-    """
-    Compute PDF of the Beta distribution.
+    r"""
+    Compute PDF of the :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
 
     Parameters
     ----------
@@ -192,7 +153,7 @@ def beta_pdf_(x: np.array,
     beta : float, optional
         The :math:`\beta` parameter. Default is 1.0.
[…]
