+++ src/pymultifit/distributions/gamma_d.py
@@ -11,7 +11,8 @@ from .utilities import gamma_sr_cdf_, gamma_sr_pdf_
 class GammaDistributionSR(BaseDistribution):
     """Class for Gamma distribution with shape and rate parameters."""
 
-    def __init__(self, amplitude: float = 1., shape: float = 1., rate: float = 1., normalize: bool = False):
+    def __init__(self, amplitude: float = 1., shape: float = 1., rate: float = 1., loc: float = 0.0,
+                 normalize: bool = False):
         if not normalize and amplitude <= 0:
             raise erH.NegativeAmplitudeError()
         elif shape <= 0:
@@ -21,16 +22,19 @@ class GammaDistributionSR(BaseDistribution):
         self.amplitude = 1. if normalize else amplitude
         self.shape = shape
         self.rate = rate
+        self.loc = loc
 
         self.norm = normalize
 
-    def _pdf(self, x: np.ndarray) -> np.ndarray:
+    def _pdf(self, x: np.array) -> np.array:
         if np.any(x < 0):
             raise erH.XOutOfRange()
-        return gamma_sr_pdf_(x, amplitude=self.amplitude, shape=self.shape, rate=self.rate, normalize=self.norm)
+        return gamma_sr_pdf_(x,
+                             amplitude=self.amplitude, shape=self.shape, rate=self.rate, loc=self.loc, normalize=self.norm)
 
-    def _cdf(self, x: np.ndarray) -> np.ndarray:
-        return gamma_sr_cdf_(x, amplitude=self.amplitude, shape=self.shape, rate=self.rate, normalize=self.norm)
+    def _cdf(self, x: np.array) -> np.array:
+        return gamma_sr_cdf_(x,
+                             amplitude=self.amplitude, shape=self.shape, rate=self.rate, normalize=self.norm)
 
     def stats(self) -> Dict[str, float]:
         a, b = self.shape, self.rate
