+++ sphinx_docs/source/substitutions.rst
@@ -1,8 +1,0 @@
-.. |utilities| replace:: :mod:`~pymultifit.distributions.utilities`
-
-.. |beta_pdf_| replace:: :mod:`~pymultifit.distributions.utilities.beta_pdf_`
-.. |beta_cdf_| replace:: :mod:`~pymultifit.distributions.utilities.beta_cdf_`
-.. |beta_logpdf_| replace:: :mod:`~pymultifit.distributions.utilities.beta_logpdf_`
-
-.. |gaussian_pdf_| replace:: :mod:`~pymultifit.distributions.utilities.gaussian_pdf_`
-.. |gaussian_cdf_| replace:: :mod:`~pymultifit.distributions.utilities.gaussian_cdf_`

[…]
