+++ src/pymultifit/distributions/chiSquare_d.py
@@ -9,13 +9,16 @@ from .gamma_d import GammaDistributionSR
 class ChiSquareDistribution(GammaDistributionSR):
     """Class for ChiSquare distribution."""
 
-    def __init__(self, amplitude: float = 1., degree_of_freedom: int = 1, normalize: bool = False):
+    def __init__(self,
+                 amplitude: float = 1.0, degree_of_freedom: int = 1, loc: float = 0.0,
+                 normalize: bool = False):
         if not normalize and amplitude <= 0:
             raise erH.NegativeAmplitudeError()
         elif not isinstance(degree_of_freedom, int) or degree_of_freedom <= 0:
             raise erH.DegreeOfFreedomError()
         self.dof = degree_of_freedom
-        super().__init__(amplitude=amplitude, shape=degree_of_freedom / 2., rate=0.5, normalize=normalize)
+        self.loc = loc
+        super().__init__(amplitude=amplitude, shape=degree_of_freedom / 2., rate=0.5, loc=loc, normalize=normalize)
 
     def stats(self) -> Dict[str, float]:
         stat_ = super().stats()
