+++ src/pymultifit/distributions/utilities.py
@@ -4,7 +4,7 @@ __all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'arc_sine_logpdf_',
            'beta_pdf_', 'beta_cdf_', 'beta_logpdf_',
            'chi_square_pdf_',
            'exponential_pdf_',
-           'folded_normal_',
+           'folded_normal_pdf_',
            'gamma_sr_pdf_', 'gamma_sr_cdf_',
            'gamma_ss_',
            'gaussian_pdf_',
@@ -387,47 +387,44 @@ def exponential_pdf_(x: np.array,
     return gamma_sr_pdf_(x, amplitude=amplitude, shape=1., rate=scale, normalize=normalize)
 
 
-def folded_normal_(x: np.array,
-                   amplitude: float = 1., mu: float = 0.0, variance: float = 1.0,
-                   normalize: bool = False) -> np.array:
-    # """
-    # Compute the folded half-normal distribution.
-    #
-    # The folded half-normal distribution is the sum of two Gaussian distributions mirrored around `mu`.
-    # It is defined as the sum of a normal distribution centered at `mu` and another mirrored at `-mu`.
-    #
-    # Parameters
-    # ----------
-    # x : np.array
-    #     Input array where the folded half-normal distribution is evaluated.
-    # amplitude : float, optional
-    #     The amplitude of the distribution. Defaults to 1. Ignored if `normalize` is set to True.
-    # mu : float, optional
-    #     The mean (`mu`) of the original normal distribution. Defaults to 0.0.
-    # variance : float, optional
-    #     The standard deviation (`sigma`) of the original normal distribution. Defaults to 1.0.
-    # normalize : bool, optional
-    #     If True, the distribution will be normalized so that the PDF is at most 1. Defaults to False.
-    #
-    # Returns
-    # -------
-    # np.array
-    #     The computed folded half-normal distribution values for the input `x`.
-    #
-    # Notes
-    # -----
-    # The folded half-normal distribution is defined as the sum of two Gaussian
-    # distributions:
-    # .. math::
-    #     f(x; \\mu, \\sigma) = g_1(x; \\mu, \\sigma) + g_2(x; -\\mu, \\sigma)
-    #
-    # where `g_1` and `g_2` are the Gaussian distributions with the specified parameters.
-    # """
+def folded_normal_pdf_(x: np.array,
+                       amplitude: float = 1., mu: float = 0.0, variance: float = 1.0,
+                       normalize: bool = False) -> np.array:
+    r"""
+    Compute PDF for :class:`~pymultifit.distributions.foldedNormal_d.FoldedNormalDistribution`.
+
+    Parameters
+    ----------
+    x : np.array
+        Input array of values where the PDF is evaluated.
+    amplitude : float, optional
+        The amplitude of the PDF. Defaults to 1.0.
+        Ignored if **normalize** is ``True``.
+    mu : float, optional
+        The mean parameter, :math:`\mu`. Defaults to 0.0.
+    variance : float, optional
+        The variance parameter, :math:`\sigma^2`. Defaults to 1.0.
+    normalize : bool, optional
+        If True, the distribution will be normalized so that the PDF is at most 1. Defaults to False.
+
+    Returns
+    -------
+    np.array
+        The computed folded half-normal distribution values for the input `x`.
+
+    Notes
+    -----
[…]
