+++ sphinx_docs/source/distributions/exponential_d.rst
@@ -15,8 +15,8 @@ The parent module for this distribution is: :mod:`pymultifit.distributions.expon
 
 This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities` module:
 
-* :mod:`~pymultifit.distributions.utilities.gamma_sr_pdf_`
-* :mod:`~pymultifit.distributions.utilities.gamma_sr_cdf_`
+* :func:`~pymultifit.distributions.utilities.gamma_sr_pdf_`
+* :func:`~pymultifit.distributions.utilities.gamma_sr_cdf_`
 
 Recommended Import
 ------------------
