+++ sphinx_docs/source/distributions/gaussian_d.rst
@@ -1,5 +1,3 @@
-.. include:: ./../substitutions.rst
-
 Gaussian Distribution
 =====================
 
@@ -8,10 +6,10 @@ The parent module for this distribution is: :mod:`pymultifit.distributions.gauss
 .. autoclass:: pymultifit.distributions.gaussian_d.GaussianDistribution
    :no-members:
 
-This class internally utilizes the following functions from the |utilities| module:
+This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities` module:
 
-* |gaussian_pdf_|
-* |gaussian_cdf_|
+* :func:`~pymultifit.distributions.utilities.gaussian_pdf_`
+* :func:`~pymultifit.distributions.utilities.gaussian_cdf_`
 
 Full Import
 -----------
