### 2024-12-24 9bcde5f
[cdf_separation-minor-6]
1. added gamma distribution to `_distributions.rst`
2. added `gamma_sr_d.rst` and `gamma_ss_d.rst`
3. minor modifications in `beta_d.rst`, `gaussian_d.py`, `gaussian_d.rst`, `laplace_f.py`, and `pymultifit.distributions.rst`
4. updated `laplace_d.py` wrt `utilities.py`
5. updated tests for `test_beta_d.py`, `test_foldedNormal_d.py` and `test_gaussian_d.py`
6. updated `utilities.py`
