+++ src/tests/test_distributions/test_gaussian_d.py
@@ -61,8 +61,46 @@ class TestGaussianDistribution:
             for _ in range(50):  # Run 50 random tests
                 loc_ = np.random.uniform(low=-10, high=10)
                 scale_ = np.random.uniform(low=0.1, high=5)
-                x = np.random.uniform(low=loc_ - 10, high=loc_ + 10, size=50)
+                x = np.linspace(start=loc_ - 10, stop=loc_ + 10, num=50)
                 distribution = GaussianDistribution(mean=loc_, standard_deviation=scale_, normalize=True)
                 expected = _cdf_pdf_scipy(x_=x, loc=loc_, scale=scale_, what=i)
                 np.testing.assert_allclose(actual=_cdf_pdf_custom(x_=x, dist_=distribution, what=i),
                                            desired=expected, rtol=1e-5, atol=1e-8)
+
+    @staticmethod
+    def test_gaussian_edge_cases():
+        x_ = np.linspace(start=-10, stop=10, num=100)
+
+        # Edge cases
+        edge_cases = [
+            (0, 1),  # Standard normal distribution (mean=0, std=1)
+            (0, 0.1),  # Narrow peak at center (small std)
+            (0, 10),  # Broad distribution (large std)
+            (5, 1),  # Mean shifted to the right
+            (-5, 1),  # Mean shifted to the left
+        ]
+
+        # Extreme cases
+        extreme_cases = [
+            (0, 1e-5),  # Very narrow peak (almost a delta function)
+            (0, 1e5),  # Extremely broad distribution
+            (1e5, 1),  # Very large mean
+            (-1e5, 1),  # Very large negative mean
+            (1e5, 1e-5),  # Extreme mean and narrow std
+            (-1e5, 1e-5),  # Extreme negative mean and narrow std
+        ]
+
+        test_cases = edge_cases + extreme_cases
+
+        for mean_, std_ in test_cases:
+            dist_ = GaussianDistribution(amplitude=1.0, mean=mean_, standard_deviation=std_, normalize=True)
+
+            # SciPy Gaussian distribution for comparison
+            scipy_pdf = norm(loc=mean_, scale=std_).pdf(x_)
+            # scipy_logpdf = norm(loc=mean_, scale=std_).logpdf(x_)
+            scipy_cdf = norm(loc=mean_, scale=std_).cdf(x_)
+
+            # Assert PDF, logPDF, and CDF are close to SciPy's implementation
+            np.testing.assert_allclose(actual=dist_.pdf(x_), desired=scipy_pdf, rtol=1e-5, atol=1e-8)
+            # np.testing.assert_allclose(actual=dist_.logpdf(x_), desired=scipy_logpdf, rtol=1e-5, atol=1e-8)
+            np.testing.assert_allclose(actual=dist_.cdf(x_), desired=scipy_cdf, rtol=1e-5, atol=1e-8)
