+++ src/pymultifit/distributions/utilities.py
@@ -7,10 +7,10 @@ __all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'arc_sine_logpdf_',
            'folded_normal_pdf_',
            'gamma_sr_pdf_', 'gamma_sr_cdf_',
            'gamma_ss_',
-           'gaussian_pdf_',
+           'gaussian_pdf_', 'gaussian_cdf_',
            'half_normal_',
            'integral_check',
-           'laplace_',
+           'laplace_pdf_', 'laplace_cdf_',
            'log_normal_',
            'norris2005',
            'norris2011',
@@ -19,7 +19,7 @@ __all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'arc_sine_logpdf_',
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, gamma, gammainc, gammaln
+from scipy.special import betainc, erf, gamma, gammainc, gammaln
 
 doc_style = 'numpy_napoleon_with_merge'
 
@@ -169,13 +169,13 @@ def beta_pdf_(x: np.array,
     -----
     The Beta PDF is defined as:
 
-    .. math:: f(y; \\alpha, \\beta) = \\frac{y^{\\alpha - 1} (1 - y)^{\\beta - 1}}{B(\\alpha, \\beta)}
+    .. math:: f(y; \alpha, \beta) = \frac{y^{\alpha - 1} (1 - y)^{\beta - 1}}{B(\alpha, \beta)}
 
-    where :math:`B(\\alpha, \\beta)` is the Beta function, and :math:`y` is a transformed value of :math:`x` such that:
+    where :math:`B(\alpha, \beta)` is the Beta function (see, :obj:`scipy.special.beta`), and :math:`y` is a transformed value of :math:`x` such that:
 
-    .. math:: y = \\frac{x - \\text{loc}}{\\text{scale}}
+    .. math:: y = \frac{x - \text{loc}}{\text{scale}}
 
-    see, :obj:`scipy.special.beta`.
+    The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
     if x.size == 0:
         return np.array([])
@@ -405,20 +405,21 @@ def folded_normal_pdf_(x: np.array,
     variance : float, optional
         The variance parameter, :math:`\sigma^2`. Defaults to 1.0.
     normalize : bool, optional
-        If True, the distribution will be normalized so that the PDF is at most 1. Defaults to False.
+        If ``True``, the distribution will be normalized to integrate to 1.
+        Defaults to False.
 
     Returns
     -------
     np.array
-        The computed folded half-normal distribution values for the input `x`.
+        The computed folded normal distribution values for the input `x`.
 
     Notes
     -----
-    The folded half-normal distribution is defined as the sum of two Gaussian distributions:
+    The PDF of :class:`~pymultifit.distributions.foldedNormal_d.FoldedNormalDistribution` is given by:
 
-    .. math:: f(x; \mu, \sigma^2) = \phi(x; \mu, \sigma) + \phi(x; -\mu, \sigma)
+    .. math:: f(x; \mu, \sigma^2) = \phi(x; \mu, \sigma) + \phi(x; -\mu, \sigma),
 
-    where :math:`\phi` is the PDF for the :class:`~pymultifit.distributions.gaussian_d.GaussianDistribution`.
+    where :math:`\phi` is the PDF of :class:`~pymultifit.distributions.gaussian_d.GaussianDistribution`.
     """
     sigma = np.sqrt(variance)
     result = np.zeros_like(a=x, dtype=float)
@@ -532,44 +533,42 @@ def gamma_sr_cdf_(x: np.array,
 def gamma_ss_(x: np.array,
[…]
