### 2024-12-25 ec79c3e
[cdf_separation-major-2]
1. minor changes in `_distributions.rst`, `custom.css`
2. corrected class references in `index.rst`
3. correction in `laplace_d.rst`, `logNormal_d.rst` and `uniform_d.rst`
4. added docstrings to `logNormal_d.py`
5. added more functionalities to `utilities.py`
6. renamed `skewNorm_f.py` -> `skewNormal_f.py` and `logNorm_f.py` -> `logNormal_f.py` for API consistency

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
