+++ src/pymultifit/fitters/__init__.py
@@ -2,6 +2,6 @@
 
 from .gaussian_f import GaussianFitter
 from .laplace_f import LaplaceFitter
-from .logNorm_f import LogNormalFitter
+from .logNormal_f import LogNormalFitter
 from .mixed_f import MixedDataFitter
-from .skewNorm_f import SkewedNormalFitter
+from .skewNormal_f import SkewedNormalFitter
