+++ src/tests/test_distributions/test_logNormal_d.py
@@ -4,6 +4,7 @@ import numpy as np
 import pytest
 from scipy.stats import lognorm
 
+from ...pymultifit import EPSILON
 from ...pymultifit.distributions import LogNormalDistribution
 from ...pymultifit.distributions.backend import errorHandling as erH
 
@@ -12,10 +13,10 @@ class TestLogNormalDistribution:
 
     @staticmethod
     def test_initialization():
-        dist = LogNormalDistribution(amplitude=2.0, mean=1.0, standard_deviation=0.5, normalize=False)
+        dist = LogNormalDistribution(amplitude=2.0, mean=1.0, std=0.5, normalize=False)
         assert dist.amplitude == 2.0
         assert dist.mean == 1.0
-        assert dist.std_ == 0.5
+        assert dist.std == 0.5
         assert not dist.norm
 
         dist_normalized = LogNormalDistribution(amplitude=2.0, normalize=True)
@@ -31,7 +32,7 @@ class TestLogNormalDistribution:
         assert distribution.amplitude == 1.0
 
         with pytest.raises(erH.NegativeStandardDeviationError, match=f"Standard deviation {erH.neg_message}"):
-            LogNormalDistribution(standard_deviation=-3.0)
+            LogNormalDistribution(std=-3.0)
 
     @staticmethod
     def test_edge_case():
@@ -42,12 +43,12 @@ class TestLogNormalDistribution:
 
     @staticmethod
     def test_stats():
-        dist_ = LogNormalDistribution(amplitude=1.0, mean=2.0, standard_deviation=3.0)
+        dist_ = LogNormalDistribution(amplitude=1.0, mean=2.0, std=3.0)
         d_stats = dist_.stats()
-        assert d_stats["mean"] == np.exp(dist_.mean + (dist_.std_**2 / 2))
+        assert d_stats["mean"] == np.exp(dist_.mean + (dist_.std**2 / 2))
         assert d_stats["median"] == np.exp(dist_.mean)
-        assert d_stats["mode"] == np.exp(dist_.mean - dist_.std_**2)
-        assert d_stats["variance"] == (np.exp(dist_.std_**2) - 1) * np.exp(2 * dist_.mean + dist_.std_**2)
+        assert d_stats["mode"] == np.exp(dist_.mean - dist_.std**2)
+        assert d_stats["variance"] == (np.exp(dist_.std**2) - 1) * np.exp(2 * dist_.mean + dist_.std**2)
 
     @staticmethod
     def test_pdf_cdf_log_norm():
@@ -55,18 +56,19 @@ class TestLogNormalDistribution:
             """Evaluate the CDF or PDF for the custom LogNormalDistribution."""
             return dist_.cdf(x_) if what == 'cdf' else dist_.pdf(x_)
 
-        def _cdf_pdf_scipy(x_, s, scale, what='cdf'):
-            return lognorm.cdf(x_, s=s, scale=scale) if what == 'cdf' else lognorm.pdf(x_, s=s, scale=scale)
+        def _cdf_pdf_scipy(x_, s, loc, scale, what='cdf'):
+            return lognorm.cdf(x_, s=s, scale=scale, loc=loc) if what == 'cdf' else lognorm.pdf(x_, s=s, scale=scale, loc=loc)
 
         for i in ['cdf', 'pdf']:
             for _ in range(50):  # Run 50 random tests
-                shape_ = np.random.uniform(low=0.1, high=2.0)  # Shape parameter (sigma in log-normal)
-                scale_ = np.random.uniform(low=0.1, high=10.0)  # Scale parameter (exp(mean) in log-normal)
+                shape_ = np.random.uniform(low=0.1, high=2.0)
+                loc_ = np.random.uniform(low=0.1, high=2.0)
[…]
