+++ src/pymultifit/distributions/utilities.py
@@ -11,11 +11,11 @@ __all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'arc_sine_logpdf_',
            'half_normal_',
            'integral_check',
            'laplace_pdf_', 'laplace_cdf_',
-           'log_normal_',
+           'log_normal_pdf_', 'log_normal_cdf_',
            'norris2005',
            'norris2011',
            'power_law_',
-           'uniform_']
+           'uniform_pdf_', 'uniform_cdf_']
 
 import numpy as np
 from custom_inherit import doc_inherit
@@ -435,7 +435,7 @@ def gamma_sr_pdf_(x: np.array,
                   amplitude: float = 1.0, shape: float = 1.0, rate: float = 1.0,
                   loc: float = 0.0, normalize: bool = False) -> np.array:
     r"""
-    Compute the PDF for :class:`~pymultifit.distributions.gamma_d.GammaDistributionSR` with :math:`\alpha` (shape) and :math:`\lambda` (rate) parameters.
+    Compute PDF for :class:`~pymultifit.distributions.gamma_d.GammaDistributionSR` with :math:`\alpha` (shape) and :math:`\lambda` (rate) parameters.
 
     Parameters
     ----------
@@ -534,7 +534,7 @@ def gamma_ss_(x: np.array,
               amplitude: float = 1., shape: float = 1., scale: float = 1.,
               normalize: bool = False) -> np.array:
     r"""
-    Compute the PDF for :class:`~pymultifit.distributions.gamma_d.GammaDistributionSS` with :math:`\alpha` (shape) and :math:`\theta` (scale) parameters.
+    Compute PDF for :class:`~pymultifit.distributions.gamma_d.GammaDistributionSS` with :math:`\alpha` (shape) and :math:`\theta` (scale) parameters.
 
     Parameters
     ----------
@@ -588,7 +588,7 @@ def gaussian_pdf_(x: np.array,
     mu : float
         The mean parameter, :math:`\mu`. Defaults to 0.
     sigma : float
-        The standard deviation parameter, :math:`\sigma`. Defaults to 1.
+        The standard deviation parameter, :math:`\std_`. Defaults to 1.
     normalize : bool, optional
         If True, the distribution is normalized so that the total area under the PDF equals 1.
         Defaults to ``False``.
@@ -603,8 +603,8 @@ def gaussian_pdf_(x: np.array,
     The Gaussian PDF is defined as:
 
     .. math::
-        f(x; \mu, \sigma) = \phi\left(\dfrac{x-\mu}{\sigma}\right) =
-        \dfrac{1}{\sqrt{2\pi\sigma}}\exp\left[-\dfrac{1}{2}\left(\dfrac{x-\mu}{\sigma}\right)^2\right]
+        f(x; \mu, \std_) = \phi\left(\dfrac{x-\mu}{\std_}\right) =
+        \dfrac{1}{\sqrt{2\pi\std_}}\exp\left[-\dfrac{1}{2}\left(\dfrac{x-\mu}{\std_}\right)^2\right]
 
     The final PDF is expressed as :math:`f(x)`.
     """
@@ -666,7 +666,7 @@ def half_normal_(x: np.array,
     amplitude : float, optional
         The amplitude of the distribution. Defaults to 1. Ignored if `normalize` is set to True.
     scale : float, optional
-        The scale parameter (`sigma`) of the distribution, corresponding to the standard deviation of the original normal distribution.
+        The scale parameter (`std`) of the distribution, corresponding to the standard deviation of the original normal distribution.
         Defaults to 1.0.
     normalize : bool, optional
         If True, the distribution will be normalized so that the PDF is at most 1. Defaults to False.
[…]

[…]
