### 2024-12-25 f799790
[cdf_separation-minor-8]
1. minor changes
2. added more tests to `test_uniform_d.py`
3. updated `uniform_d.py` and `utilities.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
