+++ src/pymultifit/distributions/skewNormal_d.py
@@ -14,13 +14,16 @@ from .backend import BaseDistribution
 class SkewNormalDistribution(BaseDistribution):
     """Class for SkewNormal distribution."""
 
-    def __init__(self, shape: float = 1., location: float = 0., scale: float = 1.):
+    def __init__(self, amplitude: float = 1.0, shape: float = 1., location: float = 0., scale: float = 1., normalize: bool = False):
+        self.amplitude = 1 if normalize else amplitude
         self.shape = shape
         self.location = location
         self.scale = scale
 
+        self.norm = normalize
+
     def _pdf(self, x: np.ndarray) -> np.ndarray:
-        return skewnorm(self.shape, loc=self.location, scale=self.scale).pdf(x)
+        return skewnorm(self.shape, loc=self.location, scale=self.scale).pdf(x) * (self.amplitude * (2 / self.scale))
 
     def pdf(self, x: np.ndarray) -> np.ndarray:
         return self._pdf(x)
