+++ src/tests/test_distributions/test_uniform_d.py
@@ -2,9 +2,18 @@
 
 import numpy as np
 import pytest
+from scipy.stats import beta, uniform
 
+from ...pymultifit import EPSILON
 from ...pymultifit.distributions import UniformDistribution
 from ...pymultifit.distributions.backend import errorHandling as erH
+
+edge_cases = [
+    (0.0, 0.0),  # Degenerate
+    (0.0, EPSILON),  # Small range
+    (-1e6, 1e6),  # Large range
+    (3, 3)  # Another degenerate case
+]
 
 
 class TestUniformDistribution:
@@ -39,19 +48,33 @@ class TestUniformDistribution:
 
     @staticmethod
     def test_stats():
-        def check_stats(dist_):
-            """check stats for the Uniform distribution"""
+        def check_stats(dist_, scipy_dist):
+            """Check stats for the custom Uniform distribution against scipy."""
             stats = dist_.stats()
-            assert stats['mean'] == 0.5 * (dist_.low + dist_.high)
-            # no median value available so check for []
-            assert stats['median'] == 0.5 * (dist_.low + dist_.high)
-            assert stats['variance'] == (1 / 12.) * (dist_.high - dist_.low)**2
+            scipy_mean, scipy_var = scipy_dist.stats(moments='mv')
+            scipy_median = scipy_dist.median()
+
+            np.testing.assert_allclose(actual=stats['mean'], desired=scipy_mean)
+            np.testing.assert_allclose(actual=stats['median'], desired=scipy_median)
+            np.testing.assert_allclose(actual=stats['variance'], desired=scipy_var)
             assert stats.get('mode', []) == []
 
-        distribution1 = UniformDistribution(amplitude=1.0, normalize=True)
-        check_stats(distribution1)
-        distribution2 = UniformDistribution(amplitude=1.0, low=-3.0, high=3.0, normalize=True)
-        check_stats(distribution2)
+        def _tester(low_, high_):
+            distribution = UniformDistribution(amplitude=1.0, low=low_, high=high_, normalize=True)
+            scipy_uniform = uniform(loc=distribution.low, scale=distribution.high)
+            scipy_beta = beta(1, 1, loc=distribution.low, scale=distribution.high)
+            check_stats(dist_=distribution, scipy_dist=scipy_uniform)
+            check_stats(dist_=distribution, scipy_dist=scipy_beta)
+
+        for _ in range(50):
+            low_val = np.random.uniform(low=0.0, high=10.0)
+            high_val = np.random.uniform(low=0.0, high=10.0)
+
+            _tester(low_=low_val, high_=high_val)
+
+        for low_val, high_val in edge_cases:
+            print(low_val, high_val)
+            _tester(low_=low_val, high_=high_val)
 
     @staticmethod
     def test_pdf_cdf_uniform():
@@ -59,21 +82,29 @@ class TestUniformDistribution:
             """Evaluate the CDF or PDF for the custom UniformDistribution."""
             return dist_.cdf(x_) if what == 'cdf' else dist_.pdf(x_)
 
-        def _cdf_pdf_scipy(x_, loc, scale, what='cdf'):
-            """Evaluate the CDF or PDF for SciPy's uniform distribution."""
-            from scipy.stats import uniform
+        def _cdf_pdf_scipy_u(x_, loc, scale, what='cdf'):
+            """Evaluate the CDF or PDF for SciPy uniform distribution."""
             return uniform.cdf(x_, loc=loc, scale=scale) if what == 'cdf' else uniform.pdf(x_, loc=loc, scale=scale)
+
+        def _cdf_pdf_scipy_b(x_, loc, scale, what='cdf'):
[…]
