+++ examples/skewed_normal.py
@@ -3,8 +3,8 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from src.pymultifit.fitters import SkewedNormalFitter
-from src.pymultifit.generators import multi_skewed_normal
+from pymultifit.fitters import SkewNormalFitter
+from pymultifit.generators import multi_skewed_normal
 
 params = [(12, 2, -7, 1), (3, -4, 2, 2), (4, 2, 7, 1.5)]
 
@@ -13,7 +13,7 @@ x = np.linspace(-15, 15, 10_000)
 noise_level = 1e-2
 y = multi_skewed_normal(x, params=params, noise_level=noise_level)
 
-fitter = SkewedNormalFitter(n_fits=3, x_values=x, y_values=y)
+fitter = SkewNormalFitter(n_fits=3, x_values=x, y_values=y)
 
 guess = [(4, 6, -6, 0.5), (3, -4, 2, 2), (4, 2, 7, 1.5)]
 
