+++ src/pymultifit/fitters/__init__.py
@@ -4,4 +4,4 @@ from .gaussian_f import GaussianFitter
 from .laplace_f import LaplaceFitter
 from .logNormal_f import LogNormalFitter
 from .mixed_f import MixedDataFitter
-from .skewNormal_f import SkewedNormalFitter
+from .skewNormal_f import SkewNormalFitter

[…]
