### 2024-12-25 aa81a01
[cdf_separation-major-3]
1. minor changes in `_distribution_api_template.py`, `_distributions.rst`, and `_fitters.rst`
2. cosmetic changes in distributions rst files
3. added PDf and CDF functions in `gamma_d.py`
4. parameter name changes in `gaussian_d.py`
5. minor changes in fitters and `generators.py`
6. added `skew_normal_pdf_` and `skew_normal_cdf_` to `utilities.py`
7. adapted `utilities.py` for `skewNormal_d.py` and `skewNormal_f.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
