+++ src/pymultifit/distributions/utilities.py
@@ -1,6 +1,6 @@
 """Created on Aug 03 17:13:21 2024"""
 
-__all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'arc_sine_logpdf_',
+__all__ = ['arc_sine_pdf_',
            'beta_pdf_', 'beta_cdf_', 'beta_logpdf_',
            'chi_square_pdf_',
            'exponential_pdf_',
@@ -15,11 +15,12 @@ __all__ = ['arc_sine_pdf_', 'arc_sine_cdf_', 'arc_sine_logpdf_',
            'norris2005',
            'norris2011',
            'power_law_',
+           'skew_normal_pdf_', 'skew_normal_cdf_',
            'uniform_pdf_', 'uniform_cdf_']
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gamma, gammainc, gammaln
+from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t
 
 doc_style = 'numpy_napoleon_with_merge'
 
@@ -46,15 +47,15 @@ def integral_check(pdf_function, x_range: tuple) -> float:
 
 
 def arc_sine_pdf_(x: np.array,
-                  amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0,
-                  normalize: bool = False) -> np.array:
+                  amplitude: float = 1.0,
+                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
     r"""
-    Compute PDF of the ArcSine distribution.
+    Compute PDF of :class:`pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
     Parameters
     ----------
     x : np.array
-        Input array of values where the PDF is evaluated.
+        Input array of values.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
         Ignored if **normalize** is ``True``.
@@ -88,77 +89,34 @@ def arc_sine_pdf_(x: np.array,
     return beta_pdf_(x=x, amplitude=amplitude, alpha=0.5, beta=0.5, loc=loc, scale=scale, normalize=normalize)
 
 
-@doc_inherit(parent=arc_sine_pdf_, style=doc_style)
-def arc_sine_cdf_(x: np.array,
-                  amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0,
-                  normalize: bool = False) -> np.array:
-    r"""
-    Compute CDF of the ArcSine distribution.
-
-    Parameters
-    ----------
-    amplitude: float, optional
-        For API consistency only.
-    normalize: bool, optional
-        For API consistency only.
-
-    Returns
-    -------
-    np.array
-        Array of the same size as :math:`x` containing the evaluated CDF values.
-
-    Notes
-    -----
-    The ArcSine CDF is defined as:
-
-    .. math:: F(y) = \frac{2}{\pi}\arcsin\left[\sqrt{y}\right]
-
-    where, :math:`y` is a transformed value of :math:`x`, defined as:
-
-    .. math:: y = \frac{x - \text{loc}}{\text{scale}}
-    """
-    return beta_cdf_(x=x, amplitude=amplitude, alpha=0.5, beta=0.5, loc=loc, scale=scale, normalize=normalize)
-
-
-@doc_inherit(parent=arc_sine_pdf_, style=doc_style)
-def arc_sine_logpdf_(x: np.array,
-                     amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0,
-                     normalize: bool = False) -> np.array:
-    r"""
-    Compute logPDF of the ArcSine Distribution.
-
-    Returns
-    -------
-    np.array
-        Array of the same shape as `x`, containing the evaluated log-PDF values.
-    """
[…]
