+++ src/pymultifit/distributions/gaussian_d.py
@@ -11,22 +11,22 @@ from .utilities import gaussian_cdf_, gaussian_pdf_
 class GaussianDistribution(BaseDistribution):
     """Class for Gaussian distribution."""
 
-    def __init__(self, amplitude: float = 1.0, mean: float = 0., standard_deviation: float = 1., normalize: bool = False):
+    def __init__(self, amplitude: float = 1.0, mean: float = 0., std: float = 1., normalize: bool = False):
         if not normalize and amplitude <= 0:
             raise erH.NegativeAmplitudeError()
-        elif standard_deviation <= 0:
+        elif std <= 0:
             raise erH.NegativeStandardDeviationError()
 
         self.amplitude = 1. if normalize else amplitude
         self.mean = mean
-        self.std_ = standard_deviation
+        self.std_ = std
         self.norm = normalize
 
     def _pdf(self, x: np.ndarray) -> np.ndarray:
-        return gaussian_pdf_(x, amplitude=self.amplitude, mu=self.mean, sigma=self.std_, normalize=self.norm)
+        return gaussian_pdf_(x, amplitude=self.amplitude, mean=self.mean, std=self.std_, normalize=self.norm)
 
     def _cdf(self, x: np.array) -> np.array:
-        return gaussian_cdf_(x, amplitude=self.amplitude, mu=self.mean, sigma=self.std_, normalize=self.norm)
+        return gaussian_cdf_(x, amplitude=self.amplitude, mean=self.mean, std=self.std_, normalize=self.norm)
 
     def stats(self) -> Dict[str, float]:
         mean_, std_ = self.mean, self.std_
