+++ src/pymultifit/distributions/gamma_d.py
@@ -29,12 +29,10 @@ class GammaDistributionSR(BaseDistribution):
     def _pdf(self, x: np.array) -> np.array:
         if np.any(x < 0):
             raise erH.XOutOfRange()
-        return gamma_sr_pdf_(x,
-                             amplitude=self.amplitude, shape=self.shape, rate=self.rate, loc=self.loc, normalize=self.norm)
+        return gamma_sr_pdf_(x, amplitude=self.amplitude, alpha=self.shape, lambda_=self.rate, loc=self.loc, normalize=self.norm)
 
     def _cdf(self, x: np.array) -> np.array:
-        return gamma_sr_cdf_(x,
-                             amplitude=self.amplitude, shape=self.shape, rate=self.rate, normalize=self.norm)
+        return gamma_sr_cdf_(x, amplitude=self.amplitude, alpha=self.shape, lambda_=self.rate, normalize=self.norm)
 
     def stats(self) -> Dict[str, float]:
         a, b = self.shape, self.rate
