+++ examples/skewed_normal.py
@@ -2,27 +2,21 @@
 
 import matplotlib.pyplot as plt
 import numpy as np
-from scipy.stats import skewnorm
 
 from pymultifit.fitters import SkewNormalFitter
 from pymultifit.generators import multi_skewed_normal
 
-params = [(2, 1, -7, 1), (3, 3, -2, 1), (4, 2, 7, 1)]
+params = [(3, 4, -7, 3), (3, -5, -2, 8), (4, 2, 7, 2)]
 
-x = np.linspace(-15, 15, 10_000)
+x = np.linspace(-30, 15, 10_000)
 
-y = multi_skewed_normal(x, params=params)
-
-y2 = np.zeros_like(x, dtype=float)
-for par in params:
-    y2 += par[0] * skewnorm.pdf(x, a=par[1], loc=par[2], scale=par[3]) * (np.sqrt(2 * np.pi * par[3]**2) / (2 * par[3]))
+noise_level = 0.1
+y = multi_skewed_normal(x, params=params, noise_level=noise_level)
 
 fitter = SkewNormalFitter(n_fits=3, x_values=x, y_values=y)
 
-guess = [(2, 1, -6, 0.5), (3, 1, 2, 2), (4, 1, 7, 1.5)]
+guess = [(2, 1, -6, 0.5), (3, -2, -2, 5), (4, 1, 7, 1.5)]
 
 fitter.fit(guess)
-f, ax = plt.subplots(1, 1, figsize=(8, 4))
-plotter = fitter.plot_fit(show_individual=True, x_label='X_data', y_label='Y_data', title='XY_plot', data_label='XY_data', axis=ax)
-ax.plot(x, y2, 'k--')
+plotter = fitter.plot_fit(show_individual=True, x_label='X_data', y_label='Y_data', title='XY_plot', data_label='XY_data')
 plt.show()
