+++ src/pymultifit/distributions/utilities.py
@@ -1,6 +1,7 @@
 """Created on Aug 03 17:13:21 2024"""
 
-__all__ = ['arc_sine_pdf_',
+__all__ = ['_sanity_check',
+           'arc_sine_pdf_',
            'beta_pdf_', 'beta_cdf_', 'beta_logpdf_',
            'chi_square_pdf_',
            'exponential_pdf_',
@@ -139,20 +140,13 @@ def beta_pdf_(x: np.array,
         return np.array([])
 
     y = (x - loc) / scale
-    pdf_ = np.zeros_like(y, dtype=float)
+    pdf_ = np.zeros_like(a=y, dtype=float)
 
     numerator = y**(alpha - 1) * (1 - y)**(beta - 1)
+    normalization_factor = gamma(alpha) * gamma(beta) / gamma(alpha + beta)
 
-    if normalize:
-        normalization_factor = gamma(alpha) * gamma(beta)
-        normalization_factor /= gamma(alpha + beta)
-        amplitude = 1.0
-    else:
-        normalization_factor = 1.0
-
-    mask_ = _beta_masking(alpha=alpha, beta=beta, y=y)
-
-    pdf_[~mask_] = amplitude * (numerator[~mask_] / normalization_factor)
+    mask_ = _beta_masking(y=y, alpha=alpha, beta=beta)
+    pdf_[~mask_] = numerator[~mask_] / normalization_factor
 
     if alpha <= 1:
         pdf_[y == 0] = np.inf
@@ -161,7 +155,16 @@ def beta_pdf_(x: np.array,
 
     # handle the cases where nans can occur with nan_to_num
     # np.inf and -np.inf to not affect the infinite values
-    return np.nan_to_num(x=pdf_ / scale, copy=False, nan=0, posinf=np.inf, neginf=-np.inf)
+    pdf_ = _remove_nans(pdf_ / scale)
+
+    if not normalize:
+        pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
+
+    return pdf_
+
+
+def _remove_nans(x: np.array) -> np.array:
+    return np.nan_to_num(x=x, copy=False, nan=0, posinf=np.inf, neginf=-np.inf)
 
 
 @doc_inherit(parent=beta_pdf_, style=doc_style)
@@ -186,10 +189,9 @@ def beta_logpdf_(x: np.array,
         return np.array([])
 
     y = (x - loc) / scale
-    logpdf_ = np.full_like(y, -np.inf, dtype=float)  # Default to -inf for invalid regions
+    mask_ = _beta_masking(y=y, alpha=alpha, beta=beta)
 
-    mask_ = _beta_masking(alpha=alpha, beta=beta, y=y)
-
+    logpdf_ = np.full_like(a=y, fill_value=-np.inf, dtype=float)
     log_numerator = (alpha - 1) * np.log(y) + (beta - 1) * np.log(1 - y)
 
     if normalize:
@@ -208,9 +210,9 @@ def beta_logpdf_(x: np.array,
     return logpdf_ - np.log(scale)
 
 
-def _beta_masking(alpha, beta, y):
+def _beta_masking(y, alpha, beta):
     out_of_range_mask = np.logical_or(y < 0, y > 1)
-    undefined_mask = np.zeros_like(y, dtype=bool)
+    undefined_mask = np.zeros_like(a=y, dtype=bool)
     if alpha <= 1:
         undefined_mask = np.logical_or(undefined_mask, y == 0)
     if beta <= 1:
@@ -249,7 +251,7 @@ def beta_cdf_(x: np.array,
     where :math:`I_x(\alpha, \beta)` is the regularized incomplete Beta function, see :obj:`~scipy.special.betainc`.
     """
[…]
