### 2024-12-26 28a81ab
[cdf_separation-minor-9]
1. docstring updated in `generators.py`
2. minor changes in `skew_normal.py` example
