+++ src/pymultifit/generators/generators.py
@@ -13,38 +13,29 @@ listOfTuplesOrArray = Union[listOfTuples, np.ndarray]
 doc_style = 'numpy_napoleon_with_merge'
 
 
-def multi_chi_squared(x: np.ndarray, params: listOfTuplesOrArray,
-                      noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
-    """Generate multi-Chi-Squared data with optional noise.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        X values.
-    params : listOfTuplesOrArray
-        List of tuples containing the parameters for each Chi-Squared distribution.
-    noise_level : float, optional
-        Standard deviation of the noise to be added to the data, by default 0.0.
-    normalize : bool, optional
-        If True, the function produces normalized data (Integration[PDF] < 1). Defaults to False.
-
-    Returns
-    -------
-    np.ndarray
-        Y values of the multi-Chi-Squared data with added noise.
-    """
-    return multi_base(x=x, distribution_func=dist.ChiSquareDistribution, params=params, noise_level=noise_level, normalize=normalize)
-
-
-@doc_inherit(parent=multi_chi_squared, style=doc_style)
 def multi_base(x: np.ndarray, distribution_func: Callable, params: listOfTuplesOrArray,
                noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
     """Generate data based on a combination of distributions with optional noise.
 
+    Parameters
+    ----------
+    x : np.array
+        Input array of values where PDF is evaluated.
+    distribution_func: Callable
+        The distribution function to be used to generate data.
+    params : listOfTuplesOrArray
+        List of tuples containing the parameters for the required distribution.
+    noise_level : float, optional
+        Standard deviation of the noise to be added to the data.
+        Defaults to 0.0.
+    normalize : bool, optional
+        If ``True``, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
+
     Returns
     -------
-    np.ndarray
-        Y values of the generated data with added noise.
+    np.array
+        Array of the same shape as :math:`x`, containing the evaluated values.
     """
     y = np.zeros_like(x, dtype=float)
 
@@ -57,139 +48,79 @@ def multi_base(x: np.ndarray, distribution_func: Callable, params: listOfTuplesO
     return y
 
 
-@doc_inherit(parent=multi_chi_squared, style=doc_style)
-def multi_exponential(x: np.ndarray, params: listOfTuplesOrArray,
+def multi_chi_squared(x: np.ndarray, params: listOfTuplesOrArray,
                       noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
-    """Generate multi-Exponential data with optional noise.
+    r"""
+    Generate multi-:class:`~pymultifit.distributions.chiSquare_d.ChiSquareDistribution` data with optional noise.
+
+    Parameters
+    ----------
+    x : np.array
+        Input array of values where PDF is evaluated.
+    params : listOfTuplesOrArray
+        List of tuples containing the parameters for the required distribution.
+    noise_level : float, optional
+        Standard deviation of the noise to be added to the data.
+        Defaults to 0.0.
+    normalize : bool, optional
+        If ``True``, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
 
     Returns
     -------
-    np.ndarray
-        Y values of the multi-Exponential data with added noise.
+    np.array
[…]
