### 2024-12-26 3fc3340
[cdf_separation-minor-9-1]
1. moved documents to docs/ instead of sphinx_docs/
