+++ docs/source/py_shenanigans/_distribution_api_runner.py
@@ -1,6 +1,6 @@
 """Created on Dec 22 18:04:13 2024"""
 
-from sphinx_docs.source.py_shenanigans._distribution_api_template import generate_distribution_rst_simple
+from docs.source.py_shenanigans._distribution_api_template import generate_distribution_rst_simple
 
 dist_names = ['ArcSine', 'Beta', 'ChiSquare', 'Exponential', 'FoldedNormal', 'Gaussian', 'HalfNormal', 'Laplace',
               'LogNormal', 'SkewNormal', 'Uniform']
