+++ .readthedocs.yml
@@ -10,10 +10,10 @@ build:
     tools:
         python: "3.9"
 
-# Build documentation in the "sphinx_docs" directory with Sphinx
+# Build documentation in the "docs" directory with Sphinx
 sphinx:
     builder: html
-    configuration: ./sphinx_docs/source/conf.py
+    configuration: ./docs/source/conf.py
 
 # recursive addition of submodules
 submodules:
