### 2024-12-26 e7af5d9
[cdf_separation-minor-10]
1. added non-working fitter rst files in fitters
2. corrected fitter functions for n_fits
