+++ src/pymultifit/fitters/foldedNormal_f.py
@@ -5,12 +5,12 @@ from .utilities import sanity_check
 from ..distributions.utilities import folded_normal_pdf_
 
 
-class FoldedHalfNormalFitter(BaseFitter):
+class FoldedNormalFitter(BaseFitter):
     """A class for fitting multiple Folded half-normal functions to the given data."""
 
-    def __init__(self, n_fits: int, x_values, y_values, max_iterations: int = 1000):
+    def __init__(self, x_values, y_values, max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
-        super().__init__(n_fits=n_fits, x_values=x_values, y_values=y_values, max_iterations=max_iterations)
+        super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 3
 
     @staticmethod
