+++ examples/log_normal.py
@@ -14,7 +14,7 @@ x = np.linspace(EPSILON, 100, 2000)
 noise_level = 0.2
 y = multi_log_normal(x, params=params, noise_level=noise_level)
 
-fitter = LogNormalFitter(n_fits=3, x_values=x, y_values=y)
+fitter = LogNormalFitter(x_values=x, y_values=y)
 
 guess = [(10, 1, 1), (3, 2, 0.2), (10, 4, 0.1)]
 
