### 2024-12-26 3c13ffa
[doc-minor-11]
1. updated `_fitters.rst` according to present fitters
2. updated docstring of `chiSquare_f.py`
3. updated `chiSquare_f.rst`
4. minor updates
