+++ docs/source/fitters/uniform_f.rst
@@ -1,24 +1,0 @@
-Uniform Distribution
-====================
-
-.. autoclass:: pymultifit.distributions.uniform_d.UniformDistribution
-   :no-members:
-
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities` module:
-
-* :mod:`~pymultifit.distributions.utilities.uniform_pdf_`
-* :mod:`~pymultifit.distributions.utilities.uniform_cdf_`
-
-Recommended Import
-------------------
-
-.. code-block:: python
-
-   from pymultifit.distributions import UniformDistribution
-
-Full Import
------------
-
-.. code-block:: python
-
-   from pymultifit.distributions.uniform_d import UniformDistribution
