### 2024-12-26 2d77b14
[signatures-minor-1]
1. added extra parameter logic in `baseFitter.py`
2. updated `chiSquare_f.py`
3. introduced new parameter setup in `laplace_f.py` and `logNormal_f.py`
4. added `DistributionBounds` in `utilities_f.py`, centralizing the boundaries of parameters
5. updated version in `version.py`
