+++ src/pymultifit/fitters/chiSquare_f.py
@@ -11,8 +11,8 @@ class ChiSquareFitter(BaseFitter):
     def __init__(self, x_values, y_values, max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
-        self.n_par = 2
+        self.n_par = 3
 
     @staticmethod
     def _fitter(x, params):
-        return chi_square_pdf_(x, normalize=False)
+        return chi_square_pdf_(x, *params, normalize=False)
