+++ src/pymultifit/fitters/logNormal_f.py
@@ -10,7 +10,7 @@ from ..distributions.utilities_d import log_normal_pdf_
 
 
 class LogNormalFitter(BaseFitter):
-    """A class for fitting multiple Log Normal functions to the given data."""
+    """Class for multi-LogNormal fitting."""
 
     def __init__(self, x_values, y_values, max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
@@ -18,7 +18,10 @@ class LogNormalFitter(BaseFitter):
             raise ValueError("The LogNormal distribution must have x > 0. "
                              "Use `EPSILON` from the package to get as close to zero as possible.")
         super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
-        self.n_par = 3
+
+        self.n_par = 4
+        self.pn_par = 3
+        self.sn_par = {'loc': 0}
 
     @staticmethod
     def _fitter(x, params):
