+++ src/pymultifit/distributions/utilities_d.py
@@ -5,11 +5,11 @@ __all__ = ['_sanity_check',
            'beta_pdf_', 'beta_cdf_', 'beta_logpdf_',
            'chi_square_pdf_',
            'exponential_pdf_',
-           'folded_normal_pdf_',
+           'folded_normal_pdf_', 'folded_normal_cdf_',
            'gamma_sr_pdf_', 'gamma_sr_cdf_',
            'gamma_ss_pdf_',
            'gaussian_pdf_', 'gaussian_cdf_',
-           'half_normal_',
+           'half_normal_pdf_', 'half_normal_cdf_',
            'integral_check',
            'laplace_pdf_', 'laplace_cdf_',
            'log_normal_pdf_', 'log_normal_cdf_',
@@ -306,7 +306,7 @@ def exponential_pdf_(x: np.array,
                      amplitude: float = 1., scale: float = 1.,
                      normalize: bool = False) -> np.array:
     r"""
-    Compute the PDF of :mod:`~pymultifit.distributions.exponential_d.ExponentialDistribution`.
+    Compute PDF of :mod:`~pymultifit.distributions.exponential_d.ExponentialDistribution`.
 
     Parameters
     ----------
@@ -361,7 +361,7 @@ def folded_normal_pdf_(x: np.array,
     Parameters
     ----------
     x : np.array
-        Input array of values.
+        Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
         Ignored if **normalize** is ``True``.
@@ -370,13 +370,13 @@ def folded_normal_pdf_(x: np.array,
     variance : float, optional
         The variance parameter, :math:`\sigma^2`. Defaults to 1.0.
     normalize : bool, optional
-        If ``True``, the distribution will be normalized to integrate to 1.
-        Defaults to False.
+        If ``True``, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
 
     Returns
     -------
     np.array
-        The computed folded normal distribution values for the input `x`.
+        Array of the same shape as :math:`x`, containing the evaluated values.
 
     Notes
     -----
@@ -400,9 +400,20 @@ def folded_normal_pdf_(x: np.array,
     return pdf_
 
 
+@doc_inherit(parent=folded_normal_pdf_, style=doc_style)
 def folded_normal_cdf_(x: np.array,
                        amplitude: float = 1., mu: float = 0.0, variance: float = 1.0,
                        normalize: bool = False) -> np.array:
+    r"""
+    Compute CDF for :class:`~pymultifit.distributions.foldedNormal_d.FoldedNormalDistribution`.
+
+    Parameters
+    ----------
+    amplitude: float, optional
+        For API consistency only.
+    normalize: bool, optional
+        For API consistency only.
+    """
     y = np.zeros_like(a=x, dtype=float)
     mask = x >= 0
     frac1, frac2 = (x[mask] - mu) / np.sqrt(2 * variance), (x[mask] + mu) / np.sqrt(2 * variance)
@@ -631,8 +642,8 @@ def gaussian_cdf_(x: np.array,
     return 0.5 * (1 + erf(num_ / den_))
 
 
-def half_normal_(x: np.array,
-                 amplitude: float = 1.0, scale: float = 1.0, normalize: bool = False) -> np.array:
+def half_normal_pdf_(x: np.array,
+                     amplitude: float = 1.0, scale: float = 1.0, normalize: bool = False) -> np.array:
     """
     Compute the half-normal distribution.
 
@@ -669,6 +680,11 @@ def half_normal_(x: np.array,
[…]
