+++ docs/source/distributions/arcSine_d.rst
@@ -10,13 +10,13 @@ ArcSine Distribution
     * :math:`\alpha_\text{beta} = 0.5`,
     * :math:`\lambda_\text{beta} = 0.5`.
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities` module:
+This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities_d` module:
 
-* :func:`~pymultifit.distributions.utilities.beta_pdf_`
-* :func:`~pymultifit.distributions.utilities.beta_cdf_`
+* :func:`~pymultifit.distributions.utilities_d.beta_pdf_`
+* :func:`~pymultifit.distributions.utilities_d.beta_cdf_`
 
 ..
-    * :func:`~pymultifit.distributions.utilities.beta_logpdf_`
+    * :func:`~pymultifit.distributions.utilities_d.beta_logpdf_`
 
 Recommended Import
 ^^^^^^^^^^^^^^^^^^
