+++ docs/source/distributions/gaussian_d.rst
@@ -4,10 +4,10 @@ Gaussian Distribution
 .. autoclass:: pymultifit.distributions.gaussian_d.GaussianDistribution
    :no-members:
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities` module:
+This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities_d` module:
 
-* :func:`~pymultifit.distributions.utilities.gaussian_pdf_`
-* :func:`~pymultifit.distributions.utilities.gaussian_cdf_`
+* :func:`~pymultifit.distributions.utilities_d.gaussian_pdf_`
+* :func:`~pymultifit.distributions.utilities_d.gaussian_cdf_`
 
 Recommended Import
 ------------------
