### 2024-12-27 06caaf5
[signatures-major-1]
1. major overhaul to `baseFitter.py`
2. baseFitter related changes to all fitter classes,
3. added `gamma_f.py`
4. updated examples
5. added `gamma_sr.py` example
