### 2024-12-27 2ee898a
[signatures-minor-3]
1. made `exponential_d.py` independent of `gamma_d.py
2. minor changes in `exponential_f.py`, `gamma_f.py`, `gaussian_f.py`, `laplace_f.py`, and `skewNormal_f.py`
3. added comprehensive `test_exponential_d.py`
4. added exponential functions in `utilities_d.py`
