+++ src/pymultifit/fitters/gaussian_f.py
@@ -5,7 +5,6 @@ from scipy.optimize import Bounds
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
-from .. import EPSILON
 from ..distributions.utilities_d import gaussian_pdf_
 
 
@@ -19,7 +18,7 @@ class GaussianFitter(BaseFitter):
 
     @staticmethod
     def fit_boundaries():
-        lb = (EPSILON, -np.inf, EPSILON)
+        lb = (0, -np.inf, 0)
         ub = (np.inf, np.inf, np.inf)
         return Bounds(lb=lb, ub=ub)
 
