+++ docs/source/distributions/beta_d.rst
@@ -14,7 +14,7 @@ Beta Distribution
    - :class:`~pymultifit.distributions.uniform_d.UniformDistribution`:
         - :math:`\alpha = \beta = 1`
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities_d` module:
+This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 
 * :func:`~pymultifit.distributions.utilities_d.beta_pdf_`
 * :func:`~pymultifit.distributions.utilities_d.beta_cdf_`
