+++ docs/source/fitters/laplace_f.rst
@@ -4,7 +4,7 @@ Laplace Distribution
 .. autoclass:: pymultifit.distributions.laplace_d.LaplaceDistribution
    :no-members:
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities` module:
+This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities` module:
 
 * :func:`~pymultifit.distributions.utilities.laplace_pdf_`
 * :func:`~pymultifit.distributions.utilities.laplace_cdf_`
