+++ docs/source/distributions/skewNormal_d.rst
@@ -4,17 +4,10 @@ SkewNormal Distribution
 .. autoclass:: pymultifit.distributions.skewNormal_d.SkewNormalDistribution
    :no-members:
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities_d` module:
+This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 
 * :mod:`~pymultifit.distributions.utilities_d.skew_normal_pdf_`
 * :mod:`~pymultifit.distributions.utilities_d.skew_normal_cdf_`
-
-Full Import
------------
-
-.. code-block:: python
-
-   from pymultifit.distributions.skewNormal_d import SkewNormalDistribution
 
 Recommended Import
 ------------------
@@ -22,3 +15,10 @@ Recommended Import
 .. code-block:: python
 
    from pymultifit.distributions import SkewNormalDistribution
+
+Full Import
+-----------
+
+.. code-block:: python
+
+   from pymultifit.distributions.skewNormal_d import SkewNormalDistribution
