+++ docs/source/fitters/gaussian_f.rst
@@ -4,7 +4,7 @@ Gaussian Distribution
 .. autoclass:: pymultifit.distributions.gaussian_d.GaussianDistribution
    :no-members:
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities` module:
+This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities` module:
 
 * :func:`~pymultifit.distributions.utilities.gaussian_pdf_`
 * :func:`~pymultifit.distributions.utilities.gaussian_cdf_`
