+++ docs/source/distributions/exponential_d.rst
@@ -11,10 +11,10 @@ Exponential Distribution
     * :math:`\alpha_\text{gammaSR} = 1`,
     * :math:`\lambda_\text{gammaSR} = \lambda_\text{expon}`.
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities_d` module:
+This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 
-* :func:`~pymultifit.distributions.utilities_d.gamma_sr_pdf_`
-* :func:`~pymultifit.distributions.utilities_d.gamma_sr_cdf_`
+* :func:`~pymultifit.distributions.utilities_d.exponential_pdf_`
+* :func:`~pymultifit.distributions.utilities_d.exponential_cdf_`
 
 Recommended Import
 ------------------
