+++ docs/source/distributions/foldedNormal_d.rst
@@ -4,17 +4,10 @@ FoldedNormal Distribution
 .. autoclass:: pymultifit.distributions.foldedNormal_d.FoldedNormalDistribution
    :no-members:
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities_d` module:
+This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 
 * :mod:`~pymultifit.distributions.utilities_d.folded_normal_pdf_`
 * :mod:`~pymultifit.distributions.utilities_d.folded_normal_cdf_`
-
-Full Import
------------
-
-.. code-block:: python
-
-   from pymultifit.distributions.foldedNormal_d import FoldedNormalDistribution
 
 Recommended Import
 ------------------
@@ -22,3 +15,10 @@ Recommended Import
 .. code-block:: python
 
    from pymultifit.distributions import FoldedNormalDistribution
+
+Full Import
+-----------
+
+.. code-block:: python
+
+   from pymultifit.distributions.foldedNormal_d import FoldedNormalDistribution
