+++ docs/source/fitters/chiSquare_f.rst
@@ -4,7 +4,7 @@ ChiSquare Fitter
 .. autoclass:: pymultifit.fitters.chiSquare_f.ChiSquareFitter
    :show-inheritance:
 
-This class internally utilizes the following functions from the :mod:`~pymultifit.distributions.utilities_f` module:
+This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_f` module:
 
 * :func:`~pymultifit.distributions.utilities_d.chi_square_pdf_`
 
