### 2024-12-27 08dcfa9
[signatures-minor-4]
1. added `001__distribution_class.ipynb`
2. renamed previous tutorials
3. minor updates in all distributions
