+++ examples/gamma_sr.py
@@ -7,15 +7,15 @@ from src.pymultifit import EPSILON
 from src.pymultifit.fitters import GammaSRFitter
 from src.pymultifit.generators import multi_gamma_sr
 
-params = [(2, 2, 1, 5), (4, 6, 2), (1, 3, 2, 9)]
+params = [(2, 2, 1, 5), (4, 6, 2, 1), (1, 3, 2, 9)]
 x = np.linspace(EPSILON, 15, 1000)
 
-noise_level = 0.1
+noise_level = 0.05
 y = multi_gamma_sr(x, params=params, noise_level=noise_level)
 
 fitter = GammaSRFitter(x, y)
 
-guess = [(1, 1, 1, 1), (3, 4, 1), (1, 2, 1, 6)]
+guess = [(2, 2, 1, 3), (3, 2, 1, 1), (1, 2, 1, 6)]
 
 fitter.fit(p0=guess)
 
