### 2024-12-28 6a487d1
[signatures-minor-5]
1. added `002__generator_class.ipynb` tutorial
2. function signature update in `generators.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
