+++ docs/source/tutorials.rst
@@ -5,6 +5,7 @@ Tutorials
    :maxdepth: 1
 
    basics1
+   basics2
    generating_data
    generating_n_model_data
    generating_mixture_model_data
