### 2024-12-28 4ed1da6
[signatures-minor-6]
1. added `003__fitter_module.ipynb`
2. updated `001__distribution_class.ipynb` and `002__generator_class.ipynb`
3. added `basics.rst` and `basics3.nblink`
4. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
