+++ docs/source/basics.rst
@@ -1,0 +1,19 @@
+Basics of ``pyMultiFit``
+========================
+
+This section introduces the foundational concepts and modules of ``pyMultiFit`` library.
+It introduces the how-to about using the template classes and methods provided by the library.
+After going through these tutorials, you will learn how to,
+
+#. Construct a custom distribution class compatible with ``pyMultiFit`` library,
+#. Use a template function to make an n-model data generator for your custom distribution,
+#. Construct a custom fitting class to fit n-model data, and
+#. Going deeper into the fitting class and obtaining fit values.
+
+.. toctree::
+   :hidden:
+   :maxdepth: 1
+
+   basics1
+   basics2
+   basics3
