### 2024-12-29 835fc6e
[signatures-minor-7]
1. updated `tutorials.rst`, `001__distribution_class.ipynb`, `002__generator_class.ipynb` and `003__fitter_module.ipynb`
2. added ArcSine to `__init__.py`
3. added an example `arcSine.py`
4. added `beta1.py` and `beta2.py`
5. added images

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
