+++ src/pymultifit/distributions/beta_d.py
@@ -10,7 +10,78 @@ from .utilities_d import beta_cdf_, beta_pdf_
 
 
 class BetaDistribution(BaseDistribution):
-    """Class for Beta distribution."""
+    r"""
+    Class for Beta distribution.
+
+    Parameters
+    ----------
+    amplitude : float, optional
+        The amplitude of the PDF. Defaults to 1.0.
+        Ignored if **normalize** is ``True``.
+    alpha : float, optional
+        The :math:`\alpha` parameter.
+        Default is 1.0.
+    beta : float, optional
+        The :math:`\beta` parameter.
+        Default is 1.0.
+    loc : float, optional
+        The location parameter, :math:`-` shifting.
+        Defaults to 0.0.
+    scale: float, optional
+        The scale parameter, :math:`-` scaling.
+        Defaults to 1.0,
+    normalize : bool, optional
+        If ``True``, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
+
+    Examples
+    --------
+    Importing libraries
+
+    .. literalinclude:: ../../../examples/basic/beta1.py
+       :language: python
+       :linenos:
+       :lineno-start: 3
+       :lines: 3-7
+
+    Generating a standard :math:`\text{Beta}(2, 30)` distribution with ``pyMultiFit`` and ``scipy``.
+
+    .. literalinclude:: ../../../examples/basic/beta1.py
+       :language: python
+       :linenos:
+       :lineno-start: 9
+       :lines: 9-12
+
+    Plotting **PDF** and **CDF**
+
+    .. literalinclude:: ../../../examples/basic/beta1.py
+       :language: python
+       :linenos:
+       :lineno-start: 14
+       :lines: 14-29
+
+    .. image:: ../../../images/beta_example1.png
+       :alt: Beta distribution (5, 30)
+       :align: center
+
+    Generating a shifted and translated :math:`\text{Beta}(2, 30)` distribution.
+
+    .. literalinclude:: ../../../examples/basic/beta2.py
+       :language: python
+       :lineno-start: 9
+       :lines: 9-12
+
+    Plotting **PDF** and **CDF**
+
+    .. literalinclude:: ../../../examples/basic/beta2.py
+       :language: python
+       :lineno-start: 14
+       :lines: 14-29
+
+    .. image:: ../../../images/beta_example2.png
+       :alt: Beta distribution (shifted and translated)
+       :align: center
+    """
 
     def __init__(self,
                  amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
