+++ docs/source/tutorials.rst
@@ -4,7 +4,6 @@ Tutorials
 .. toctree::
    :maxdepth: 1
 
-   tutorials/GALLERY_HEADER
    basics
    generating_data
    generating_n_model_data
