+++ docs/source/basics.rst
@@ -8,7 +8,6 @@ After going through these tutorials, you will learn how to,
 #. Construct a custom distribution class compatible with ``pyMultiFit`` library,
 #. Use a template function to make an n-model data generator for your custom distribution,
 #. Construct a custom fitting class to fit n-model data, and
-#. Going deeper into the fitting class and obtaining fit values.
 
 .. toctree::
    :hidden:
