### 2024-12-29 a7369ec
[signatures-minor-8]
1. updated the fitter rst files
2. minor changes to `exponential_f.py`
3. added `mixed_f.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
