+++ src/pymultifit/fitters/exponential_f.py
@@ -1,5 +1,7 @@
 """Created on Nov 30 11:30:45 2024"""
 
+import numpy as np
+
 from .backend import BaseFitter
 from .utilities_f import sanity_check
 from ..distributions.utilities_d import exponential_pdf_
@@ -11,7 +13,15 @@ class ExponentialFitter(BaseFitter):
     def __init__(self, x_values, y_values, max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
-        self.n_par = 2
+        self.n_par = 3
+        self.pn_par = 2
+        self.sn_par = {'loc': 0}
+
+    @staticmethod
+    def fit_boundaries():
+        lb = (0, 0, -np.inf)
+        ub = (np.inf, np.inf, np.inf)
+        return lb, ub
 
     @staticmethod
     def fitter(x, params):
