+++ examples/gamma_sr.py
@@ -4,7 +4,7 @@ import numpy as np
 from matplotlib import pyplot as plt
 
 from src.pymultifit import EPSILON
-from src.pymultifit.fitters import GammaSRFitter
+from src.pymultifit.fitters import GammaFitterSR
 from src.pymultifit.generators import multi_gamma_sr
 
 params = [(2, 2, 1, 5), (4, 6, 2, 1), (1, 3, 2, 9)]
@@ -13,7 +13,7 @@ x = np.linspace(EPSILON, 15, 1000)
 noise_level = 0.05
 y = multi_gamma_sr(x, params=params, noise_level=noise_level)
 
-fitter = GammaSRFitter(x, y)
+fitter = GammaFitterSR(x, y)
 
 guess = [(2, 2, 1, 3), (3, 2, 1, 1), (1, 2, 1, 6)]
 
