+++ src/pymultifit/fitters/backend/baseFitter.py
@@ -18,8 +18,8 @@ _gaussian = 'GaussianFitter'
 _lNormal = 'LogNormalFitter'
 _skNormal = 'SkewNormalFitter'
 _laplace = 'LaplaceFitter'
-_gammaSR = 'GammaSRFitter'
-_gammaSS = 'GammaSSFitter'
+_gammaSR = 'GammaFitterSR'
+_gammaSS = 'GammaFitterSS'
 
 
 class BaseFitter:
@@ -75,28 +75,101 @@ class BaseFitter:
         return adjusted_p0
 
     def _covariance(self):
+        """
+        Store the covariance matrix of the fitted model.
+
+        Returns
+        -------
+        np.array
+            An array containing the covariance matrix of the fitted model.
+        """
         if self.covariance is None:
             raise RuntimeError("Fit not performed yet. Call fit() first.")
         return self.covariance
 
     @staticmethod
-    def _format_param(value, t_low=0.001, t_high=10_000):
-        """Formats the parameter value based on its magnitude."""
+    def _format_param(value, t_low=0.001, t_high=10_000) -> str:
+        r"""
+        Formats the parameter value to scientific notation based on its magnitude.
+
+        Parameters
+        ----------
+        value: float
+            The value of the parameter to be formatted.
+        t_low: float, optional
+            The lower bound below which the formatting should be applied to the value.
+            Defaults to 0.001.
+        t_high: float, optional
+            The upper bound above which the formatting should be applied to the value.
+            Defaults to 10,000.
+
+        Returns
+        -------
+        str:
+            A formatted string of the parameter value.
+        """
         return f'{value:.3E}' if t_high < abs(value) or abs(value) < t_low else f'{value:.3f}'
 
-    def _n_fitter(self, x, *params):
+    def _n_fitter(self, x: np.array, *params) -> np.array:
+        r"""
+        Perform N-fitting by summing over multiple parameter sets.
+
+        Parameters
+        ----------
+        x : np.array
+            Input array of values for which the composite function is evaluated.
+        params : tuple
+            A tuple containing all parameters to be fitted. This is reshaped into  an array of size (``self.n_fits``, ``self.n_par``) where:
+            - ``self.n_fits`` is the number of individual fits.
+            - ``self.n_par`` is the number of parameters per fit.
+
+        Returns
+        -------
+        np.array
+            An array containing the composite fitted values for the input ``x``.
+        """
         y = np.zeros_like(a=x, dtype=float)
         params = np.reshape(a=params, newshape=(self.n_fits, self.n_par))
         for par in params:
             y += self.fitter(x=x, params=par)
         return y
 
-    def _params(self):
+    def _params(self) -> np.array:
+        r"""
+        Store the fitted parameters of the fitted model.
+
+        Returns
+        -------
+        np.array
+            The parameters obtained after performing the fit.
+
+        Raises
+        ------
+        RuntimeError
+            If the fit has not been performed yet (i.e., ``self.params`` is ``None``).
+
+        Notes
+        -----
+        This method assumes that the fitting process assigns values to ``self.params``.
+        """
         if self.params is None:
             raise RuntimeError("Fit not performed yet. Call fit() first.")
         return self.params
 
     def _plot_individual_fitter(self, plotter):
+        r"""
+        Plot individual fits from the composite fitter.
+
+        Parameters
+        ----------
+        plotter : matplotlib.axes.Axes
+            The axis object where the plots will be rendered.
+
+        Notes
+        -----
+        - `self.params` must contain the fitted parameters reshaped into (``self.n_fits``, ``self.n_par``) before calling this method.
+        - Each plot will be labeled with the class name and the index of the fit, along with the formatted parameters.
+        """
         x = self.x_values
         params = np.reshape(a=self.params, newshape=(self.n_fits, self.n_par))
[…]
