+++ docs/source/pymultifit.fitters.rst
@@ -12,14 +12,6 @@ Subpackages
 Submodules
 ----------
 
-pymultifit.fitters.mixed\_f module
-----------------------------------
-
-.. automodule:: pymultifit.fitters.mixed_f
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 pymultifit.fitters.norris\_f module
 -----------------------------------
 
@@ -32,14 +24,6 @@ pymultifit.fitters.powerLaw\_f module
 -------------------------------------
 
 .. automodule:: pymultifit.fitters.powerLaw_f
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pymultifit.fitters.skewNorm\_f module
--------------------------------------
-
-.. automodule:: pymultifit.fitters.skewNormal_f
    :members:
    :undoc-members:
    :show-inheritance:

[…]
