### 2024-12-29 b6bbb03
[signatures-minor-9]
1. updated `test_gamma_sr_d.py` for loc as well
2. minor changes in `exponential_d.py`, `foldedNormal_d.py`
3. added boundaries in `foldedNormal_f.py`
4. updated `gamma_d.py`
5. major docstring changes in `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
