+++ src/pymultifit/distributions/utilities_d.py
@@ -50,12 +50,12 @@ def arc_sine_pdf_(x: np.array,
                   amplitude: float = 1.0,
                   loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
     r"""
-    Compute PDF of :class:`pymultifit.distributions.arcSine_d.ArcSineDistribution`.
+    Compute PDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
     Parameters
     ----------
     x : np.array
-        Input array of values.
+        Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
         Ignored if **normalize** is ``True``.
@@ -84,7 +84,7 @@ def arc_sine_pdf_(x: np.array,
 
     .. math:: y = \frac{x - \text{loc}}{\text{scale}}
 
-    The final PDF is expressed as :math:`f(y)/\text{scale}`
+    The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
     return beta_pdf_(x=x, amplitude=amplitude, alpha=0.5, beta=0.5, loc=loc, scale=scale, normalize=normalize)
 
@@ -98,7 +98,7 @@ def beta_pdf_(x: np.array,
     Parameters
     ----------
     x : np.array
-        Input array of values.
+        Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
         Ignored if **normalize** is ``True``.
@@ -109,10 +109,10 @@ def beta_pdf_(x: np.array,
         The :math:`\beta` parameter.
         Default is 1.0.
     loc : float, optional
-        The location parameter :math:`-` for shifting.
+        The location parameter, for shifting.
         Default is 0.0.
     scale : float, optional
-        The scale parameter - scaling.
+        The scale parameter, for scaling.
         Default is 1.0.
     normalize : bool, optional
         If ``True``, the distribution is normalized so that the total area under the PDF equals 1.
@@ -141,7 +141,7 @@ def beta_pdf_(x: np.array,
     y = (x - loc) / scale
     pdf_ = np.zeros_like(a=y, dtype=float)
 
-    numerator = y**(alpha - 1) * (1 - y)**(beta - 1)
+    numerator = y ** (alpha - 1) * (1 - y) ** (beta - 1)
     normalization_factor = gamma(alpha) * gamma(beta) / gamma(alpha + beta)
 
     mask_ = _beta_masking(y=y, alpha=alpha, beta=beta)
@@ -171,18 +171,12 @@ def beta_logpdf_(x: np.array,
                  amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
     r"""
-    Compute logPDF of the Beta distribution.
+    Compute logPDF for :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
 
     Parameters
     ----------
     x : np.array
-        Input array of values where the logPDF is evaluated.
-
-
-    Returns
-    -------
-    np.array
-        Array of the same shape as :math:`x`, containing the evaluated log-PDF values.
+        Input array of values where logPDF is evaluated.
     """
     if x.size == 0:
         return np.array([])
@@ -230,16 +224,11 @@ def beta_cdf_(x: np.array,
     Parameters
     ----------
     x : np.array
-        The input array for which to compute the Beta PDF.
+        Input array of values.
     amplitude : float, optional
[…]
