+++ src/tests/test_distributions/test_gamma_sr_d.py
@@ -4,6 +4,7 @@ import numpy as np
 import pytest
 from scipy.stats import gamma
 
+from pymultifit.distributions import GammaDistributionSS
 from ...pymultifit import EPSILON
 from ...pymultifit.distributions import GammaDistributionSR
 from ...pymultifit.distributions.backend import errorHandling as erH
@@ -46,22 +47,35 @@ class TestGammaDistributionSR:
 
     @staticmethod
     def test_stats():
-        def check_stats_sr(dist_):
-            """check stats for the gammaSS distribution"""
-            stats = dist_.stats()
-            assert stats['mean'] == dist_.shape / dist_.rate
-            # no median value available so check for []
-            assert stats.get('median', []) == []
-            if dist_.shape >= 1:
-                assert stats['mode'] == (dist_.shape - 1) / dist_.rate
-            else:
-                assert stats['mode'] == 0
-            assert stats['variance'] == dist_.shape / dist_.rate
+        def check_stats_sr(custom_, scipy_):
+            stats = custom_.stats()
 
-        dist1 = GammaDistributionSR(amplitude=1.0, shape=1.0, rate=1.0, normalize=True)
-        check_stats_sr(dist1)
-        dist2 = GammaDistributionSR(amplitude=1.0, shape=0.5, rate=1.0, normalize=True)
-        check_stats_sr(dist2)
+            # Check mean
+            scipy_mean = scipy_.mean()
+            assert np.isclose(stats['mean'], scipy_mean, rtol=1e-6), f"Mean mismatch: {stats['mean']} vs {scipy_mean}"
+
+            # Check variance
+            scipy_variance = scipy_.var()
+            assert np.isclose(stats['variance'], scipy_variance,
+                              rtol=1e-6), f"Variance mismatch: {stats['variance']} vs {scipy_variance}"
+
+            # scipy doesn't provide mode for the distribution so I'm not doing that test
+
+            # Wikipedia doesn't have any analytical form for the median, so i'm skipping this test
+            assert stats.get('median', []) == [], "Median check failed (should be empty)."
+
+        for _ in range(50):
+            shape_ = np.random.uniform(low=0.1, high=5.0)
+            scale_ = np.random.uniform(low=0.1, high=2.0)
+            loc_ = np.random.uniform(low=-5.0, high=5.0)
+
+            dist1 = GammaDistributionSR(shape=shape_, rate=scale_, loc=loc_, normalize=True)
+            dist2 = GammaDistributionSS(shape=shape_, scale=1 / scale_, loc=loc_, normalize=True)
+
+            scipy_dist = gamma(a=shape_, scale=1 / scale_, loc=loc_)
+
+            check_stats_sr(custom_=dist1, scipy_=scipy_dist)
+            check_stats_sr(custom_=dist2, scipy_=scipy_dist)
 
     @staticmethod
     def test_pdf_cdf():
@@ -69,19 +83,23 @@ class TestGammaDistributionSR:
             return dist_.cdf(x_) if what == 'cdf' else dist_.pdf(x_)
 
         def _cdf_pdf_scipy(x_, shape, rate, loc, what='cdf'):
-            return [gamma.cdf(x_, a=shape, scale=1 / rate) if what == 'cdf' else
-                    gamma.pdf(x_, a=shape, scale=1 / rate, loc=loc)][0]
+            return [gamma.cdf(x_, a=shape, scale=rate, loc=loc) if what == 'cdf' else
+                    gamma.pdf(x_, a=shape, scale=rate, loc=loc)][0]
 
-        for i in ['cdf', 'pdf']:
+        for i in ['pdf', 'cdf']:
             for _ in range(50):  # Run 50 random tests
[…]

[…]
