### 2024-12-29 9a96e3b
[signatures-minor-10]
1. updated fitters with boundaries and other necessary parameters

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
