+++ src/pymultifit/fitters/norris_f.py
@@ -1,31 +1,0 @@
-"""Created on Nov 30 05:33:38 2024"""
-
-from .backend import BaseFitter
-from .utilities_f import sanity_check
-from ..distributions.utilities_d import norris2005, norris2011
-
-
-class Norris2005Fitter(BaseFitter):
-    """A class for fitting multiple Norris 2005 models to the given data."""
-
-    def __init__(self, x_values, y_values, max_iterations: int = 1000):
-        x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
-        super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
-        self.n_par = 3
-
-    @staticmethod
-    def fitter(x, params):
-        return norris2005(x, *params, normalize=False)
-
-
-class Norris2011Fitter(BaseFitter):
-    """A class for fitting multiple Norris 2011 models to the given data."""
-
-    def __init__(self, x_values, y_values, max_iterations: int = 1000):
-        x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
-        super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
-        self.n_par = 3
-
-    @staticmethod
-    def fitter(x, params):
-        return norris2011(x, *params, normalize=False)
