### 2024-12-29 995fd56
[signatures-minor-11]
1. added `_error_handling.rst`
2. updated `api_index.rst`, `conf.py`
3. added exceptions raising to `arcSine_d.py` and `beta_d.py`
4. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
