+++ src/pymultifit/generators/generators.py
@@ -5,7 +5,7 @@ from typing import List, Tuple, Type, Union
 import numpy as np
 from custom_inherit import doc_inherit
 
-from .. import distributions as dist, GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, POWERLAW, SKEW_NORMAL
+from .. import distributions as dist, GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL
 from ..distributions.backend import BaseDistribution
 
 listOfTuples = List[Tuple[float, ...]]
@@ -21,7 +21,7 @@ def multi_base(x: np.array, distribution_func: Type[BaseDistribution], params: l
     Parameters
     ----------
     x : np.array
-        Input array of values where PDF is evaluated.
+        Input array of values.
     distribution_func: Type[BaseDistribution]
         The distribution function to be used to generate data.
     params : listOfTuplesOrArray
@@ -59,7 +59,7 @@ def multi_chi_squared(x: np.array, params: listOfTuplesOrArray,
     Parameters
     ----------
     x : np.array
-        Input array of values where PDF is evaluated.
+        Input array of values.
     params : listOfTuplesOrArray
         List of tuples containing the parameters for the required distribution.
     noise_level : float, optional
@@ -74,80 +74,101 @@ def multi_chi_squared(x: np.array, params: listOfTuplesOrArray,
     np.array
         Array of the same shape as :math:`x`, containing the evaluated values.
     """
-    return multi_base(x=x, distribution_func=dist.ChiSquareDistribution, params=params, noise_level=noise_level, normalize=normalize)
+    return multi_base(x=x,
+                      distribution_func=dist.ChiSquareDistribution, params=params, noise_level=noise_level,
+                      normalize=normalize)
 
 
 @doc_inherit(parent=multi_chi_squared, style=doc_style)
 def multi_gamma_sr(x: np.array, params: listOfTuplesOrArray,
                    noise_level: float = 0.0, normalize: bool = False) -> np.array:
     r"""Generate multi-:class:`~pymultifit.distributions.gamma_d.GammaDistributionSR` data with optional noise."""
-    return multi_base(x=x, distribution_func=dist.GammaDistributionSR, params=params, noise_level=noise_level, normalize=normalize)
+    return multi_base(x=x,
+                      distribution_func=dist.GammaDistributionSR, params=params, noise_level=noise_level,
+                      normalize=normalize)
 
 
 @doc_inherit(parent=multi_chi_squared, style=doc_style)
 def multi_gamma_ss(x: np.array, params: listOfTuplesOrArray,
                    noise_level: float = 0.0, normalize: bool = False) -> np.array:
     r"""Generate multi-:class:`~pymultifit.distributions.gamma_d.GammaDistributionSS` data with optional noise."""
-    return multi_base(x=x, distribution_func=dist.GammaDistributionSS, params=params, noise_level=noise_level, normalize=normalize)
+    return multi_base(x=x,
+                      distribution_func=dist.GammaDistributionSS, params=params, noise_level=noise_level,
+                      normalize=normalize)
 
 
 @doc_inherit(parent=multi_chi_squared, style=doc_style)
 def multi_exponential(x: np.array, params: listOfTuplesOrArray,
                       noise_level: float = 0.0, normalize: bool = False) -> np.array:
     r"""Generate multi-:class:`~pymultifit.distributions.exponential_d.ExponentialDistribution` data with optional noise."""
-    return multi_base(x=x, distribution_func=dist.ExponentialDistribution, params=params, noise_level=noise_level, normalize=normalize)
+    return multi_base(x=x,
+                      distribution_func=dist.ExponentialDistribution, params=params, noise_level=noise_level,
+                      normalize=normalize)
 
 
[…]
