+++ src/pymultifit/fitters/__init__.py
@@ -1,7 +1,11 @@
 """Created on Aug 03 20:34:39 2024"""
 
+from .chiSquare_f import ChiSquareFitter
+from .exponential_f import ExponentialFitter
+from .foldedNormal_f import FoldedNormalFitter
 from .gamma_f import GammaFitterSR, GammaFitterSS
 from .gaussian_f import GaussianFitter
+from .halfNormal_f import HalfNormalFitter
 from .laplace_f import LaplaceFitter
 from .logNormal_f import LogNormalFitter
 from .mixed_f import MixedDataFitter
