+++ src/pymultifit/distributions/__init__.py
@@ -12,9 +12,7 @@ from .gaussian_d import GaussianDistribution
 from .halfNormal_d import HalfNormalDistribution
 from .laplace_d import LaplaceDistribution
 from .logNormal_d import LogNormalDistribution
-from .norris_d import Norris2005Distribution, Norris2011Distribution
 from .others import cubic, line, linear, nth_polynomial, quadratic
-from .powerLaw_d import PowerLawDistribution
 from .skewNormal_d import SkewNormalDistribution
 from .uniform_d import UniformDistribution
 
