+++ src/pymultifit/distributions/arcSine_d.py
@@ -1,5 +1,6 @@
 """Created on Aug 14 02:02:42 2024"""
 
+from .backend import errorHandling as erH
 from .beta_d import BetaDistribution
 
 
@@ -51,12 +52,25 @@ class ArcSineDistribution(BetaDistribution):
     .. image:: ../../../images/arcsine_example.png
        :alt: ArcSine distribution
        :align: center
+
+    Raises
+    ------
+    :class:`~pymultifit.distributions.backend.errorHandling.NegativeAmplitudeError`
+        If the provided value of amplitude is negative.
+    :class:`~pymultifit.distributions.backend.errorHandling.NegativeScaleError`
+        If the provided value of scale is negative.
     """
 
     def __init__(self, amplitude: float = 1., loc: float = 0.0, scale: float = 1.0, normalize: bool = False):
+        if not normalize and amplitude <= 0:
+            raise erH.NegativeAmplitudeError()
+        if scale < 0:
+            raise erH.NegativeScaleError()
+
         self.amplitude = 1 if normalize else amplitude
         self.loc = loc
         self.scale = scale
 
         self.norm = normalize
-        super().__init__(amplitude=self.amplitude, alpha=0.5, beta=0.5, loc=self.loc, scale=self.scale, normalize=self.norm)
+        super().__init__(amplitude=self.amplitude, alpha=0.5, beta=0.5, loc=self.loc, scale=self.scale,
+                         normalize=self.norm)
