+++ docs/source/pymultifit.distributions.rst
@@ -1,14 +1,6 @@
 pymultifit.distributions package
 ================================
 
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   pymultifit.distributions.backend
-
 Submodules
 ----------
 
