+++ docs/source/pymultifit.distributions.backend.rst
@@ -1,21 +1,0 @@
-pymultifit.distributions.backend package
-========================================
-
-Submodules
-----------
-
-pymultifit.distributions.backend.errorHandling module
------------------------------------------------------
-
-.. automodule:: pymultifit.distributions.backend.errorHandling
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pymultifit.distributions.backend
-   :members:
-   :undoc-members:
-   :show-inheritance:
