### 2024-12-30 e873ab2
[signatures-minor-12]
updated documentation for `chiSquare_d.py` and `exponential_d.py`
