+++ src/pymultifit/distributions/exponential_d.py
@@ -9,7 +9,79 @@ from .utilities_d import exponential_cdf_, exponential_pdf_
 
 
 class ExponentialDistribution(BaseDistribution):
-    """Class for Exponential distribution."""
+    r"""
+    Class for Exponential distribution.
+
+    Parameters
+    ----------
+    amplitude : float, optional
+        The amplitude of the PDF. Defaults to 1.0.
+        Ignored if **normalize** is ``True``.
+    scale: float, optional
+        The scale parameter, :math:`\lambda`.
+        Defaults to 1.0.
+    loc : float, optional
+        The location parameter, for shifting.
+        Defaults to 0.0.
+    normalize : bool, optional
+        If ``True``, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
+
+    Examples
+    --------
+    Importing libraries
+
+    .. literalinclude:: ../../../examples/basic/expon.py
+       :language: python
+       :linenos:
+       :lineno-start: 3
+       :lines: 3-7
+
+    Generating a standard Exponential(:math:`\lambda =1.5`) distribution with ``pyMultiFit`` and ``scipy``.
+
+    .. literalinclude:: ../../../examples/basic/expon.py
+       :language: python
+       :linenos:
+       :lineno-start: 9
+       :lines: 9-12
+
+    Plotting **PDF** and **CDF**
+
+    .. literalinclude:: ../../../examples/basic/expon.py
+       :language: python
+       :linenos:
+       :lineno-start: 14
+       :lines: 14-29
+
+    .. image:: ../../../images/expon_example1.png
+       :alt: Expon(1.5)
+       :align: center
+
+    Generating a translated Exponential(:math:`\lambda=1.5`) distribution with :math:`\text{loc} = 3`.
+
+    .. literalinclude:: ../../../examples/basic/expon.py
+       :language: python
+       :lineno-start: 32
+       :lines: 32
+
+    Plotting **PDF** and **CDF**
+
+    .. literalinclude:: ../../../examples/basic/expon.py
+       :language: python
+       :lineno-start: 34
+       :lines: 34-49
+
+    .. image:: ../../../images/expon_example2.png
+       :alt: Expon(1.5, 3)
+       :align: center
+
+    Raises
+    ------
+    :class:`~pymultifit.distributions.backend.errorHandling.NegativeAmplitudeError`
+        If the provided value of amplitude is negative.
+    :class:`~pymultifit.distributions.backend.errorHandling.NegativeScaleError`
+        If the provided value of scale is negative.
+    """
 
     def __init__(self, amplitude: float = 1.0, scale: float = 1.0, loc: float = 0.0, normalize: bool = False):
         if not normalize and amplitude <= 0:
