+++ src/pymultifit/distributions/utilities_d.py
@@ -1,7 +1,8 @@
 """Created on Aug 03 17:13:21 2024"""
 
-__all__ = ['arc_sine_pdf_',
-           'beta_pdf_', 'beta_cdf_', 'beta_logpdf_',
+__all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans',
+           'arc_sine_pdf_',
+           'beta_pdf_', 'beta_cdf_',
            'chi_square_pdf_',
            'exponential_pdf_', 'exponential_cdf_',
            'folded_normal_pdf_', 'folded_normal_cdf_',
@@ -9,7 +10,6 @@ __all__ = ['arc_sine_pdf_',
            'gamma_ss_pdf_',
            'gaussian_pdf_', 'gaussian_cdf_',
            'half_normal_pdf_', 'half_normal_cdf_',
-           'integral_check',
            'laplace_pdf_', 'laplace_cdf_',
            'log_normal_pdf_', 'log_normal_cdf_',
            'skew_normal_pdf_', 'skew_normal_cdf_',
@@ -17,30 +17,9 @@ __all__ = ['arc_sine_pdf_',
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t
+from scipy.special import betainc, erf, gammainc, gammaln, owens_t
 
 doc_style = 'numpy_napoleon_with_merge'
-
-
-def integral_check(pdf_function, x_range: tuple) -> float:
-    """
-    Compute the integral of a given PDF function over a specified range.
-
-    Parameters
-    ----------
-    pdf_function : function
-        The PDF function to integrate.
-    x_range : tuple
-        The range (a, b) over which to integrate the PDF.
-
-    Returns
-    -------
-    float
-        The integral result of the PDF function over the specified range.
-    """
-    from scipy.integrate import quad
-    integral = quad(lambda x: pdf_function(x), x_range[0], x_range[1])[0]
-    return integral
 
 
 def arc_sine_pdf_(x: np.array,
@@ -138,7 +117,7 @@ def beta_pdf_(x: np.array,
     y = (x - loc) / scale
     pdf_ = np.zeros_like(a=y, dtype=float)
 
-    numerator = y ** (alpha - 1) * (1 - y) ** (beta - 1)
+    numerator = y**(alpha - 1) * (1 - y)**(beta - 1)
     normalization_factor = gamma(alpha) * gamma(beta) / gamma(alpha + beta)
 
     mask_ = _beta_masking(y=y, alpha=alpha, beta=beta)
@@ -159,22 +138,11 @@ def beta_pdf_(x: np.array,
     return pdf_
 
 
-def _remove_nans(x: np.array) -> np.array:
-    return np.nan_to_num(x=x, copy=False, nan=0, posinf=np.inf, neginf=-np.inf)
-
-
 @doc_inherit(parent=beta_pdf_, style=doc_style)
 def beta_logpdf_(x: np.array,
                  amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
-    r"""
-    Compute logPDF for :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
-
-    Parameters
-    ----------
-    x : np.array
-        Input array of values where logPDF is evaluated.
-    """
+    r"""Compute logPDF for :class:`~pymultifit.distributions.beta_d.BetaDistribution`."""
     if x.size == 0:
         return np.array([])
 
@@ -198,17 +166,6 @@ def beta_logpdf_(x: np.array,
[…]
