+++ docs/source/distributions/arcSine_d.rst
@@ -3,12 +3,7 @@ ArcSine Distribution
 
 .. autoclass:: pymultifit.distributions.arcSine_d.ArcSineDistribution
    :no-members:
-
-.. note::
-    The :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution` is a special case of the :class:`~pymultifit.distributions.beta_d.BetaDistribution`,
-
-    * :math:`\alpha_\text{beta} = 0.5`,
-    * :math:`\lambda_\text{beta} = 0.5`.
+   :show-inheritance:
 
 This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 

[…]
