### 2024-12-30 4e6ddbb
[doc-minor-12]
1. updated `003__fitter_module.ipynb`
2. added `004__fitter_module2.ipynb`
3. fixed the API documentation, quite a lot
4. table fix changes to `custom.css`
5. added `basics4.nblink`, `generators.rst` and `utilities_f.rst`
6. added error retrieving capability in `mixed_f.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
