+++ docs/source/basics.rst
@@ -16,3 +16,4 @@ After going through these tutorials, you will learn how to,
    basics1
    basics2
    basics3
+   basics4
