+++ docs/source/distributions/utilities_d.rst
@@ -3,11 +3,13 @@ Distribution utilities
 
 .. py:currentmodule:: pymultifit.distributions.utilities_d
 
-The :mod:`utilities_d` module serves as a cornerstone of the ``pyMultiFit`` library, containing majority of the core utility functions utilized across various classes and components.
+.. automodule:: pymultifit.distributions.utilities_d
+
+The :mod:`~pymultifit.distributions.utilities_d` module serves as a cornerstone of the ``pyMultiFit`` library, containing majority of the core utility functions utilized across various classes and components.
 It provides a comprehensive suite of mathematical and statistical tools, including probability distribution functions (PDFs), cumulative distribution functions (CDFs), some internal scaling and masking utilities, and data preprocessing methods.
 
-Available for use
------------------
+Available for user
+------------------
 
 .. autofunction:: arc_sine_pdf_
 .. autofunction:: beta_pdf_
