+++ docs/source/generators/generators.rst
@@ -1,0 +1,7 @@
+Generators
+==========
+
+.. automodule:: pymultifit.generators.generators
+   :members:
+   :undoc-members:
+   :show-inheritance:
