+++ docs/source/pymultifit.fitters.rst
@@ -1,29 +1,0 @@
-pymultifit.fitters package
-==========================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   pymultifit.fitters.backend
-
-Submodules
-----------
-
-pymultifit.fitters.utilities module
------------------------------------
-
-.. automodule:: pymultifit.fitters.utilities_f
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pymultifit.fitters
-   :members:
-   :undoc-members:
-   :show-inheritance:
