+++ docs/source/pymultifit.rst
@@ -1,30 +1,0 @@
-pyMultiFit
-==================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 2
-
-   pymultifit.fitters
-   pymultifit.generators
-
-Submodules
-----------
-
-pymultifit.version module
--------------------------
-
-.. automodule:: pymultifit.version
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pymultifit
-   :members:
-   :undoc-members:
-   :show-inheritance:
