### 2024-12-31 49bf64b
[doc-minor-13]
1. updated `tutorials.rst`, removing some tutorials
2. updated `index.rst`
3. added `benchmarks.rst`, and `_constants.rst`
4. docstring changes in `mixed_f.py`
5. minor changes in `mixed_f.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
