+++ src/pymultifit/fitters/mixed_f.py
@@ -1,7 +1,7 @@
 """Created on Aug 10 23:08:38 2024"""
 
 import itertools
-from typing import Optional
+from typing import Optional, Tuple, Union, List, Callable
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -52,40 +52,17 @@ model_dict = {LINE: [_Line, 2],
 
 
 class MixedDataFitter:
-    """
-    A class to fit a mixture of different models to data.
+    r"""
+    Class to fit a mixture of different models to data.
 
-    Attributes
-    ----------
-    x_values : array_like
-        The x-values for the data.
-    y_values : array_like
-        The y-values for the data.
-    model_list : list of str
-        List of models to fit (e.g., ['gaussian', 'gaussian', 'line']).
-    params : array_like, optional
-        The fitted parameters of the model after fitting.
-    covariance : array_like, optional
-        The covariance of the fitted parameters.
-    model_function : callable
-        The composite model function used for fitting.
+    :param x_values: The x-values for the data.
+    :param y_values: The y-values for the data.
+    :param model_list: List of models to fit (e.g., `LINE`, `GAUSSIAN`, `LOG_NORMAL`)
+    :param max_iterations: The maximum number of iterations for fitting procedure.
     """
 
-    def __init__(self, x_values, y_values, model_list, max_iterations: int = 1000):
-        """
-        Initializes the MixedDataFitter with data and a list of models.
-
-        Parameters
-        ----------
-        x_values : array_like
-            The x-values for the data.
-        y_values : array_like
-            The y-values for the data.
-        model_list : list of str
-            List of models to fit (e.g., ['gaussian', 'gaussian', 'line']).
-        max_iterations: int, optional
-            The max number of iterations for fitting procedure.
-        """
+    def __init__(self, x_values: Union[List, np.ndarray], y_values: Union[List, np.ndarray],
+                 model_list: List[str], max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
 
         self.x_values = x_values
@@ -103,30 +80,27 @@ class MixedDataFitter:
         return (f"{self.__class__.__name__}(x_values={self.x_values}, y_values={self.y_values}, "
                 f"model_list={self.model_list}, max_iterations={self.max_iterations})")
 
-    def _create_model_function(self):
+    def _create_model_function(self) -> Callable:
         """
         Creates a composite model function based on the specified models.
 
-        Returns
-        -------
-        callable
-            A function that can be used for fitting.
+        :return: A composite model for fitting.
         """
 
-        def _composite_model(x, *params):
+        def _composite_model(x: np.ndarray, *params) -> np.ndarray:
             """
             Compute the composite model.
 
             Parameters
             ----------
-            x : array_like
+            x : np.ndarray
                 The x-values where the model is evaluated.
             params : tuple
                 Parameters for the model components.
 
             Returns
             -------
-            y : array_like
+            y : np.ndarray
                 The computed y-values from the composite model.
             """
             y = np.zeros_like(x, dtype=float)
@@ -141,14 +115,11 @@ class MixedDataFitter:
 
         return _composite_model
 
-    def _expected_param_count(self):
+    def _expected_param_count(self) -> int:
         """
         Calculates the expected number of parameters based on the model list.
 
-        Returns
-        -------
-        int
-            The expected number of parameters.
+        :return: The number of parameters.
         """
         count = 0
         for model in self.model_list:
@@ -157,14 +128,11 @@ class MixedDataFitter:
 
         return count
 
-    def _get_bounds(self):
+    def _get_bounds(self) -> Tuple[np.ndarray, np.ndarray]:
         """
         Sets the bounds for each parameter based on the model list.
 
-        Returns
-        -------
-        tuple of array_like
[…]
