+++ docs/source/benchmarks.rst
@@ -1,0 +1,2 @@
+Benchmarking with ``scipy``
+===========================
