+++ docs/source/fitters/mixed_f.rst
@@ -2,6 +2,7 @@ Mixed Data Fitter
 ==================
 
 .. autoclass:: pymultifit.fitters.mixed_f.MixedDataFitter
+   :members:
    :show-inheritance:
    :private-members:
    :undoc-members:
