+++ docs/source/tutorials.rst
@@ -5,7 +5,3 @@ Tutorials
    :maxdepth: 1
 
    basics
-   generating_data
-   generating_n_model_data
-   generating_mixture_model_data
-   fitting_mixture_model
