+++ docs/source/api_index.rst
@@ -17,3 +17,4 @@ The documentation first goes through a birdseye view for these three modules, fo
    _fitters
    _generators
    _error_handling
+   _constants
