### 2024-12-31 35b4e15
[benchmarks-minor-1]
1. added `accuracy.ipynb` and `_bm_accuracy.nblink`
2. added `bench.rst` and updated `index.rst`
3. added `functions.py` in benchmarks module

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
