### 2024-12-31 bd3a8b8
[benchmarks-minor-2]
1. corrected the `foldedNormal_d.py` and `halfNormal_d.py`
2. updated `test_foldedNormal_d.py` and `test_halfNormal_d.py`
3. pdf changes in `utilities_d.py`
