+++ src/pymultifit/distributions/utilities_d.py
@@ -17,7 +17,7 @@ __all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans',
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gammainc, gammaln, owens_t
+from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t
 
 doc_style = 'numpy_napoleon_with_merge'
 
@@ -56,7 +56,7 @@ def arc_sine_pdf_(x: np.array,
 
     .. math:: f(y) = \frac{1}{\pi \sqrt{y(1-y)}}
 
-    where, :math:`y` is a transformed value of :math:`x`, defined as:
+    where, :math:`y` is the transformed value of :math:`x`, defined as:
 
     .. math:: y = \frac{x - \text{loc}}{\text{scale}}
 
@@ -105,7 +105,7 @@ def beta_pdf_(x: np.array,
 
     .. math:: f(y; \alpha, \beta) = \frac{y^{\alpha - 1} (1 - y)^{\beta - 1}}{B(\alpha, \beta)}
 
-    where :math:`B(\alpha, \beta)` is the Beta function (see, :obj:`scipy.special.beta`), and :math:`y` is a transformed value of :math:`x` such that:
+    where :math:`B(\alpha, \beta)` is the Beta function (see, :obj:`scipy.special.beta`), and :math:`y` is the transformed value of :math:`x` such that:
 
     .. math:: y = \frac{x - \text{loc}}{\text{scale}}
 
@@ -240,7 +240,7 @@ def chi_square_pdf_(x: np.array,
     .. math:: f(y\ |\ k) = \dfrac{y^{(k/2) - 1} e^{-y/2}}{2^{k/2} \Gamma(k/2)}
 
     where :math:`\Gamma(k)` is the :obj:`~scipy.special.gamma` function,
-    and :math:`y` is a transformed value of :math:`x`, defined as:
+    and :math:`y` is the transformed value of :math:`x`, defined as:
 
     .. math:: y = x - \text{loc}
 
@@ -293,7 +293,7 @@ def exponential_pdf_(x: np.array,
         0 &; y < 0.
         \end{cases}
 
-    where, :math:`y` is a transformed value of :math:`x`, defined as:
+    where, :math:`y` is the transformed value of :math:`x`, defined as:
 
     .. math:: y = x - \text{loc}
 
@@ -334,7 +334,7 @@ def exponential_cdf_(x: np.array,
 
 
 def folded_normal_pdf_(x: np.array,
-                       amplitude: float = 1., mean: float = 0.0, variance: float = 1.0,
+                       amplitude: float = 1., mean: float = 0.0, sigma: float = 1.0,
                        loc: float = 0.0, normalize: bool = False) -> np.array:
     r"""
     Compute PDF for :class:`~pymultifit.distributions.foldedNormal_d.FoldedNormalDistribution`.
@@ -349,9 +349,12 @@ def folded_normal_pdf_(x: np.array,
     mean : float, optional
         The mean parameter, :math:`\mu`.
         Defaults to 0.0.
-    variance : float, optional
-        The variance parameter, :math:`\sigma^2`.
+    sigma : float, optional
+        The standard deviation parameter, :math:`\sigma`.
         Defaults to 1.0.
+    loc : float, optional
+        The location parameter, for shifting.
+        Defaults to 0.0.
     normalize : bool, optional
[…]
