### 2024-12-31 bf5de7e
[benchmarks-minor-3]
1. corrected `chiSquare_d.py`
2. added chi_square_pdf_ and chi_square_cdf_ in `utilities_d.py`
3. updated `accuracy.ipynb`
4. updated `test_chiSquare_d.py`
