+++ benchmarks/functions.py
@@ -6,6 +6,15 @@ from matplotlib import pyplot as plt
 from pymultifit import EPSILON
 
 
+def test_and_plot(general_case, edge_case, custom_dist, scipy_dist, title):
+    accuracy_g = compare_accuracy(general_case[0], custom_dist, scipy_dist)
+    plot_accuracy(general_case[0], accuracy_g, f"General Case {title}")
+
+    for x_ in edge_case:
+        accuracy_e = compare_accuracy(x_, custom_dist, scipy_dist)
+        plot_accuracy(x_, accuracy_e, f"Edge Case {title}")
+
+
 def compare_accuracy(x, custom_dist, scipy_dist):
     pdf_custom = custom_dist.pdf(x)
     cdf_custom = custom_dist.cdf(x)
@@ -22,13 +31,13 @@ def compare_accuracy(x, custom_dist, scipy_dist):
 
 # Plotting Function
 def plot_accuracy(x, results, title_suffix):
-    plt.figure(figsize=(10, 8))
+    plt.figure(figsize=(12, 4))
 
     # Absolute Difference (PDF)
-    plt.subplot(2, 2, 1)
+    plt.subplot(1, 2, 1)
     plt.plot(x, results["pdf_abs_diff"], label="PDF Absolute Diff", marker='.')
     plt.xscale("log")
-    # plt.yscale("log")
+    plt.yscale("symlog")
     plt.xlabel("x")
     plt.ylabel("Absolute Difference (PDF)")
     plt.title(f"Absolute Difference:\nPDF {title_suffix}")
@@ -36,10 +45,10 @@ def plot_accuracy(x, results, title_suffix):
     plt.grid(True)
 
     # Absolute Difference (CDF)
-    plt.subplot(2, 2, 2)
+    plt.subplot(1, 2, 2)
     plt.plot(x, results["cdf_abs_diff"], label="CDF Absolute Diff", marker='.')
     plt.xscale("log")
-    # plt.yscale("log")
+    plt.yscale("symlog")
     plt.xlabel("x")
     plt.ylabel("Absolute Difference (CDF)")
     plt.title(f"Absolute Difference:\nCDF {title_suffix}")
