+++ src/pymultifit/benchmark_functions/__init__.py
@@ -1,3 +1,0 @@
-"""Created on Dec 31 05:45:31 2024"""
-
-from .functions import compare_accuracy, plot_accuracy
