### 2024-12-31 fa32982
[benchmarks-minor-4]
1. added `speed.ipynb`
2. updated `functions.py` and `accuracy.ipynb`
