+++ benchmarks/functions.py
@@ -1,5 +1,7 @@
 """Created on Dec 31 05:45:40 2024"""
 
+from timeit import default_timer as timer
+
 import numpy as np
 from matplotlib import pyplot as plt
 
@@ -33,7 +35,6 @@ def compare_accuracy(x, custom_dist, scipy_dist):
 def plot_accuracy(x, results, title_suffix):
     plt.figure(figsize=(12, 4))
 
-    # Absolute Difference (PDF)
     plt.subplot(1, 2, 1)
     plt.plot(x, results["pdf_abs_diff"], label="PDF Absolute Diff", marker='.')
     plt.xscale("log")
@@ -44,7 +45,6 @@ def plot_accuracy(x, results, title_suffix):
     plt.legend()
     plt.grid(True)
 
-    # Absolute Difference (CDF)
     plt.subplot(1, 2, 2)
     plt.plot(x, results["cdf_abs_diff"], label="CDF Absolute Diff", marker='.')
     plt.xscale("log")
@@ -56,4 +56,91 @@ def plot_accuracy(x, results, title_suffix):
     plt.grid(True)
 
     plt.tight_layout()
-    plt.savefig('plots/comparison {title_suffix}.png'.format(title_suffix=title_suffix))
+    plt.savefig('plots/{title_suffix}.png'.format(title_suffix=title_suffix))
+
+
+#####################################################################################################################################################
+# SPEED
+#####################################################################################################################################################
+
+
+def evaluate_speed(custom_dist, scipy_dist, n_points_list, compute_cdf=False, repetitions=100):
+    """
+    Evaluate and compare the speed of custom and SciPy distributions for PDF/CDF computations.
+
+    Parameters:
+    - custom_dist: object
+        Custom distribution instance.
+    - scipy_dist: object
+        SciPy distribution instance.
+    - n_points_list: list
+        List of number of points to evaluate (used for scalability testing).
+    - compute_cdf: bool
+        Whether to test the speed of CDF (True) or PDF (False) computations.
+    - repetitions: int
+        Number of repetitions to average the execution times.
+
+    Returns:
+    - avg_times_class: list of average execution times for the custom distribution.
+    - avg_times_scipy: list of average execution times for the SciPy distribution.
+    """
+    avg_times_class = []
+    avg_times_scipy = []
+
+    for n_points in n_points_list:
+        x = np.linspace(1e-10, 10, n_points)
+
+        # Timing for custom distribution
+        class_times = []
+        for _ in range(repetitions):
+            start_class = timer()
+            _ = custom_dist.cdf(x) if compute_cdf else custom_dist.pdf(x)
+            end_class = timer()
+            class_times.append(end_class - start_class)
+        avg_times_class.append(np.mean(class_times))
+
+        # Timing for SciPy distribution
+        scipy_times = []
+        for _ in range(repetitions):
+            start_scipy = timer()
+            _ = scipy_dist.cdf(x) if compute_cdf else scipy_dist.pdf(x)
+            end_scipy = timer()
+            scipy_times.append(end_scipy - start_scipy)
+        avg_times_scipy.append(np.mean(scipy_times))
+
+    return avg_times_class, avg_times_scipy
+
+
+def plot_speed_and_ratios(n_points_list, times_class, times_scipy, title_suffix, save_as="speed_comparison"):
+    n_points_list = np.array(n_points_list)
+
+    mean_c = np.array([np.mean(i) for i in times_class])
[…]
