### 2024-12-31 2b59648
[benchmarks-minor-4-1]
1. updated `test_chiSquare_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
