### 2024-12-31 fba008e
[benchmarks-minor-5]
1. updated `accuracy.ipynb`
2. finished `speed.ipynb`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
