### 2025-01-02 103d020
[benchmarks-minor-5]
1. minor changes in `004__fitter_module2.ipynb` and `arcSine_d.py`
2. added sdss spectra for testing purposes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
