### 2025-01-02 7cfb17c
[doc-minor-14-1]
1. updated `environment.yaml` and `requirements[dev].txt`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
