### 2025-01-02 8e18ccb
[doc-minor-15]
1. added more docstring examples
2. updated distribution docstrings
3. added scipy_like methods in various distributions
