+++ src/pymultifit/distributions/skewNormal_d.py
@@ -1,15 +1,85 @@
 """Created on Aug 03 21:35:28 2024"""
 
+from functools import wraps
 from typing import Dict
 
 import numpy as np
+from custom_inherit import doc_inherit
 
 from .backend import BaseDistribution
-from .utilities_d import skew_normal_cdf_, skew_normal_pdf_
+from .utilities_d import doc_style, skew_normal_cdf_, skew_normal_pdf_
 
 
 class SkewNormalDistribution(BaseDistribution):
-    """Class for SkewNormal distribution."""
+    r"""
+    Class for SkewNormal distribution.
+
+    :param amplitude: The amplitude of the PDF. Defaults to 1.0. Ignored if **normalize** is ``True``.
+    :type amplitude: float, optional
+
+    :param shape: The mean parameter, :math:`\mu`. Defaults to 0.0.
+    :type shape: float, optional
+
+    :param scale: The scale parameter, for scaling. Defaults to 1.0,
+    :type scale: float, optional
+
+    :param location: The location parameter, for shifting. Defaults to 0.0.
+    :type location: float, optional
+
+    :param normalize: If ``True``, the distribution is normalized so that the total area under the PDF equals 1. Defaults to ``False``.
+    :type normalize: bool, optional
+
+    :raise NegativeAmplitudeError: If the provided value of amplitude is negative.
+    :raise NegativeStandardDeviationError: If the provided value of standard deviation is negative.
+
+    Examples
+    --------
+    Importing libraries:
+
+    .. literalinclude:: ../../../examples/basic/skewnormal.py
+       :language: python
+       :linenos:
+       :lineno-start: 3
+       :lines: 3-7
+
+    Generating a standard Skew Normal(:math:`\xi=1, \mu = 0, \sigma = 1`) distribution with ``pyMultiFit`` and ``scipy``:
+
+    .. literalinclude:: ../../../examples/basic/skewnormal.py
+       :language: python
+       :linenos:
+       :lineno-start: 9
+       :lines: 9-12
+
+    Plotting **PDF** and **CDF**:
+
+    .. literalinclude:: ../../../examples/basic/skewnormal.py
+       :language: python
+       :linenos:
+       :lineno-start: 14
+       :lines: 14-29
+
+    .. image:: ../../../images/skew_norm_example1.png
+       :alt: SkewNormal(1, 0, 1)
+       :align: center
+
+    Generating a translated Skew Normal(:math:`\xi=3, \mu=-3, \sigma=3`) distribution:
+
+    .. literalinclude:: ../../../examples/basic/skewnormal.py
+       :language: python
+       :lineno-start: 32
+       :lines: 32
+
+    Plotting **PDF** and **CDF**:
+
+    .. literalinclude:: ../../../examples/basic/skewnormal.py
+       :language: python
+       :lineno-start: 34
+       :lines: 34-49
+
+    .. image:: ../../../images/skew_norm_example2.png
+       :alt: Skew Normal(3, -3, 3)
+       :align: center
+    """
 
     def __init__(self, amplitude: float = 1.0, shape: float = 1., location: float = 0., scale: float = 1., normalize: bool = False):
         self.amplitude = 1 if normalize else amplitude
@@ -18,6 +88,29 @@ class SkewNormalDistribution(BaseDistribution):
         self.scale = scale
 
         self.norm = normalize
+
+    @classmethod
+    @doc_inherit(parent=BaseDistribution.scipy_like, style=doc_style)
+    def scipy_like(cls, a, loc: float = 0.0, scale: float = 1.0):
+        """
+
+        Parameters
+        ----------
+        a : float
+            The shape parameter for the distribution.
+
+        loc : float, optional
[…]
