### 2025-01-02 1286d6c
[doc-minor-15-1]
1. testing for `beta_d.rst`
