### 2025-01-03 efc196e
[doc-minor-16]
1. things might be broken in this commit, have to thoroughly check everything

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
