### 2025-01-03 930567b
[scipy_like-major-1]
1. added `scipy_like` function to all distributions
2. added code examples to `uniform_d.py`
