### 2025-01-03 6cc1a5c
[doc-minor-17]
updated docstrings
