+++ src/pymultifit/generators/generators.py
@@ -9,16 +9,16 @@ from .. import distributions as dist, doc_style, GAUSSIAN, LAPLACE, LINE, LOG_NO
 from ..distributions.backend import BaseDistribution
 
 listOfTuples = List[Tuple[float, ...]]
-listOfTuplesOrArray = Union[listOfTuples, np.array]
+listOfTuplesOrArray = Union[listOfTuples, np.ndarray]
 
 
-def multi_base(x: np.array, distribution_func: Type[BaseDistribution], params: listOfTuplesOrArray,
-               noise_level: float = 0.0, normalize: bool = False) -> np.array:
+def multi_base(x: np.ndarray, distribution_func: Type[BaseDistribution], params: listOfTuplesOrArray,
+               noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
     """Generate data based on a combination of distributions with optional noise.
 
     Parameters
     ----------
-    x : np.array
+    x : np.ndarray
         Input array of values.
     distribution_func: Type[BaseDistribution]
         The distribution function to be used to generate data.
@@ -33,7 +33,7 @@ def multi_base(x: np.array, distribution_func: Type[BaseDistribution], params: l
 
     Returns
     -------
-    np.array
+    np.ndarray
         Array of the same shape as :math:`x`, containing the evaluated values.
     """
     y = np.zeros_like(x, dtype=float)
@@ -49,14 +49,14 @@ def multi_base(x: np.array, distribution_func: Type[BaseDistribution], params: l
     return y
 
 
-def multi_chi_squared(x: np.array, params: listOfTuplesOrArray,
-                      noise_level: float = 0.0, normalize: bool = False) -> np.array:
+def multi_chi_squared(x: np.ndarray, params: listOfTuplesOrArray,
+                      noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
     r"""
     Generate multi-:class:`~pymultifit.distributions.chiSquare_d.ChiSquareDistribution` data with optional noise.
 
     Parameters
     ----------
-    x : np.array
+    x : np.ndarray
         Input array of values.
     params : listOfTuplesOrArray
         List of tuples containing the parameters for the required distribution.
@@ -69,7 +69,7 @@ def multi_chi_squared(x: np.array, params: listOfTuplesOrArray,
 
     Returns
     -------
-    np.array
+    np.ndarray
         Array of the same shape as :math:`x`, containing the evaluated values.
     """
     return multi_base(x=x,
@@ -78,8 +78,8 @@ def multi_chi_squared(x: np.array, params: listOfTuplesOrArray,
 
 
 @doc_inherit(parent=multi_chi_squared, style=doc_style)
-def multi_gamma_sr(x: np.array, params: listOfTuplesOrArray,
-                   noise_level: float = 0.0, normalize: bool = False) -> np.array:
+def multi_gamma_sr(x: np.ndarray, params: listOfTuplesOrArray,
+                   noise_level: float = 0.0, normalize: bool = False) -> np.ndarray:
     r"""Generate multi-:class:`~pymultifit.distributions.gamma_d.GammaDistributionSR` data with optional noise."""
     return multi_base(x=x,
                       distribution_func=dist.GammaDistributionSR, params=params, noise_level=noise_level,
@@ -87,8 +87,8 @@ def multi_gamma_sr(x: np.array, params: listOfTuplesOrArray,
 
 
 @doc_inherit(parent=multi_chi_squared, style=doc_style)
-def multi_gamma_ss(x: np.array, params: listOfTuplesOrArray,
-                   noise_level: float = 0.0, normalize: bool = False) -> np.array:
+def multi_gamma_ss(x: np.ndarray, params: listOfTuplesOrArray,
[…]
