+++ src/pymultifit/fitters/skewNormal_f.py
@@ -8,7 +8,7 @@ from ..distributions.utilities_d import skew_normal_pdf_
 
 
 class SkewNormalFitter(BaseFitter):
-    """A class for fitting multiple Skewed Normal functions to the given data."""
+    """A class for fitting multiple SkewNormal distributions to the given data."""
 
     def __init__(self, x_values, y_values, max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
