+++ src/pymultifit/fitters/logNormal_f.py
@@ -12,7 +12,7 @@ from ..distributions.utilities_d import log_normal_pdf_
 
 
 class LogNormalFitter(BaseFitter):
-    """Class for multi-LogNormal fitting."""
+    """A class for fitting multiple LogNormal distributions to the given data."""
 
     def __init__(self, x_values, y_values, max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
