### 2025-01-03 1cff8f5
[doc-minor-19]
1. added `test_skewNormal_d.py`
