+++ docs/source/distributions/uniform_d.rst
@@ -5,6 +5,7 @@ Uniform Distribution
    :members:
    :inherited-members:
    :show-inheritance:
+   :member-order: groupwise
 
 This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 
