+++ docs/source/distributions/gamma_sr_d.rst
@@ -5,6 +5,7 @@ Gamma Distribution (SR)
    :members:
    :inherited-members:
    :show-inheritance:
+   :member-order: groupwise
 
 
 .. note::
