+++ docs/source/distributions/gamma_ss_d.rst
@@ -5,6 +5,7 @@ Gamma Distribution (SS)
    :members:
    :inherited-members:
    :show-inheritance:
+   :member-order: groupwise
 
 
 .. note::
