### 2025-01-03 03e27dd
[doc-minor-19-1]
1. minor correction in `test_gamma_sr_d.py`
