### 2025-01-04 b7b9b63
[tutorials-minor-1]
1. added `exploration1.ipynb`
2. updated the link in exploration1.nblink
3. updated `baseFitter.py` and `tutorials.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
