### 2025-01-07 68fcf64
[tutorials-minor-2]
1. minor changes in `sdss_.py`
2. added `spec-0293-51994-0078.fits`
3. updated `tutorials.rst`
4. added `exploration2.ipynb` and `exploration2.nblink`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
