+++ rough_docstring.txt
@@ -6,7 +6,7 @@ PFD generic docstring
 
     Parameters
     ----------
-    x : np.array
+    x : np.ndarray
         Input array of values.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
@@ -23,7 +23,7 @@ PFD generic docstring
 
     Returns
     -------
-    np.array
+    np.ndarray
         Array of the same shape as :math:`x`, containing the evaluated values.
     """
 
