+++ docs/source/distributions/others.rst
@@ -1,0 +1,10 @@
+Other function
+==============
+
+.. py:currentmodule:: pymultifit.distributions.others
+
+.. autofunction:: line
+.. autofunction:: linear
+.. autofunction:: quadratic
+.. autofunction:: cubic
+.. autofunction:: nth_polynomial
