+++ docs/source/generators/generators.rst
@@ -1,7 +1,17 @@
 Generators
 ==========
 
-.. automodule:: pymultifit.generators.generators
-   :members:
-   :undoc-members:
-   :show-inheritance:
+.. py:currentmodule:: pymultifit.generators.generators
+
+.. autofunction:: multi_base
+.. autofunction:: multi_chi_squared
+.. autofunction:: multi_exponential
+.. autofunction:: multi_folded_normal
+.. autofunction:: multi_gamma_sr
+.. autofunction:: multi_gamma_ss
+.. autofunction:: multi_gaussian
+.. autofunction:: multi_half_normal
+.. autofunction:: multi_laplace
+.. autofunction:: multi_log_normal
+.. autofunction:: multi_skew_normal
+.. autofunction:: multiple_models
