+++ docs/source/baseDistributionMethods.rst
@@ -1,20 +1,0 @@
-.. _base_distribution_methods:
-
-Methods
--------
-
-+-------------------+---------------------------------------------------------------+
-| **Method**        | **Description**                                               |
-+===================+===============================================================+
-| ``cdf(x)``        | Compute the cumulative density function (CDF) for the         |
-|                   | distribution.                                                 |
-+-------------------+---------------------------------------------------------------+
-| ``pdf(x)``        | Compute the probability density function (PDF) for the        |
-|                   | distribution.                                                 |
-+-------------------+---------------------------------------------------------------+
-| ``scipy_like()``  | Create an instance of the class in a format similar to|
-|                   | scipy's parameterization.                                     |
-+-------------------+---------------------------------------------------------------+
-| ``stats()``       | Computes and returns the statistical properties of the        |
-|                   | distribution.                                                 |
-+-------------------+---------------------------------------------------------------+
