### 2025-01-07 8695156
[main-minor-9-2]
testing issues with readthedocs

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
