+++ setup.py
@@ -25,7 +25,7 @@ setup(name='pymultifit',
       license=metadata_['__license__'],
       url=metadata_['__url__'],
       description=metadata_['__description__'],
-      packages=find_packages(where="src", exclude=["src/tests"]),
+      packages=find_packages(where="src", exclude=["examples", "*tests*"]),
       package_dir={"": "src"},
       long_description=readme,
       long_description_content_type="text/markdown",
