+++ docs/source/_distributions.rst
@@ -23,7 +23,7 @@ Derived Distributions
    * - Name
      - Description
      - PDF function
-   * - :class:`ArcSineDistribution <pymultifit.distributions.arcSine_d.ArcSineDistribution>`
+   * - :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`
      - ArcSine distribution.
      - :math:`\dfrac{1}{\pi\sqrt{x(1-x)}}`
    * - :class:`BetaDistribution <pymultifit.distributions.beta_d.BetaDistribution>`
