+++ docs/source/distributions/arcSine_d.rst
@@ -1,8 +1,6 @@
 ArcSine Distribution
 ====================
 
-.. py:currentmodule:: pymultifit.distributions
-
 .. autoclass:: pymultifit.distributions.arcSine_d.ArcSineDistribution
    :members:
    :inherited-members:
@@ -12,8 +10,8 @@ ArcSine Distribution
 
 This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 
-* :func:`~utilities_d.beta_pdf_`
-* :func:`~utilities_d.beta_cdf_`
+* :func:`~pymultifit.distributions.utilities_d.beta_pdf_`
+* :func:`~pymultifit.distributions.utilities_d.beta_cdf_`
 
 Recommended Import
 ^^^^^^^^^^^^^^^^^^
