### 2025-01-07 c072baa
[main-minor-9-5]
testing issues with readthedocs

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
