+++ docs/source/distributions/arcSine_d.rst
@@ -1,7 +1,9 @@
 ArcSine Distribution
 ====================
 
-.. autoclass::`~pymultifit.distributions.arcSine_d.ArcSineDistribution`
+.. py:currentmodule:: pymultifit.distributions
+
+.. autoclass:: pymultifit.distributions.arcSine_d.ArcSineDistribution
    :members:
    :inherited-members:
    :show-inheritance:
@@ -10,8 +12,8 @@ ArcSine Distribution
 
 This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
 
-* :func:`~pymultifit.distributions.utilities_d.beta_pdf_`
-* :func:`~pymultifit.distributions.utilities_d.beta_cdf_`
+* :func:`~utilities_d.beta_pdf_`
+* :func:`~utilities_d.beta_cdf_`
 
 Recommended Import
 ^^^^^^^^^^^^^^^^^^
