### 2025-01-07 03170ff
[main-minor-9-6]
testing issues with readthedocs

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
