+++ src/pymultifit/distributions/utilities_d.py
@@ -15,6 +15,8 @@ __all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans',
            'skew_normal_pdf_', 'skew_normal_cdf_',
            'uniform_pdf_', 'uniform_cdf_']
 
+from typing import Union
+
 import numpy as np
 from custom_inherit import doc_inherit
 from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t
@@ -22,15 +24,15 @@ from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t
 from .. import doc_style
 
 
-def arc_sine_pdf_(x: np.array,
+def arc_sine_pdf_(x: np.ndarray,
                   amplitude: float = 1.0,
-                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
+                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.ndarray:
     r"""
     Compute PDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
     Parameters
     ----------
-    x : np.array
+    x : np.ndarray
         Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
@@ -47,7 +49,7 @@ def arc_sine_pdf_(x: np.array,
 
     Returns
     -------
-    np.array
+    np.ndarray
         Array of the same shape as :math:`x`, containing the evaluated PDF values.
 
     Notes
@@ -65,15 +67,15 @@ def arc_sine_pdf_(x: np.array,
     return beta_pdf_(x=x, amplitude=amplitude, alpha=0.5, beta=0.5, loc=loc, scale=scale, normalize=normalize)
 
 
-def beta_pdf_(x: np.array,
+def beta_pdf_(x: np.ndarray,
               amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
-              loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
+              loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.ndarray:
     r"""
     Compute PDF of :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
 
     Parameters
     ----------
-    x : np.array
+    x : np.ndarray
         Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
@@ -96,7 +98,7 @@ def beta_pdf_(x: np.array,
 
     Returns
     -------
-    np.array
+    np.ndarray
         Array of the same shape as `x`, containing the evaluated PDF values.
 
     Notes
@@ -141,9 +143,9 @@ def beta_pdf_(x: np.array,
 
 
 @doc_inherit(parent=beta_pdf_, style=doc_style)
-def beta_logpdf_(x: np.array,
+def beta_logpdf_(x: np.ndarray,
                  amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
-                 loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
+                 loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.ndarray:
     r"""Compute logPDF for :class:`~pymultifit.distributions.beta_d.BetaDistribution`."""
     if x.size == 0:
         return np.array([])
@@ -171,15 +173,15 @@ def beta_logpdf_(x: np.array,
 
 
 @doc_inherit(parent=beta_pdf_, style=doc_style)
-def beta_cdf_(x: np.array,
+def beta_cdf_(x: np.ndarray,
               amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
[…]
