### 2025-01-07 7156772
[main-minor-9-7]
testing issues with readthedocs

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
