### 2025-01-09 206404e
[property-fix-minor-1]
1. fixed the issue with property complexity.
