+++ src/pymultifit/distributions/chiSquare_d.py
@@ -122,25 +122,14 @@ class ChiSquareDistribution(BaseDistribution):
         return chi_square_cdf_(x, amplitude=self.amplitude, degree_of_freedom=self.dof, loc=self.loc, scale=self.scale,
                                normalize=self.norm)
 
-    @property
-    def mean(self) -> float:
-        return (self.scale * self.dof) + self.loc
-
-    @property
-    def variance(self) -> float:
-        return 2 * self.dof * self.scale**2
-
-    @property
-    def stddev(self) -> float:
-        return np.sqrt(self.variance)
-
-    @property
-    def mode(self):
-        return max(self.dof - 2, 0)
-
     def stats(self) -> Dict[str, float]:
-        return {'mean': self.mean,
-                'median': self.median,
-                'mode': self.mode,
-                'variance': self.variance,
-                'std': self.stddev}
+        df = self.dof
+        s, l_ = self.scale, self.loc
+
+        mean_ = (s * df) + l_
+        mode_ = max(df - 2, 0)
+        variance_ = 2 * df * s**2
+        return {'mean': mean_,
+                'mode': mode_,
+                'variance': variance_,
+                'std': np.sqrt(variance_)}
