### 2025-01-09 46989e2
[property-fix-minor-2]
1. minor fixes
