### 2025-01-12 344f10f
[streamlining-minor-1]
1. added a frozen preprocessing function in `baseFitter.py`
2. minor corrections

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
