+++ examples/gaussian.py
@@ -3,8 +3,8 @@
 import numpy as np
 from matplotlib import pyplot as plt
 
-from src.pymultifit.fitters import GaussianFitter
-from src.pymultifit.generators import multi_gaussian
+from pymultifit.fitters import GaussianFitter
+from pymultifit.generators import multi_gaussian
 
 params = [(20, -20, 2), (4, -5.5, 10), (5, -1, 0.5), (10, 3, 1), (4, 15, 3)]
 
@@ -19,5 +19,7 @@ guess = [(10, -18, 1), (4, -5.5, 10), (5, -1, 0.5), (10, 3, 1), (4, 15, 3)]
 
 fitter.fit(guess)
 
-plotter = fitter.plot_fit(show_individual=True, x_label='X_data', y_label='Y_data', title='XY_plot', data_label='XY_data')
+f, ax = plt.subplots(1, 1, figsize=(12, 6))
+plotter = fitter.plot_fit(show_individual=True, x_label='X_data', y_label='Y_data', title='XY_plot',
+                          data_label='XY_data', axis=ax)
 plt.show()
