### 2025-01-13 018d7cc
[streamlining-minor-2]
1. updated rst files
2. updated for next version in `conf.py`, and `version.py`
3. added Line class for `others.py`, moved polynomial functions to `polynomials.py`
4. minor changes in `utilities_d.py`
5. major updated in `mixed_f.py`
6. added `others.py` in fitters as well

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
