+++ src/pymultifit/fitters/mixed_f.py
@@ -9,15 +9,34 @@ from mpyez.backend.uPlotting import LinePlot
 from mpyez.ezPlotting import plot_xy
 from scipy.optimize import curve_fit
 
-from .utilities_f import sanity_check, _Line
-from .. import GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, NORMAL, SKEW_NORMAL
-from .. import distributions as dist
+# importing from files to avoid circular import
+from .chiSquare_f import ChiSquareFitter
+from .exponential_f import ExponentialFitter
+from .foldedNormal_f import FoldedNormalFitter
+from .gamma_f import GammaFitterSR, GammaFitterSS
+from .gaussian_f import GaussianFitter
+from .halfNormal_f import HalfNormalFitter
+from .laplace_f import LaplaceFitter
+from .logNormal_f import LogNormalFitter
+from .others import LineFitter
+from .skewNormal_f import SkewNormalFitter
+from .utilities_f import sanity_check
+from .. import (GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL, CHI_SQUARE, EXPONENTIAL, FOLDED_NORMAL, GAMMA_SR,
+                GAMMA_SS, NORMAL, HALF_NORMAL)
 
-model_dict = {LINE: [_Line, 2],
-              GAUSSIAN: [dist.GaussianDistribution, 3],
-              LOG_NORMAL: [dist.LogNormalDistribution, 3],
-              SKEW_NORMAL: [dist.SkewNormalDistribution, 4],
-              LAPLACE: [dist.LaplaceDistribution, 3]}
+# mock initialize the internal classes for auto MixedDataFitter class
+fitter_dict = {CHI_SQUARE: ChiSquareFitter,
+               EXPONENTIAL: ExponentialFitter,
+               FOLDED_NORMAL: FoldedNormalFitter,
+               GAMMA_SR: GammaFitterSR,
+               GAMMA_SS: GammaFitterSS,
+               GAUSSIAN: GaussianFitter,
+               NORMAL: GaussianFitter,
+               HALF_NORMAL: HalfNormalFitter,
+               LAPLACE: LaplaceFitter,
+               LOG_NORMAL: LogNormalFitter,
+               SKEW_NORMAL: SkewNormalFitter,
+               LINE: LineFitter}
 
 
 class MixedDataFitter:
@@ -31,7 +50,7 @@ class MixedDataFitter:
     """
 
     def __init__(self, x_values: Union[List, np.ndarray], y_values: Union[List, np.ndarray],
-                 model_list: List[str], max_iterations: int = 1000):
+                 model_list: List[str], max_iterations: int = 1000, fitter_dictionary=None):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
 
         self.x_values = x_values
@@ -41,13 +60,49 @@ class MixedDataFitter:
         self.params = None
         self.covariance = None
 
-        # Validate the model list and create the model function
+        self.fitter_dict = fitter_dictionary or fitter_dict
+
         self._validate_models()
         self.model_function = self._create_model_function()
 
     def __repr__(self):
         return (f"{self.__class__.__name__}(x_values={self.x_values}, y_values={self.y_values}, "
                 f"model_list={self.model_list}, max_iterations={self.max_iterations})")
+
+    def _instantiate_fitter(self, model: str, return_values: Union[str, List[str]] = 'class'):
+        """
+        Instantiate the fitter for the specified model and return requested values.
+
+        :param model: The model name as a string.
+        :param return_values: The specific attribute(s) or instance to return.
+                              Options are 'class', 'n_par', and 'bounds'.
+                              Can be a string (for one value) or a list of strings.
[…]

[…]
