### 2025-01-13 279389a
[streamlining-minor-3]
1. docs/sources
    1. added `exploration3.nblink`
    2. updated `conf.py`
    3. updated `tutorials.rst`
2. src/pymultifit
    1. updated `mixed_f.py` to include a fitter_dictionary for easy user mixed fitting access
    2. updated `generators.py` for better `multiple_model` data generation
3, tutorials
    1. added `exploration3.ipynb`
    2. updated `004__fitter_module2.ipynb`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
