### 2025-01-14 c352303
[streamlining-minor-4]
1. src/pymultifit/fitters
    1. renamed `get_parameters` to `get_model_parameters` in `baseFitter.py`
    2. added `get_value_error_pair` and modified `get_parameters` function in `mixed_f.py`
2. tutorials updated

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
