### 2025-01-14 1463fad
[streamlining-minor-5]
1. updated tutorials
2. removed the restriction for adding `fit_boundaries` in `mixed_f.py` as well
