### 2025-01-14 ca530d9
[streamlining-minor-6]
1. frozen parameters are now working in `mixed_f.py`
2. cosmetic changes in `mixed_f.py`
