+++ src/pymultifit/fitters/mixed_f.py
@@ -7,7 +7,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from mpyez.backend.uPlotting import LinePlot
 from mpyez.ezPlotting import plot_xy
-from scipy.optimize import curve_fit
+from scipy.optimize import Bounds, curve_fit
 
 # importing from files to avoid circular import
 from .chiSquare_f import ChiSquareFitter
@@ -21,8 +21,8 @@ from .logNormal_f import LogNormalFitter
 from .others import LineFitter
 from .skewNormal_f import SkewNormalFitter
 from .utilities_f import sanity_check
-from .. import (GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL, CHI_SQUARE, EXPONENTIAL, FOLDED_NORMAL, GAMMA_SR,
-                GAMMA_SS, NORMAL, HALF_NORMAL)
+from .. import (epsilon, GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL, CHI_SQUARE, EXPONENTIAL, FOLDED_NORMAL, GAMMA_SR, GAMMA_SS, NORMAL,
+                HALF_NORMAL)
 
 # mock initialize the internal classes for auto MixedDataFitter class
 fitter_dict = {CHI_SQUARE: ChiSquareFitter,
@@ -68,46 +68,6 @@ class MixedDataFitter:
     def __repr__(self):
         return (f"{self.__class__.__name__}(x_values={self.x_values}, y_values={self.y_values}, "
                 f"model_list={self.model_list}, max_iterations={self.max_iterations})")
-
-    def _instantiate_fitter(self, model: str, return_values: Union[str, List[str]] = 'class'):
-        """
-        Instantiate the fitter for the specified model and return requested values.
-
-        :param model: The model name as a string.
-        :param return_values: The specific attribute(s) or instance to return.
-                              Options are 'class', 'n_par', and 'bounds'.
-                              Can be a string (for one value) or a list of strings.
-        :return: The requested values as a single value or a tuple.
-        :raises ValueError: If the model is not recognized or return_values are invalid.
-        """
-        try:
-            fitter_instance = self.fitter_dict[model](x_values=[], y_values=[])
-        except KeyError:
-            raise ValueError(f"Model '{model}' not recognized. Ensure it is defined in the fitter dictionary.")
-
-        valid_options = {'class', 'n_par', 'bounds'}
-        if isinstance(return_values, str):
-            return_values = [return_values]  # Convert to list for uniform processing
-
-        if not all(val in valid_options for val in return_values):
-            invalid_options = [val for val in return_values if val not in valid_options]
-            raise ValueError(f"Invalid return_values {invalid_options}. Expected values: {valid_options}")
-
-        result = []
-        for val in return_values:
-            if val == 'class':
-                result.append(fitter_instance)
-            elif val == 'n_par':
-                result.append(fitter_instance.n_par)
-            elif val == 'bounds':
-                try:
-                    result.append(fitter_instance.fit_boundaries())
-                except NotImplementedError:
-                    # in case the boundaries are not defined, put -np.inf, and np.inf to work with
-                    n_par = fitter_instance.n_par
-                    result.append([[-np.inf] * n_par, [np.inf] * n_par])
-
-        return result[0] if len(result) == 1 else tuple(result)
 
     def _create_model_function(self) -> Callable:
         """
@@ -157,81 +117,6 @@ class MixedDataFitter:
 
         return count
 
-    def _get_bounds(self) -> Tuple[np.ndarray, np.ndarray]:
-        """
-        Sets the bounds for each parameter based on the model list.
-
[…]
