### 2025-01-19 #221 @SimonMolinsky
@syedalimohsinbukhari could you check your documentation on Read the Docs? I got the 404 error when using this link: https://pymultifit.readthedocs.com/latest/
