### 2025-01-19 #221 @syedalimohsinbukhari
Hi @SimonMolinsky

The documentation is up and working on https://pymultifit.readthedocs.io/latest, not on the .com version. I don't know how I messed up io with com, but I'll correct it rightaway

*EDIT* The link is now corrected in the README.md file
