### 2025-01-19 #221 @SimonMolinsky
Hi!

I've checked the benchmarks and tutorials, and a few comments from me:

- https://pymultifit.readthedocs.io/latest/bench.html - `This benchmarking file compares the performance and accuracy of the custom distributions implemented in your_library against their counterparts in SciPy (scipy.stats).`

Change the placeholder named `your_library` to the name of your package.

- **benchmarking** - generally it's ok, but it would be nice to introduce some summary at the top of the page (not for every plot, but for accuracy and speed pages)

- The tutorials are great!

I have the last two questions:

1. How have you achieved a computing time decrease over `Scipy`? You use mostly `numpy`; haven't you skipped some checks performed by `scipy` functions to achieve those results? Or is there another reason for the faster computations?
2. Do you plan to maintain and develop this package in the future?
