### 2025-01-19 #221 @syedalimohsinbukhari
Hi,

Thanks for taking the time to check it out. I will surely include the summary on the benchmarking pages and remove that placeholder mark.

> I have the last two questions:
>
> 1. How have you achieved a computing time decrease over `Scipy`? You use mostly `numpy`; haven't you skipped some checks performed by `scipy` functions to achieve those results? Or is there another reason for the faster computations?

From my own understanding, the reason for the decrease in computational time from `Scipy` mostly stems from the fact that my implementations are singular in purpose and do not require as many checks as the scipy requires. This did require a lot of tweaking in some distributions to match the Scipy's values, the results of which you can see in the accuracy benchmark file.

My reasoning for this answer comes from slower/comparable computation times with Beta/ArcSine/SkewNormal distributions that **required** extensive checks/calling other distributions in order to match the values by Scipy that incurred heavy overheads.

For example, BetaDistribution -> beta_pdf_ uses various masks for different values of alpha and beta parameters

https://github.com/syedalimohsinbukhari/pyMultiFit/blob/e1777ea6a5391839e14b9cefd43c78ebff0d73d6/src/pymultifit/distributions/utilities_d.py#L117-L143

Also, SkewNormalDistribution -> skew_normal_pdf_ internally calls gaussian_pdf_ and gaussian_cdf_ functions

https://github.com/syedalimohsinbukhari/pyMultiFit/blob/e1777ea6a5391839e14b9cefd43c78ebff0d73d6/src/pymultifit/distributions/utilities_d.py#L1131-L1143

> 3. Do you plan to maintain and develop this package in the future?

Yes, I do.
