### 2025-01-19 #221 @SimonMolinsky
@syedalimohsinbukhari

Considering your input, your package has the potential to be a great tool in the `pyOpenSci`!

**You might submit `pymultifit` to the `pyOpenSci` ecosystem.**

I'm still slightly unsure about the overlap between your package and `scipy`, but you have provided extensive tutorials and benchmarks for the community, and this is a positive signal! Moreover, you intend to maintain `pymultifit` in the future, so I think that the overlap between `scipy` and `pymultifit` will be less noticeable.

Again, feel free to submit your package! I will close this thread when you submit the package.

Have a nice day!
