### 2025-01-19 #221 @syedalimohsinbukhari
Hi @SimonMolinsky

Thank you for getting back, and thank you for your assessment. I will submit the package, but I just want to be clear as to what you mean by the overlap between `pyMultiFit` and `scipy` (just for my own clearance).
