### 2025-01-20 #221 @SimonMolinsky
I meant the capabilities (not functionalities and their  API implementation).
