+++ benchmarks/functions.py
@@ -3,7 +3,9 @@
 from timeit import default_timer as timer
 
 import numpy as np
+import pandas as pd
 from matplotlib import pyplot as plt
+from matplotlib.ticker import FixedLocator
 
 from pymultifit import EPSILON
 
@@ -125,7 +127,7 @@ def plot_speed_and_ratios(n_points_list, times_class, times_scipy, title_suffix,
     plt.savefig(f"plots/{save_as}_{title_suffix}.png")
 
 
-def cdf_pdf_plots(custom_dist, scipy_dist, n_points, save_as: str, repetitions: int = 10):
+def cdf_pdf_plots(custom_dist, scipy_dist, n_points, save_as: str, repetitions: int = 15):
     p_times_class, p_times_scipy = evaluate_speed(custom_dist=custom_dist, scipy_dist=scipy_dist,
                                                   n_points_list=n_points, compute_cdf=False, repetitions=repetitions)
     plot_speed_and_ratios(n_points, p_times_class, p_times_scipy, "PDF Computations", save_as=save_as)
@@ -133,3 +135,134 @@ def cdf_pdf_plots(custom_dist, scipy_dist, n_points, save_as: str, repetitions: 
     c_times_class, c_times_scipy = evaluate_speed(custom_dist=custom_dist, scipy_dist=scipy_dist,
                                                   n_points_list=n_points, compute_cdf=False, repetitions=repetitions)
     plot_speed_and_ratios(n_points, c_times_class, c_times_scipy, "CDF Computations", save_as=save_as)
+
+    return (p_times_class, c_times_class), (p_times_scipy, c_times_scipy)
+
+
+def plot_distribution_comparison(data_dict, title_labels=('PDF', 'CDF')):
+    """
+    Automates the plotting of PDF and CDF timing comparisons for multiple distributions.
+
+    Parameters:
+    - data_dict: Dictionary with keys as distribution names and values as tuples
+                 [(custom_PDF, scipy_PDF), (custom_CDF, scipy_CDF)]
+    - title_labels: List of subplot titles (e.g., ['PDF', 'CDF'])
+
+    Example usage:
+    plot_distribution_comparison({
+        'Exp': ([pdf_custom_exp, pdf_scipy_exp], [cdf_custom_exp, cdf_scipy_exp]),
+        'Unif': ([pdf_custom_unif, pdf_scipy_unif], [cdf_custom_unif, cdf_scipy_unif]),
+        'Laplace': ([pdf_custom_laplace, pdf_scipy_laplace], [cdf_custom_laplace, cdf_scipy_laplace])
+    })
+    """
+
+    f, ax = plt.subplots(len(title_labels), 1, figsize=(18, 6 * len(title_labels)))
+
+    if len(title_labels) == 1:
+        ax = [ax]  # Ensure consistent indexing for single plot
+
+    for i, title in enumerate(title_labels):
+        log_data = []
+        xtick_labels = []
+
+        for dist_name, (pdf_cdf_pair) in data_dict.items():
+            pdf_or_cdf = pdf_cdf_pair[i]  # i = 0 for PDF, i = 1 for CDF
+            log_data.append(np.log10(pdf_or_cdf[0]))  # Custom timing
+            log_data.append(np.log10(pdf_or_cdf[1]))  # SciPy timing
+            xtick_labels.append(f'custom\n{dist_name}')
+            xtick_labels.append(f'scipy\n{dist_name}')
+
+        ax[i].boxplot(log_data, meanline=True, showmeans=True)
+        ax[i].set_xticklabels(xtick_labels, rotation=60, ha="center")
+        ax[i].set_title(title)
+
+    plt.xlabel('Log[Time] [s]')
+    plt.tight_layout()
+    plt.show()
+
+
+def generate_data_dict(data_list, label_list):
+    """
[…]
