+++ environment.yaml
@@ -8,6 +8,8 @@ dependencies:
     - matplotlib
     - scipy
     - custom-inherit
+    - seaborn
+    - pandas
     - pytest
     - astropy
     - jupyter
