### 2025-01-21 2e023e9
[summary-minor-2]
1. benchmarks
    1. updated `functions.py`
    2. updated `speed.ipynb` and `summary.ipynb`
2. docs/source
    1. added `_bm_summary.nblink`
    2. updated `bench.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
