### 2025-01-21 e2fb240
[summary-minor-4]
1. structural changes in API
2. added uniform edge cases and general cases in benchmarks
3. minor modifications
