+++ docs/source/_generators.rst
@@ -1,57 +1,0 @@
-``generators``
-==============
-
-Template function
------------------
-
-.. list-table::
-   :align: center
-   :header-rows: 1
-
-   * - Name
-     - Description
-   * - :meth:`~pymultifit.generators.generators.multi_base`
-     - The template function for the generator functions.
-
-Standalone generators
----------------------
-
-.. list-table::
-   :align: center
-   :header-rows: 1
-
-   * - Name
-     - Description
-   * - :meth:`~pymultifit.generators.generators.multiple_models`
-     - Generator for data based on a combination of different models.
-
-Derived generators
-------------------
-
-.. list-table::
-   :align: center
-   :header-rows: 1
-
-   * - Name
-     - Description
-   * - :meth:`~pymultifit.generators.generators.multi_chi_squared`
-     - Chi-Squared distribution generator.
-   * - :meth:`~pymultifit.generators.generators.multi_exponential`
-     - Exponential distribution generator.
-   * - :meth:`~pymultifit.generators.generators.multi_folded_normal`
-     - Folded Normal distribution generator.
-   * - :meth:`~pymultifit.generators.generators.multi_gaussian`
-     - Gaussian distribution generator.
-   * - :meth:`~pymultifit.generators.generators.multi_half_normal`
-     - Half-Normal distribution generator.
-   * - :meth:`~pymultifit.generators.generators.multi_laplace`
-     - Laplace distribution generator.
-   * - :meth:`~pymultifit.generators.generators.multi_log_normal`
-     - Log-Normal distribution generator.
-   * - :meth:`~pymultifit.generators.generators.multi_skew_normal`
-     - Skewed Normal distribution generator.
-
-.. toctree::
-   :hidden:
-
-   GeneratorFunctions         <generators/generators>
