+++ docs/source/benchmarks/benchmarks.rst
@@ -12,9 +12,6 @@ The purpose of these benchmarks is to evaluate:
 2. **Accuracy**:
     - Compare the results of key operations (e.g., PDF, CDF) against SciPy's outputs to ensure numerical correctness.
 
-3. **Usability**:
-    - Highlight edge cases and parameter ranges where the custom implementations deviate or fail compared to SciPy.
-
 Each benchmark will output timing information and accuracy metrics, providing insights into the trade-offs between speed and precision.
 
 Key Benchmarked Operations
@@ -29,7 +26,7 @@ Key Benchmarked Operations
     - Used widely in hypothesis testing and statistical applications.
 
 Accuracy Metrics:
-    - Mean Absolute Error (MAE) and Relative Error will be calculated between custom and SciPy distributions for PDF and CDF outputs.
+    - Absolute Error will be calculated between custom and SciPy distributions for PDF and CDF outputs.
 
 Benchmark Workflow
 -------------------
@@ -44,10 +41,10 @@ Each distribution (e.g., Gaussian, Laplace) will be benchmarked as follows:
     - Measure execution times for PDF and CDF for both the custom and SciPy implementations.
 
 3. Compare Accuracy:
-    - Compute differences (e.g., Mean Absolute Error) between the outputs of custom and SciPy distributions for identical inputs.
+    - Compute differences (e.g., Absolute Error) between the outputs of custom and SciPy distributions for identical inputs.
 
 4. Output Results:
-    - Print or store results in an easily interpretable format (tables or plots).
+    - Print/plot or store results in an easily interpretable format (tables or plots).
 
 Expected Results
 -----------------
